% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{add.sub.version}
\alias{add.sub.version}
\title{Add Sub Version to Specification}
\usage{
add.sub.version(
  specification,
  sub.version,
  description,
  inherit.outcomes = T,
  can.seed.new.engine = T
)
}
\description{
Add Sub Version to Specification
}
