% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERVENTIONS_intervention_effects.R
\name{anchor.intervention.effect.location.and.version}
\alias{anchor.intervention.effect.location.and.version}
\title{Anchor Intervention Effect's Location and Version}
\usage{
anchor.intervention.effect.location.and.version(
  intervention.effect,
  location,
  specification.metadata,
  error.prefix = ""
)
}
\arguments{
\item{intervention.effect}{A 'jheem.intervention.effect' object that does not have a location or version set}

\item{specification.metadata}{A 'jheem.specification.metadata' object for the desired version. Its location need not match the location given as argument to this function.}
}
\value{
A new intervention effect that has location and version set
}
\description{
Anchor Intervention Effect's Location and Version
}
