% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{apply.ontology.mapping}
\alias{apply.ontology.mapping}
\title{Apply an Ontology Mapping to Data}
\usage{
apply.ontology.mapping(
  ontology.mapping,
  from.arr,
  to.dim.names = NULL,
  fun = "sum",
  na.rm = F,
  error.prefix = ""
)
}
\arguments{
\item{ontology.mapping}{An object of class 'ontology.mapping' - obtained either through \link{register.ontology.mapping}, \link{get.ontology.mapping}, \link{get.mappings.to.align.ontologies}}

\item{from.arr}{An array to transform. Can have any underlying data type, although the function is optimized for numeric data}

\item{to.dim.names}{The dimnames that the resulting transformed data should have (can represent a subset of the dimnames that the ontology mapping knows how to transform). If NULL, the ontology mapping will keep the dimnames in from.arr except for dimensions that are transformed}

\item{fun}{A function used to combine values, or the name of a function. The function must be able to accept any number of arguments (including none). The default, 'sum', is optimized for numeric arrays}

\item{na.rm}{Passed through to fun}

\item{error.prefix}{A character value to be prepended to any error messages thrown by the function}
}
\description{
Apply an Ontology Mapping to Data
}
