% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_main.R
\name{are.simulations.present}
\alias{are.simulations.present}
\title{Check if Simulations Have Been Saved}
\usage{
are.simulations.present(
  version,
  locations,
  calibration.code,
  interventions,
  n.sim,
  sub.version = NULL,
  root.dir = get.jheem.root.directory()
)
}
\arguments{
\item{version, calibration.code, locations}{The version, locations, and calibration code for which to gauge progress}

\item{root.dir}{The root directory from which to search for calibrations}

\item{as.pct}{Whether the progress should be shown as a percentage instead of a fraction}

\item{round.to.digits}{How many digits to round progress to}
}
\description{
Summarize the progress of a calibration task
}
