% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology.R
\name{as.list.ontology}
\alias{as.list.ontology}
\title{Convert an ontology back to a plain list}
\usage{
\method{as.list}{ontology}(ont)
}
\arguments{
\item{ont}{An ontology as created by \code{\link{ontology}}}
}
\description{
Convert an ontology back to a plain list
}
