% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology.R
\name{as.ontology}
\alias{as.ontology}
\title{Convert a named list to an ontology}
\usage{
as.ontology(x, incomplete.dimensions = NULL)
}
\arguments{
\item{x}{A named list of charactera vectors}
}
\description{
Convert a named list to an ontology
}
