% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_main.R
\name{assemble.simulations.from.calibration}
\alias{assemble.simulations.from.calibration}
\title{Assemble Simulations From A Calibration}
\usage{
assemble.simulations.from.calibration(
  version,
  location,
  calibration.code,
  root.dir = get.jheem.root.directory("Cannot set up calibration: "),
  allow.incomplete = F,
  chains = NULL,
  verbose = T
)
}
\arguments{
\item{version}{The version of the model specification (must have been previously registered)}

\item{location}{A single character value representing the location for the calibration}

\item{chains}{}
}
\description{
Assemble Simulations From A Calibration
}
