% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_transmute_calibration.R
\name{assemble.transmuted.simulations}
\alias{assemble.transmuted.simulations}
\title{Assemble Transmuted Simulations}
\usage{
assemble.transmuted.simulations(
  transmute.code,
  location,
  n.sim,
  to.version = NULL,
  to.sub.version = NULL,
  chunks = NULL,
  allow.incomplete = F,
  root.dir = get.jheem.root.directory()
)
}
\description{
Assemble Transmuted Simulations
}
