% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_simulation.R
\name{burn.simset}
\alias{burn.simset}
\title{Burn Simulations From a Simulation Set}
\usage{
burn.simset(simset, keep = NULL)
}
\arguments{
\item{keep}{Either a number of simulations to keep or a fraction to keep (between 0 and 1). The latest simulations will be kept.}
}
\description{
Subset a simulation set by trimming off a portion of earlier simulations.
}
\details{
Decimal values above 1 are rounded down to the nearest whole number.
}
