% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_main.R
\name{cache.mcmc.summary}
\alias{cache.mcmc.summary}
\title{Cache A Summary of an MCMC}
\usage{
cache.mcmc.summary(
  version,
  location,
  calibration.code,
  root.dir = get.jheem.root.directory(),
  get.one.set.of.parameters = F,
  throw.error.if.incomplete = T
)
}
\description{
Cache A Summary of an MCMC
}
