% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_transmutation.R
\name{can.transmute.versions}
\alias{can.transmute.versions}
\title{Assess Whether Simulations from one Version/Specification can be Transmuted to Another Version}
\usage{
can.transmute.versions(
  from.version,
  to.version,
  from.sub.version = NULL,
  to.sub.version = NULL,
  print.details = T
)
}
\arguments{
\item{from.version, to.version}{The versions from and to which we want to transmute}

\item{from.sub.version, to.sub.versions}{The sub-versions from and to which we want to transmute}

\item{print.details}{Whether to print the reasons why transmutation cannot happen}
}
\description{
'Transmutation' of a simulation is calculating a set of outcomes for another version without re-running the (expensive) differential equation solver
}
