% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERSIONS_version_manager.R
\name{clear.all.managers}
\alias{clear.all.managers}
\title{Clear JHEEM Version, Calibration, Intervention, and Ontology Mappings}
\usage{
clear.all.managers()
}
\description{
Clear JHEEM Version, Calibration, Intervention, and Ontology Mappings
}
