% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_main.R
\name{clear.calibration.cache}
\alias{clear.calibration.cache}
\title{Clear a Calibration Cache}
\usage{
clear.calibration.cache(
  version,
  location,
  calibration.code,
  root.dir = get.jheem.root.directory("Cannot set up calibration: "),
  remove.mcmc.summary = F,
  allow.remove.incomplete = F
)
}
\arguments{
\item{version}{The version of the model specification (must have been previously registered)}

\item{location}{A single character value representing the location for the calibration}

\item{allow.remove.incomplete}{}
}
\description{
Clear a Calibration Cache
}
