% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_transmute_calibration.R
\name{clear.transmute.calibration.cache}
\alias{clear.transmute.calibration.cache}
\title{Clear a Calibration Cache}
\usage{
clear.transmute.calibration.cache(
  transmute.code,
  location,
  n.sim,
  to.version = NULL,
  to.sub.version = NULL,
  allow.remove.incomplete = F,
  root.dir = get.jheem.root.directory()
)
}
\arguments{
\item{location}{A single character value representing the location for the calibration}

\item{allow.remove.incomplete}{}
}
\description{
Clear a Calibration Cache
}
