% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPERS_array_helpers.R
\name{collapse.array.according.to.indices}
\alias{collapse.array.according.to.indices}
\title{Collapse an Array According to Indices}
\usage{
collapse.array.according.to.indices(
  arr,
  small.indices,
  large.indices,
  small.n,
  check.consistency = T
)
}
\description{
Collapse an Array According to Indices
}
