% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{combine.ontology.mappings}
\alias{combine.ontology.mappings}
\title{Combine multiple ontology mappings into a single mapping}
\usage{
combine.ontology.mappings(...)
}
\arguments{
\item{...}{Either ontology.mapping objects or lists containing ontology.mapping objects#'}
}
\description{
Combine multiple ontology mappings into a single mapping
}
