% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_main.R
\name{compare.sims}
\alias{compare.sims}
\title{Compare Two Simsets}
\usage{
compare.sims(likelihood, sim1, sim2, piecewise = T, log = F)
}
\arguments{
\item{sim1, sim2}{'jheem.simulation.set' objects to compare}

\item{piecewise}{Should they be compared on each likelihood individually, if a joint likelihood is supplied?}

\item{log}{Whether to use log likelihood}
}
\description{
Compare how two simsets score on a likelihood
}
