% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_main.R
\name{compute.likelihood}
\alias{compute.likelihood}
\title{Compute Likelihood}
\usage{
compute.likelihood(
  likelihood,
  sim,
  log = T,
  use.optimized.get = F,
  check.consistency = T,
  debug = F
)
}
\arguments{
\item{log}{Whether to use log likelihood}

\item{use.optimized.get}{Should the optimized sim$get be used? The optimized get only works on simsets of the same internal structure as that defined in the specification at the time this likelihood was instantiated, so use it for the MCMC but leave it FALSE for analysis.}

\item{check.consistency}{- Whether to spend time checking to make sure everything is internally consistent. Setting to F is faster, but may generate weird error messages if there are bugs}
}
\description{
Compute Likelihood
}
