% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_main.R
\name{copy.calibration.info}
\alias{copy.calibration.info}
\title{Copy Calibration Info to a New Code}
\usage{
copy.calibration.info(
  from.code,
  to.code,
  n.iter = NULL,
  n.burn = NULL,
  thin = NULL,
  solver.metadata = NULL
)
}
\arguments{
\item{from.code}{The code for the the calibration info to copy}

\item{to.code}{The calibration code to be used for the copied calibration info}
}
\description{
Copy Calibration Info to a New Code
}
