% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{copy.data.manager}
\alias{copy.data.manager}
\title{Create a copy of a JHEEM Data Manager}
\usage{
copy.data.manager(
  data.manager = get.default.data.manager(),
  name,
  description,
  set.as.default = F
)
}
\arguments{
\item{data.manager}{A jheem.data.manager object to be copied}

\item{name}{The name of the new data manager}

\item{description}{A short description of the new data manager. If NULL, the description of the copied data manager.}

\item{set.as.default}{Should the new data manager be set as the default data manager}
}
\description{
Create a copy of a JHEEM Data Manager
}
