% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_basic_likelihood.R
\name{create.basic.likelihood.instructions.with.included.multiplier}
\alias{create.basic.likelihood.instructions.with.included.multiplier}
\title{Create JHEEM Basic Likelihood Instructions With Included Multiplier}
\usage{
create.basic.likelihood.instructions.with.included.multiplier(
  outcome.for.data,
  outcome.for.sim,
  dimensions = character(0),
  denominator.dimensions = dimensions,
  dimension.values = NULL,
  levels.of.stratification = NULL,
  from.year = -Inf,
  to.year = Inf,
  omit.years = NULL,
  sources.to.use = NULL,
  minimum.error.sd = 0,
  included.multiplier,
  included.multiplier.sd,
  included.multiplier.correlation = NULL,
  included.multiplier.correlation.structure = c("compound.symmetry",
    "autoregressive.1")[1],
  correlation.different.years = 0.5,
  correlation.different.strata = 0.1,
  correlation.different.sources = 0.3,
  correlation.same.source.different.details = 0.3,
  observation.correlation.form = c("compound.symmetry", "autoregressive.1")[1],
  error.variance.term = NULL,
  error.variance.type = NULL,
  weights = list(),
  equalize.weight.by.year = T,
  name = outcome.for.sim,
  na.rm = NULL
)
}
\arguments{
\item{outcome.for.data}{A single character vector specifying the outcome that data managers will use to find data.}

\item{outcome.for.sim}{A single character vector specifying the simulation outcome to use.}

\item{dimensions}{A character vector of dimensions, excluding year, from which stratifications will be generated.}

\item{denominator.dimensions}{(Optional) Analogous to 'dimensions' but for the simulation denominator outcome when applicable. Must be a subset of 'dimensions'.}

\item{levels.of.stratification}{An integer vector specifying how the dimensions should be combined to form strata. '0' indicates totals (not stratified) while '1' indicates strata that are each stratified by one dimension at a time, '2' indicates strat that are each stratified by a combination of two dimensions at a time, etc. May not exceed the number of dimensions. Defaults to NULL, which is equivalent to '0'.}

\item{from.year}{-Inf or a single integer value specifying the earliest year for which data should be pulled.}

\item{to.year}{Inf or a single integer value specifying the latest year for which data should be pulled.}

\item{omit.years}{(Optional) An integer vector of years to ignore.}

\item{sources.to.use}{A character vector of sources from which to pull data.}

\item{minimum.error.sd}{A single, positive numeric value. If any standard deviations are calculated to a value below this, they will be replaced with this value. This can help keep the likelihood from computing to negative infinity.}

\item{included.multiplier}{The value of the multiplier that has already been used in calculating the simulation value of the outcome (represents the mean of the distribution around the multiplier).}

\item{included.multiplier.sd}{The standard deviation of the uncertainty around the included multiplier.}

\item{included.multiplier.correlation}{The correlation between values of the multiplier across different strata and years.}

\item{included.multiplier.correlation.structure}{Which correlation form to use for values of the multiplier across various strata and years.}

\item{correlation.different.years}{A single numeric value specifying the correlation between observations of different years.}

\item{correlation.different.strata}{A single numeric value specifying the correlation between observations from different strata.}

\item{correlation.different.sources}{A single numeric value specifying the correlation between observations from different sources.}

\item{correlation.same.source.different.details}{A single numeric value specifying the correlation between observations of the same source but which have different 'details'.}

\item{observation.correlation.form}{Which correlation form to use in building the measurement error covariance matrix. May be 'compound.symmetry' or 'autoregressive.1'.}

\item{error.variance.type, error.variance.term}{The type of error variance(s) to be used and the corresponding value or function required for each. See details for the allowable values. For more than one type, supply them in a vector. If you therefore need more than one term, supply these in a list.}

\item{weights}{A list containing only numeric vectors and objects of class 'jheem.likelihood.weights', which are generated by \code{\link{create.likelihood.weights}}}

\item{equalize.weight.by.year}{A Boolean indicating whether every year should be weighted equally before the weights from 'weights' are applied.}

\item{na.rm}{Setting this TRUE will pass 'na.rm=T' along to the pull function. If set to NULL, 'na.rm=T' will be used for outcomes with scale 'proportion', 'rate', or 'ratio'.}
}
\description{
Create JHEEM Basic Likelihood Instructions With Included Multiplier
}
\details{
Error variance may be specified in a variety of ways, and more than
one can be used in the same likelihood. 'error.variance.type' is a vector
containing single character values describing the type(s) of variance to use,
and there is one 'error.variance.term' corresponding to each. The types "sd"
(standard deviation), "variance", "cv" (covariance), and "exp.of.variance"
(exponent of variance) require a single numeric value as their term. The
types "data.sd", "data.variance", and "data.cv" indicate that error data will
be pulled from the data manager and matched directly to data points, and they
require NULL as their term. Finally, the type "function.sd" requires as its
term a function that takes only arguments "data", "details", "version", and "location"
and returns a
numeric array of the same dimensions as "data". Supplying multiple types (and
therefore terms) results in the variances created by each being summed prior
to use in the measurement error covariance matrix.

As an example, one type could be specified with an 'error.variance.type' of
"cv" and an 'error.variance.term' of 0.13 (or list(0.13)). Two or more
types might look like an 'error.variance.type' of `c("data.sd",
"function.sd") and an 'error.variance.term' of list(NULL, my_sd_function). It
is possible to use the same type multiple times, although this is not
recommended.
}
\seealso{
Other Basic Likelihoods: 
\code{\link{create.basic.likelihood.instructions}()},
\code{\link{create.basic.likelihood.instructions.with.specified.outcome}()},
\code{\link{create.time.lagged.comparison.likelihood.instructions}()},
\code{\link{create.time.lagged.comparison.likelihood.instructions.with.included.multiplier}()}
}
\concept{Basic Likelihoods}
