% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_basic_likelihood.R
\name{create.basic.likelihood.instructions.with.specified.outcome}
\alias{create.basic.likelihood.instructions.with.specified.outcome}
\title{Create JHEEM Basic Likelihood Instructions With Specified Outcome}
\usage{
create.basic.likelihood.instructions.with.specified.outcome(
  outcome.for.sim,
  outcome.value,
  from.year,
  to.year,
  omit.years = NULL,
  correlation.different.years = 0.5,
  observation.correlation.form = c("compound.symmetry", "autoregressive.1")[1],
  error.variance.term = NULL,
  error.variance.type = NULL,
  weights = list(),
  equalize.weight.by.year = T,
  name = outcome.for.sim
)
}
\arguments{
\item{outcome.for.sim}{A single character vector specifying the simulation outcome to use.}

\item{outcome.value}{A single, non-NA numeric value that is the value of the outcome at the totals level (i.e., not stratified).}

\item{from.year}{-Inf or a single integer value specifying the earliest year for which data should be pulled.}

\item{to.year}{Inf or a single integer value specifying the latest year for which data should be pulled.}

\item{omit.years}{(Optional) An integer vector of years to ignore.}

\item{correlation.different.years}{A single numeric value specifying the correlation between observations of different years.}

\item{observation.correlation.form}{Which correlation form to use in building the measurement error covariance matrix. May be 'compound.symmetry' or 'autoregressive.1'.}

\item{error.variance.type, error.variance.term}{The type of error variance(s) to be used and the corresponding value or function required for each. See details for the allowable values. For more than one type, supply them in a vector. If you therefore need more than one term, supply these in a list.}

\item{weights}{A list containing only numeric vectors and objects of class 'jheem.likelihood.weights', which are generated by \code{\link{create.likelihood.weights}}}

\item{equalize.weight.by.year}{A Boolean indicating whether every year should be weighted equally before the weights from 'weights' are applied.}
}
\description{
Create JHEEM Basic Likelihood Instructions With Specified Outcome
}
