% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_custom_likelihood.R
\name{create.custom.likelihood.instructions}
\alias{create.custom.likelihood.instructions}
\title{Create Custom Likelihood Instructions}
\usage{
create.custom.likelihood.instructions(
  name,
  compute.function,
  get.data.function = NULL,
  weights = NULL,
  verbose = F
)
}
\arguments{
\item{compute.function}{A function that takes only arguments "sim", "log", and "weights", with optional "data" and "debug" arguments.}

\item{weights}{A list containing only numeric vectors and objects of class 'jheem.likelihood.weights', which are generated by \code{\link{create.likelihood.weights}}}
}
\description{
Create instructions for a likelihood that computes using a custom function.
}
\details{
All weights used in the compute function should arrive in its "weights" argument in two ways:
From the "weights" argument of the likelihood instructions (which apply to these instructions only),
and additional weights passed top-down from joint likelihoods that will include this likelihood.
So, no weights should be passed in the "data" function, for example, to maintain consistency and accuracy.
}
