% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{create.data.manager}
\alias{create.data.manager}
\title{Create a JHEEM Data Manager}
\usage{
create.data.manager(name, description, set.as.default = F)
}
\arguments{
\item{name}{The name of the data manager}

\item{description}{A short description}
}
\description{
Create a JHEEM Data Manager
}
