% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERVENTIONS_intervention_effects.R
\name{create.intervention.effect}
\alias{create.intervention.effect}
\title{Create an Effect for an Intervention to Have on a Model Quantity}
\usage{
create.intervention.effect(
  quantity.name,
  start.time,
  effect.values,
  times,
  end.time = Inf,
  apply.effects.as,
  scale,
  allow.values.less.than.otherwise,
  allow.values.greater.than.otherwise,
  bindings = NULL,
  ...
)
}
\arguments{
\item{quantity.name}{The name of the quantity, defined in a jheem.model.specification, which an intervention affects}

\item{start.time}{The time at which the intervention begins to take effect (ie, depart from what the value of the quantity would otherwise have been)}

\item{effect.values}{The values to apply to the quantity at times. Must be either (1) a numeric vector, (2) a character vector, (3) an expression or call, or (4) a list containing only numeric vectors, character vectors, expressions, or calls}

\item{times}{The times as which values apply to the quantity.}

\item{end.time}{The time at which the intervention stops taking effect (ie, returns to what the value of the quantity otherwise would have been)}

\item{apply.effects.as}{A single character value indicating how values should be applied to the quantity. Choices are (1) 'value' - the quantity takes the given value, (2) 'multiplier' - the value is multiplied by the value the quantity would otherwise have taken, or (3) 'addend' - the value is added to the value the quantity would otherwise have taken.}

\item{scale}{The scale at which values are applied to the quantity}

\item{allow.values.less.than.otherwise, allow.values.greater.than.otherwise}{Logical indicators of whether the intervention may cause the value of the quantity to be less/greater than it otherwise would have been}

\item{bindings}{A named list. The names must correspond to variables on which the intervention effect applies, if it is a variable. The elements of the list must be either (1) numeric values or (2) functions must take only "location" and "specification.metadata" as arguments and return numeric vectors with no NA values}

\item{...}{Additional arguments to be passed to any bindings which are functions}
}
\description{
Create an Effect for an Intervention to Have on a Model Quantity
}
