% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERVENTIONS_main.R
\name{create.intervention}
\alias{create.intervention}
\title{Create an Intervention}
\usage{
create.intervention(
  ...,
  code = NULL,
  name = NULL,
  parameters = NULL,
  parameter.distribution = NULL,
  generate.parameters.function = NULL,
  overwrite.existing.intervention = F
)
}
\arguments{
\item{...}{Either (a) objects of class 'target.population', (b) objects of class 'intervention.effect', or (d) lists containing only 'target.population' or 'intervention.effect' objects. Must contain at least one 'target.population' object and at least one 'intervention.effect' object}

\item{code}{A brief (2-30 character) string that will uniquely identify this intervention}

\item{parameter.distribution}{}
}
\description{
Create an Intervention
}
\details{
Creates a 'jheem.intervention' object where all the intervention effects given in ... apply to all the target populations given in ...
}
