% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{create.jheem.engine}
\alias{create.jheem.engine}
\title{Create a JHEEM Engine (to run simulations for a given version and location)}
\usage{
create.jheem.engine(version, location)
}
\arguments{
\item{version}{The name of the version to run. Must have had a model specification registered with \code{\link{register.model.specification}}}

\item{location}{A single character value indicating the location of the model}
}
\description{
Create a JHEEM Engine (to run simulations for a given version and location)
}
\seealso{
Other Functions to create and modify a jheem.engine: 
\code{\link{crunch.jheem.engine}()},
\code{\link{run.jheem.engine}()},
\code{\link{set.element.functional.form.from.time}()},
\code{\link{set.element.functional.form.future.slope}()},
\code{\link{set.element.functional.form.interaction.alphas}()},
\code{\link{set.element.functional.form.main.effect.alphas}()},
\code{\link{set.element.ramp.times}()},
\code{\link{set.element.values.from.parameters}()},
\code{\link{set.element.value}()}
}
\concept{Functions to create and modify a jheem.engine}
