% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_transmutation.R
\name{create.jheem.transmuter}
\alias{create.jheem.transmuter}
\title{Create a Transmuter}
\usage{
create.jheem.transmuter(
  simulation.set,
  to.version,
  from.year = simulation.set$from.year,
  to.year = simulation.set$to.year,
  to.sub.version = NULL
)
}
\arguments{
\item{simulation.set}{The simulation.set to transmute}

\item{to.version, to.sub.version}{The version and sub.version to transmute to}
}
\description{
Create a Transmuter
}
\details{
A jheem.transmuter can convert simulations from one version/specification to another using its transmute() method
}
