% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_ifelse_likelihood_instructions.R
\name{create.location.based.ifelse.likelihood.instructions}
\alias{create.location.based.ifelse.likelihood.instructions}
\title{Create Location Based If-Else Likelihood Instructions}
\usage{
create.location.based.ifelse.likelihood.instructions(
  ...,
  name = NULL,
  locations.list = list()
)
}
\arguments{
\item{...}{'jheem.likelihood.instructions' objects that all have the same 'outcome.for.sim'.}

\item{name}{Name for the ifelse; will default to}

\item{locations.list}{A list of character vectors of locations that will use the corresponding instruction.
The list should be length one shorter than the number of instructions. For example, if you have three instructions,
the first vector has locations that will use the first instruction and second vector has locations that will
use the second instruction. The third instruction will be used for all other locations.}
}
\description{
Create instructions that can instantiate a different 'jheem.likelihood' depending on the location.
}
