% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_functional_forms.R
\name{create.logistic.spline.functional.form}
\alias{create.logistic.spline.functional.form}
\title{Create a Logistic Spline Functional Form}
\usage{
create.logistic.spline.functional.form(
  knot.times,
  knot.values,
  min = -Inf,
  max = Inf,
  knot.link = c("identity", "log", "logit")[2],
  knot.min = NA,
  knot.max = NA,
  knots.are.on.transformed.scale = F,
  overwrite.knot.values.with.alphas = F,
  fraction.of.asymptote.after.end = 0.05,
  fraction.of.asymptote.before.start = 0.025,
  fraction.of.asymptote.for.change.dir = 0.02,
  fraction.asymptote.link = "log",
  overwrite.fraction.asymptote.with.alphas = F,
  error.prefix = "Cannot create logistic spline functional.form: "
)
}
\arguments{
\item{knot.times}{The (numeric) times at which the knots of the spline apply. Must have names set}

\item{knot.values}{Initial values for the knots. Must have names set, matching the names of knot times}

\item{min, max}{The minimum and maximum values this functional form can create. The default (NA) sets to the min/max for the link specified (eg, 0-Inf for a log link, 0-1 for a logistic link)}

\item{knot.link}{The name of a transformation to the scale at which the knot values have alphas ADDED. The knots are back-transformed from this scale prior to the spline scale being applied}

\item{knot.min, knot.max}{The min and max values that KNOTS can take before being splined (note, the min and max values the functional form can take, after splining, are determined by parameters min and max)}

\item{knots.are.on.transformed.scale}{Logical indicating whether the knots are already on the transformed scale (ie knot.link has been applied)}

\item{fraction.of.asymptote.after.end, fraction.of.asymptote.before.start}{The fraction of the total change in the first or last segment that would take place before the first knot or after the last}

\item{fraction.of.asymptote.for.change.dir}{The fraction of the total change in a segment that should take place before between two segments}

\item{fraction.asymptote.link}{The name of a transformation to the scale at which alphas are added to the betas for the fraction.asymptote.<x> values}

\item{overwrite.fraction.asymptote.with.alphas}{Logical indicating whether, when alphas are applied to the fraction.of.asymptote.<x>, they OVERWRITE the values. If F, alphas are interpreted as either additive or multiplicative to the knot.values, depending on the value of fraction.asymptote.alphas.are.multipliers}

\item{error.prefix}{String to prepend to any error message in case of an error}
}
\description{
Create a Logistic Spline Functional Form
}
