% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_functional_forms.R
\name{create.logistic.tail.functional.form}
\alias{create.logistic.tail.functional.form}
\title{Create a "Logistic Tail" Functional Form}
\usage{
create.logistic.tail.functional.form(
  intercept,
  slope,
  anchor.year,
  logistic.after.frac.of.span = 0.5,
  min = 0,
  max = 1,
  intercept.link = "logistic",
  intercept.min = NA,
  intercept.max = NA,
  slope.link = intercept.link,
  slope.min = NA,
  slope.max = NA,
  parameters.are.on.transformed.scale = T,
  overwrite.parameters.with.alphas = F,
  error.prefix = "Cannot create logistic-tail functional.form: "
)
}
\arguments{
\item{intercept, slope}{Scalar values or arrays representing the intercept and slope. Note that the final form constrains the slope to be non-negative}

\item{anchor.year}{The year at which the functional form evaluates to the intercept}

\item{logistic.after.frac.of.span}{The fraction of the the total span (max - min) after which the model follows a logistic curve}

\item{min, max}{The upper and lower limits to which the functional form can evaluate}

\item{intercept.link, slope.link}{The names of the transformations to the scale on which alphas are added to calculate the intercept and slope. One of "identity", "log", "logit"}

\item{intercept.min, intercept.max, slope.min, slope.max}{The min and max values to which intercept and slope van evaluate}

\item{overwrite.parameters.with.alphas}{Logical indicating whether, when alphas are applied, they OVERWRITE values. If F, alphas are added to the transformed value (on the transformed scale)}

\item{error.prefix}{String to prepend to any error message in case of an error}
}
\description{
Create a "Logistic Tail" Functional Form
}
\details{
A "logistic tail" functional form follows a linear (intercept + slope) in its first part, and a logistic-linear form in its second part. This allows a linear trend to taper off
}
