% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERVENTIONS_foreground.R
\name{create.model.foreground}
\alias{create.model.foreground}
\title{Create a Model Foreground Object}
\usage{
create.model.foreground(..., error.prefix = "")
}
\arguments{
\item{...}{The intervention.effects and target.populations for this foreground. Can be either: (1) an alternating set of jheem.intervention.effect objects and target.population objects (ie, one intervention.effect followed by one target.population, then the next pair, or vice versa) or (2) one list of intervention effects and one list of target populations, both of the same length}
}
\description{
Create a Model Foreground Object
}
\details{
A model foreground object applies multiple effects to multiple subpopulations *simultaneously*, with all effects targeting the same model quantity
}
