% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERVENTIONS_criteria_based.R
\name{create.monotonic.criteria.based.intervention}
\alias{create.monotonic.criteria.based.intervention}
\title{Create a "Guess-and-Check" Intervention that Must Satisfy Some Criteria by Varying Parameters}
\usage{
create.monotonic.criteria.based.intervention(
  base.intervention,
  completion.criteria,
  max.iterations,
  n.iterations.after.satisfying.criteria = 5,
  max.iterations.first.sim = max.iterations,
  n.iterations.after.satisfying.criteria.first.sim = ceiling(max.iterations.first.sim/2),
  max.failure.rate = 0,
  discount.prior.n = 0.5,
  code = NULL,
  name = NULL,
  overwrite.existing.intervention = F
)
}
\arguments{
\item{base.intervention}{A single.iteration.intervention, as created by \code{\link{create.intervention}} or \code{\link{join.interventions}}}

\item{completion.criteria}{Either a single jheem.intervention.criterion object, as created by \code{\link{create.intervention.criterion}} or a list containing only jheem.intervention.criterion objects}

\item{max.iterations}{The maximum number of iterations to run in trying to satisfy intervention criteria. If criteria are not satisfied after max.iterations, will 'give up' (and either produce degenerate simulations or throw an error, depending on max.failure.rate)}

\item{n.iterations.after.satisfying.criteria}{The number of iterations to run after we have satisfied the criteria to try to get as close as possible to the target}

\item{max.failure.rate}{The maximum proportion of simulations in a simulation set which can fail to find a solution that satisfies intervention criteria within max.iterations. If this threshold is exceeded, will throw an error; otherwise, will return degenerate simulations when it could not find a solution}
}
\description{
Create a "Guess-and-Check" Intervention that Must Satisfy Some Criteria by Varying Parameters
}
