% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERVENTIONS_criteria_based.R
\name{create.monotonic.criterion}
\alias{create.monotonic.criterion}
\title{Create a criterion for monotonic parameter, criteria-based interventions}
\usage{
create.monotonic.criterion(
  parameter.name,
  outcome,
  parameter.scale,
  parameter.initial.value,
  min.acceptable.parameter.value = NULL,
  max.acceptable.parameter.value = NULL,
  target.value,
  min.acceptable.value,
  max.acceptable.value,
  subsequent.thresholds.apply.after.iteration = numeric(),
  dimension.values = list(),
  draw.parameter.from.previous.sims = !is.function(parameter.initial.value),
  ...
)
}
\arguments{
\item{parameter.name}{The name of the parameter to be modified to achieve the outcome}

\item{outcome}{A character vector indicating the name of the outcome to which the criterion applies}

\item{parameter.scale}{The scale on which the parameter operates}

\item{parameter.initial.value}{Either a single scalar numeric or a function that takes argument sim and returns a single scalar numeric}

\item{min.acceptable.parameter.value, max.acceptable.parameter.value}{Indicators of the min and max possible values of the parameter to prevent unreasonably large (or small) parameter values crashing the algorithm. Must be single, numeric values or NULL, in which case the limits allowed by the 'parameter.scale' will be used.}

\item{target.value}{A single, numeric value that we want the outcome to get to}

\item{min.acceptable.value, max.acceptable.value}{Indicators of the min and max possible values of the outcome we would be willing to accept. Must be either (1) single, numeric values or (2) numeric vectors. In this case, the first values are the min/max threshold until subsequent.thresholds.apply.after.iteration[1], the second values are the thresholds prior to subsequent.thresholds.apply.after.iteration[2], etc}

\item{subsequent.thresholds.apply.after.iteration}{The number of iterations after which the 2nd, 3rd, etc thresholds in min.acceptable.value, max.acceptable.value apply}

\item{dimension.values}{The dimension values for which we should pull the outcome. Must be a named list of either character vectors, integer vectors, or logical vectors}

\item{...}{An alternative way of supplying dimensions values - must be named character vectors, integer vectors, or logical vectors}

\item{draw.parameters.from.previous.sims}{Whether the initial value for the parameter after the first simulation in a set should be the given initial value, or should be based on what worked for previous simulations in the set}
}
\description{
Create a criterion for monotonic parameter, criteria-based interventions
}
\details{
If the intervention algorithm finds it needs to keep raising (or lowering) the parameter value so high (or low) that it reaches the bounds, the parameter will be considered optimized.
This often occurs when the parameter's outcome is already beyond the target and would be generally deemed acceptable anyway.
}
