% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_functional_forms.R
\name{create.natural.spline.functional.form}
\alias{create.natural.spline.functional.form}
\title{Create a Natural Spline Functional Form}
\usage{
create.natural.spline.functional.form(
  knot.times,
  knot.values,
  link = c("identity", "log", "logit")[1],
  min = NA,
  max = NA,
  knot.link = link,
  knot.min = min,
  knot.max = max,
  knots.are.on.transformed.scale = F,
  overwrite.knot.values.with.alphas = F,
  before.time = NULL,
  before.modifier = NULL,
  after.time = NULL,
  after.modifier = NULL,
  modifiers.apply.to.change = T,
  modifier.link = link,
  after.modifier.increasing.change.link = modifier.link,
  after.modifier.decreasing.change.link = modifier.link,
  before.modifier.increasing.change.link = modifier.link,
  before.modifier.decreasing.change.link = modifier.link,
  modifier.min = min,
  modifier.max = max,
  overwrite.modifiers.with.alphas = F,
  error.prefix = "Cannot create natural spline functional.form: "
)
}
\arguments{
\item{knot.times}{The (numeric) times at which the knots of the spline apply. Must have names set}

\item{knot.values}{Initial values for the knots. Must have names set, matching the names of knot times}

\item{link}{The name of a transformation to the scale at which the SPLINE should apply. The knots are transformed to this scale, the spline is applied, and then the splined values are back-transformed. One of 'identity', 'log', or 'logistic'}

\item{min, max}{The minimum and maximum values this functional form can create. The default (NA) sets to the min/max for the link specified (eg, 0-Inf for a log link, 0-1 for a logistic link)}

\item{knot.link}{The name of a transformation to the scale at which the knot values have alphas ADDED. The knots are back-transformed from this scale prior to the spline scale being applied}

\item{knot.min, knot.max}{The min and max values that KNOTS can take before being splined (note, the min and max values the functional form can take, after splining, are determined by parameters min and max)}

\item{knots.are.on.transformed.scale}{Logical indicating whether the knots are already on the transformed scale (ie knot.link has been applied)}

\item{before.time, after.time}{The times at which the knots calculated from before.modifier and after.modifier apply, respectively}

\item{before.modifier, after.modifier}{If either of these is set, an additional knot is added to the specified knots. For before.modifier, this knot is calculated as before.modifier added or multiplier by the value of the first knot (on the model scale); after.modifier is added or multiplied by the value of the last knot}

\item{modifiers.apply.to.change}{If TRUE, after.modifier (the untransformed-by-link value) is multiplied by the change between the penultimate and last knots and added to the last knot to generate an additional knot at after.time (similarly, before.modifier is multiplied by the change from the first to the second and subtracted from the first). If FALSE, the transformed after.multiplier is added to the last knot after the last knot is transformed to modifier.link scale, and the resulting value is back-transformed to yield the knot at after.time (similarly, before.modifier is added to the first knot on modifier.link scale)}

\item{modifier.link}{The name of a transformation to the scale at which the before.modifer and after.modifier have alphas added}

\item{modifier.min, modifier.max}{The min and max values for before.modifier and after.modifier}

\item{overwrite.modifiers.with.alphas}{Logical indicating whether, when alphas are applied to the before.modifier or after.modifier, they OVERWRITE the modifiers. If F, alphas are added on the scale specified by modifier.link}

\item{error.prefix}{String to prepend to any error message in case of an error}
}
\description{
Create a Natural Spline Functional Form
}
