% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{create.ontology.mapping}
\alias{create.ontology.mapping}
\title{Create a mapping that allows transformations of arrays}
\usage{
create.ontology.mapping(
  mappings,
  from.dimensions,
  to.dimensions = from.dimensions,
  name = NULL,
  error.prefix = "Error creating ontology mapping: "
)
}
\arguments{
\item{mappings}{A character matrix, with a number of columns == length(from.dimensions) + length(to.dimensions). The first length(from.dimensions) columns represent values for the from.dimensions (in the same order as from.dimensions), and the last length(to.dimensions) colums represent values for to.dimensions (in the order of to.dimensions). Each row represents a mapping from a set of values for from.dimension to a set of values for to.dimensions. NB: any dimnames on mappings will be ignored}

\item{from.dimensions}{A character vector of names of the dimensions we are mapping from}

\item{to.dimensions}{A character vector of names of the dimensions we are mapping to}

\item{error.prefix}{A text string to prepend to any error messages generated in executing the function}
}
\description{
Create a mapping that allows transformations of arrays
}
