% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{create.other.catchall.ontology.mapping}
\alias{create.other.catchall.ontology.mapping}
\title{Create an ontology mapping that lumps otherwise unspecified categories into an "other" category}
\usage{
create.other.catchall.ontology.mapping(
  dimension,
  from.values,
  to.values,
  other.category.value = "other"
)
}
\arguments{
\item{dimension}{A single character value indicating what dimension this mapping applies to}

\item{from.values}{The possible values for the dimension in the ontology to map from}

\item{to.values}{The possible values for the dimension in the ontology to map to}

\item{other.category.value}{What the "other" category should be called in the to ontology}
}
\description{
Create an ontology mapping that lumps otherwise unspecified categories into an "other" category
}
