% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_outcome_location_mappings.R
\name{create.outcome.location.mapping}
\alias{create.outcome.location.mapping}
\title{Create Outcome Location Mapping}
\usage{
create.outcome.location.mapping(
  location.mappings,
  outcome.name,
  version = jheem.kernel$version,
  location = jheem.kernel$location,
  sub.version = NULL,
  jheem.kernel = NULL
)
}
\arguments{
\item{location.mappings}{is a named list, where the names correspond to modeled locations, and the elements are vectors of observed locations}
}
\description{
Create Outcome Location Mapping
}
