% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{create.outcome.metadata}
\alias{create.outcome.metadata}
\title{Create a metadata object for tracked quantities or tracked transitions}
\usage{
create.outcome.metadata(
  display.name,
  description,
  scale,
  axis.name,
  units,
  singular.unit = units,
  display.as.percent = scale == "proportion"
)
}
\arguments{
\item{display.name}{The (nicely formatted) name of the quantity - to use in titling figures and other reporting. Should be capitalized}

\item{description}{A phrase or single sentence describing the quantity. Should be captialized}

\item{scale}{The scale of the quantity. Either "rate", "ratio", "proportion", "time", "number", "non.negative.number"}

\item{axis.name}{A name to use in labeling axes on figures that display this quantity. Should be capitalized}

\item{units}{The units the quantity takes (eg, "cases", "%", "tests/year")}

\item{singular.unit}{The unit to use if the value of the quantity is exactly == 1}

\item{display.as.percent}{A logical indicator for whether, when plotted or otherwise displayed, the quantity should be converted to a percent. Only valid if scale=='proportion'}
}
\value{
An object of class 'outcome.metadata' which can be passed to \code{\link{track.transition}}, \code{\link{track.quantity}}, \code{\link{track.dynamic.quantity}}
}
\description{
Create a metadata object for tracked quantities or tracked transitions
}
