% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_simset_collection.R
\name{create.simset.collection}
\alias{create.simset.collection}
\title{Create a Collection of JHEEM Simulation-Sets}
\usage{
create.simset.collection(
  version,
  calibration.code,
  locations,
  interventions,
  n.sim = NULL,
  sub.version = NULL,
  root.dir = get.jheem.root.directory("Cannot create simset collection: ")
)
}
\arguments{
\item{version, sub.version}{The version and (optional) sub-version for the simsets}

\item{locations}{A character vector of locations for which to collect simsets}

\item{interventions}{The interventions for which to collect interventions. Can be either (a) a character vector of intervention codes or (b) a list of 'jheem.intervention' objects or single character intervention codes}

\item{n.sim}{Optional - the number of simulations which collected simsets should contain. If NULL, the number will attempt to be figured out}

\item{root.dir}{The root directory relative to which all simset files (and all JHEEM files) are saved}
}
\value{
An object of class 'jheem.simset.collection'
}
\description{
Create a Collection of JHEEM Simulation-Sets
}
