% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{create.solver.metadata}
\alias{create.solver.metadata}
\title{Create Solver Metadata}
\usage{
create.solver.metadata(
  method = "ode45",
  package = NULL,
  atol = 0.01,
  rtol = 0.001
)
}
\description{
Create Solver Metadata
}
