% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_functional_forms.R
\name{create.static.functional.form}
\alias{create.static.functional.form}
\title{Create an Static (Time-Invariant) Functional Form}
\usage{
create.static.functional.form(
  value,
  link = c("identity", "log", "logit")[1],
  min = NA,
  max = NA,
  value.name = "value",
  value.is.on.transformed.scale = F,
  overwrite.parameters.with.alphas = F,
  error.prefix = "Cannot create static functional.form: "
)
}
\arguments{
\item{value}{array representing the value}

\item{link}{The name of the transformation to the scale on which alphas are added to the value. One of "identity", "log", "logit"}

\item{min, max}{The minimum and maximum values this functional form can create. The default (NA) sets to the min/max for the link specified (eg, 0-Inf for a log link, 0-1 for a logistic link)}

\item{value.is.on.transformed.scale}{Logical indicating whether the value is already on the transformed scale at which alphas apply}

\item{overwrite.parameters.with.alphas}{Logical indicating whether, when alphas are applied, they OVERWRITE values. If F, alphas are added to the transformed value (on the transformed scale)}

\item{error.prefix}{String to prepend to any error message in case of an error}
}
\description{
Create an Static (Time-Invariant) Functional Form
}
