% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS_style_manager.R
\name{create.style.manager}
\alias{create.style.manager}
\title{Create A Style Manager}
\usage{
create.style.manager(
  color.sim.by = "stratum",
  color.data.by = "stratum",
  shade.data.by = "location",
  linetype.sim.by = "simset",
  shape.data.by = "source",
  shape.sim.by = "simset",
  sim.palette = ggsci::pal_jama(),
  data.palette = ggsci::pal_jama(),
  linetypes = c("solid", "dashed", "dotted", "dotdash"),
  shapes = 21:25,
  shade.increment = -64,
  alpha.ribbon = 0.2,
  alpha.line = 1,
  linewidth.baseline = 1,
  linewidth.slope = 3
)
}
\arguments{
\item{sim.palette, data.palette}{Must be either a character vector of colors to choose from or a function that takes a number of colors needed and returns a character vector of colors.}

\item{shapes}{A numeric vector indicating which of the 25 ggplot shapes to select from. Each value must be an integer between 1 and 25.}

\item{shade.increment}{A single numeric value less than 0 indicating the extent of shading to apply (to alternate values of the 'shade.data.by'/'shade.sim.by' variables)}

\item{alpha.line}{A single numeric value which determines the transparency of the simulation lines. If NULL, the width of the lines is used (see "linewidth.slope" argument for details).}

\item{linewidth.baseline}{A single numeric value greater than zero which determines the simset linewidth in combination with the "linewidth.slope" argument.}

\item{linewidth.slope}{A single numeric value specifying how simset linewidth will be determined for based on how many simulations it has (its "n.sim").
The formula used is linewidth = baseline/(slope * log10(n.sim) + 1), where "baseline" is the argument "linewidth.baseline". If this slope is the default of 3, and linewidth.baseline is the default of 1, this will result in simsets with 1000 simulations having a linewidth of 0.1
while simsets with only 1 simulation will have a linewidth of 1. If you would like all simsets to plot with the same thickness, use a slope of 0.}
}
\description{
Create A Style Manager
}
