% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERVENTIONS_target_populations.R
\name{create.target.population}
\alias{create.target.population}
\title{Create a Target Population to Which Interventions May be Applied}
\usage{
create.target.population(..., name, invert = F)
}
\arguments{
\item{...}{Dimension values that denote the categories that comprise this target population. Each element must be namned with the name of the dimension, and can be either (1) character values, (2) integers, or (3) logical vectors}

\item{name}{A short (<= 20 characters) descriptive name for the target population. Should be nicely formatted for public display}

\item{invert}{A logical indicating whether the target population should be the INVERSE of the characteristics specified in ... (ie, everyone who does NOT fall into those categories)}
}
\description{
Create a Target Population to Which Interventions May be Applied
}
\seealso{
Other Functions to create Target Populations: 
\code{\link{diff.target.populations}()},
\code{\link{intersect.target.populations}()},
\code{\link{union.target.populations}()}
}
\concept{Functions to create Target Populations}
