% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_basic_likelihood.R
\name{create.time.lagged.comparison.likelihood.instructions}
\alias{create.time.lagged.comparison.likelihood.instructions}
\title{Create JHEEM Basic Likelihood Instructions}
\usage{
create.time.lagged.comparison.likelihood.instructions(
  outcome.for.data,
  outcome.for.sim,
  dimensions = character(0),
  denominator.dimensions = dimensions,
  dimension.values = NULL,
  levels.of.stratification = NULL,
  from.year = -Inf,
  to.year = Inf,
  omit.years = NULL,
  sources.to.use = NULL,
  correlation.different.years = 0.5,
  correlation.different.strata = 0.1,
  correlation.different.sources = 0.3,
  correlation.same.source.different.details = 0.3,
  observation.correlation.form = c("compound.symmetry", "autoregressive.1")[1],
  error.variance.term = NULL,
  error.variance.type = NULL,
  weights = list(),
  equalize.weight.by.year = T,
  use.lognormal.approximation = T
)
}
\arguments{
\item{outcome.for.data}{A single character vector specifying the outcome that data managers will use to find data.}

\item{outcome.for.sim}{A single character vector specifying the simulation outcome to use.}

\item{dimensions}{A character vector of dimensions, excluding year, from which stratifications will be generated.}

\item{denominator.dimensions}{(Optional) Analogous to 'dimensions' but for the simulation denominator outcome when applicable. Must be a subset of 'dimensions'.}

\item{levels.of.stratification}{An integer vector specifying how the dimensions should be combined to form strata. '0' indicates totals (not stratified) while '1' indicates strata that are each stratified by one dimension at a time, '2' indicates strat that are each stratified by a combination of two dimensions at a time, etc. May not exceed the number of dimensions. Defaults to NULL, which is equivalent to '0'.}

\item{from.year}{-Inf or a single integer value specifying the earliest year for which data should be pulled.}

\item{to.year}{Inf or a single integer value specifying the latest year for which data should be pulled.}

\item{omit.years}{(Optional) An integer vector of years to ignore.}

\item{sources.to.use}{A character vector of sources from which to pull data.}

\item{correlation.different.years}{A single numeric value specifying the correlation between observations of different years.}

\item{correlation.different.strata}{A single numeric value specifying the correlation between observations from different strata.}

\item{correlation.different.sources}{A single numeric value specifying the correlation between observations from different sources.}

\item{correlation.same.source.different.details}{A single numeric value specifying the correlation between observations of the same source but which have different 'details'.}

\item{observation.correlation.form}{Which correlation form to use in building the measurement error covariance matrix. May be 'compound.symmetry' or 'autoregressive.1'.}

\item{error.variance.term}{If applicable for the 'error.variance.type' used, a single numeric value. If the 'error.variance.type' is 'function.sd', then this term should be a function that takes arguments 'data' and 'details' and returns a numeric array of the same dimensions as ‘data’, with no NA values, that represents the sd for each measurement in data.}

\item{error.variance.type}{A single character value specifying how the error variance is described. May refer to a single quantity with "sd" (standard deviation), "variance", "cv" (covariance); or may request the error variance be sourced from data with "data.sd", "data.variance", and "data.ci" (confidence interval). The error variance may be specified with more nuance if set to 'function.sd' (see description in 'error.variance.term' definition).}

\item{weights}{A list containing only numeric vectors and objects of class 'jheem.likelihood.weights', which are generated by \code{\link{create.likelihood.weights}}}

\item{equalize.weight.by.year}{A Boolean indicating whether every year should be weighted equally before the weights from 'weights' are applied.}

\item{use.lognormal.approximation}{A single logical value.}
}
\description{
Create JHEEM Basic Likelihood Instructions
}
