% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{crunch.jheem.engine}
\alias{crunch.jheem.engine}
\title{Perform Pre-Calculations Needed to Run a Simulation From a JHEEM Engine}
\usage{
crunch.jheem.engine(jheem.engine, start.year, end.year, check.consistency = T)
}
\arguments{
\item{jheem.engine}{A jheem.engine object (created by \code{\link{create.jheem.engine}})}

\item{start.year, end.year}{The years across which the simulation is to run}

\item{check.consistency}{Whether to perform consistency checks during calculations, which can catch errors, but imposes a computational cost. In general, this is likely only necessary on the first time an engine is crunched or run}
}
\description{
Perform Pre-Calculations Needed to Run a Simulation From a JHEEM Engine
}
\seealso{
Other Functions to create and modify a jheem.engine: 
\code{\link{create.jheem.engine}()},
\code{\link{run.jheem.engine}()},
\code{\link{set.element.functional.form.from.time}()},
\code{\link{set.element.functional.form.future.slope}()},
\code{\link{set.element.functional.form.interaction.alphas}()},
\code{\link{set.element.functional.form.main.effect.alphas}()},
\code{\link{set.element.ramp.times}()},
\code{\link{set.element.values.from.parameters}()},
\code{\link{set.element.value}()}
}
\concept{Functions to create and modify a jheem.engine}
