% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERVENTIONS_target_populations.R
\name{diff.target.populations}
\alias{diff.target.populations}
\title{Create a Target Population that is the DIFFERENCE between two other Target Populations}
\usage{
\method{diff}{target.populations}(pop1, pop2, name = NULL)
}
\arguments{
\item{pop1, pop2}{Target Population Objects}
}
\description{
Create a Target Population that is the DIFFERENCE between two other Target Populations
}
\details{
The resulting target population contains everyone who IS in pop1 but NOT in pop2
}
\seealso{
Other Functions to create Target Populations: 
\code{\link{create.target.population}()},
\code{\link{intersect.target.populations}()},
\code{\link{union.target.populations}()}
}
\concept{Functions to create Target Populations}
