% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPERS_array_helpers.R
\name{distribute.dimension.values}
\alias{distribute.dimension.values}
\title{Distribute dimension values}
\usage{
distribute.dimension.values(arr, dimension.values.to.distribute)
}
\arguments{
\item{arr}{An array -- no validation is performed}

\item{dimension.values.to.distribute}{A named list of the dimensions and their values which should be redistributed}
}
\value{
An array with the same number of dimensions as 'arr', but without the dimension values specified in 'dimension.values.to.distribute'.
}
\description{
This function redistributes certain values in a dimension across the remaining values in that dimension.
This is useful to remove problematic dimension values -- like race = 'unknown' -- while keeping the marginal sums and the
relevant ratios between the remaining dimension values the same as they were before.
}
\details{
If multiple dimensions should be redistributed, they will be redistributed in the order that they exist in 'dimension.values.to.distribute'.
This can result in different outputs for different orders.
}
