% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{examine.get.mappings.to.align.ontologies}
\alias{examine.get.mappings.to.align.ontologies}
\title{Identify any Reasons for a Failure to Get Mappings to Align Ontologies}
\usage{
examine.get.mappings.to.align.ontologies(
  ontology.1,
  ontology.2,
  align.on.dimensions = intersect(names(ontology.1), names(ontology.2)),
  allow.non.overlapping.incomplete.dimensions = F,
  include.dim.names = NULL,
  print.ontologies = T
)
}
\arguments{
\item{ontology.1, ontology.2}{The ontologies of two data elements to be transformed}

\item{align.on.dimensions}{The dimensions which should match in the transformed data}

\item{allow.non.overlapping.incomplete.dimensions}{Logical indicating whether, when we map to an incomplete dimension, there can be no overlap in the from and to values for that dimension}

\item{include.dim.names}{Optional argument, specifying specific dimension values that must be present in the transformed data}

\item{print.ontologies}{Whether to print the full ontologies in the details}
}
\description{
For use in figuring out why a call to \code{\link{get.mappings.to.align.ontologies}} did not work

If no mappings can be found, prints detailed information on why we were unable to find mappings
}
