% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{examine.get.ontology.mapping}
\alias{examine.get.ontology.mapping}
\title{Identify any Reasons for a Failure to Map Ontologies}
\usage{
examine.get.ontology.mapping(
  from.ontology,
  to.ontology,
  allow.non.overlapping.incomplete.dimensions = F,
  print.ontologies = T
)
}
\arguments{
\item{from.ontology}{The ontology to which the data to be transformed conform. Must be an 'ontology' object as created by \code{\link{ontology}}}

\item{to.ontology}{The ontology to which the transformed data should conform. Must be an 'ontology' object as created by \code{\link{ontology}}}

\item{allow.non.overlapping.incomplete.dimensions}{Logical indicating whether, when we map to an incomplete dimension, there can be no overlap in the from and to values for that dimension}

\item{print.ontologies}{Whether to print the full ontologies in the details}
}
\description{
For use in figuring out why a call to \code{\link{get.ontology.mapping}} did not work

If no mapping can be found, prints detailed information on why we were unable to find a mapping
}
