% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{examine.recent.failed.ontology.mapping}
\alias{examine.recent.failed.ontology.mapping}
\title{Get Detailed Information on a Recent Failed Attempt to Map Ontologies}
\usage{
examine.recent.failed.ontology.mapping(which = 1, print.ontologies = T)
}
\arguments{
\item{which}{The index of the recent failure. 1 denotes the most recent failed attempt, 2 the second most recent, and so on}

\item{print.ontologies}{Whether to print the full ontologies in the details}
}
\description{
Prints detailed information on why we were unable to find a mapping
}
