% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS_plot_simulations.R
\name{execute.plotly.plot}
\alias{execute.plotly.plot}
\title{Execute Plotly Plot
A utility that performs the plot-rendering half of the plot simulations in plotly operation
for applications like webtools that need the step to be separate.}
\usage{
execute.plotly.plot(
  prepared.plot.data,
  outcomes = NULL,
  split.by = NULL,
  facet.by = NULL,
  plot.which = "sim.and.data",
  summary.type = c("individual.simulation", "mean.and.interval",
    "median.and.interval")[1],
  plot.year.lag.ratio = F,
  n.facet.rows = NULL,
  style.manager = get.default.style.manager(),
  debug = F,
  hide.legend = FALSE
)
}
\description{
Execute Plotly Plot
A utility that performs the plot-rendering half of the plot simulations in plotly operation
for applications like webtools that need the step to be separate.
}
