% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS_simplot.R
\name{execute.simplot}
\alias{execute.simplot}
\title{Execute Simplot
A utility that performs the plot-rendering half of the simplot operation
for applications like webtools that need the step to be separate.}
\usage{
execute.simplot(
  prepared.plot.data,
  outcomes = NULL,
  split.by = NULL,
  facet.by = NULL,
  plot.which = c("sim.and.data", "sim.only", "data.only")[1],
  summary.type = c("individual.simulation", "mean.and.interval",
    "median.and.interval")[1],
  plot.year.lag.ratio = F,
  n.facet.rows = NULL,
  style.manager = get.default.style.manager(),
  debug = F
)
}
\arguments{
\item{prepared.plot.data}{The named list output by "prepare.plot"}

\item{outcomes}{A character vector of which simulation outcomes to plot}

\item{split.by}{At most one dimension}

\item{facet.by}{Any number of dimensions but cannot include the split.by dimension}

\item{plot.which}{Should simulation data and calibration data be plotted ('sim.and.data'), or only simulation data ('sim.only')}

\item{n.facet.rows}{An optional argument to specify how many rows the figure should facet plots into}

\item{style.manager}{We are going to have to define this down the road. It's going to govern how we do lines and sizes and colors. For now, just hard code those in, and we'll circle back to it}
}
\description{
Execute Simplot
A utility that performs the plot-rendering half of the simplot operation
for applications like webtools that need the step to be separate.
}
