% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_main.R
\name{extract.last.simulation.from.calibration}
\alias{extract.last.simulation.from.calibration}
\title{Extract the Last Simulation From A Calibration}
\usage{
extract.last.simulation.from.calibration(
  version,
  location,
  calibration.code,
  root.dir = get.jheem.root.directory("Cannot set up calibration: "),
  allow.incomplete = T,
  include.first.sim = F,
  chains = NULL
)
}
\arguments{
\item{version}{The version of the model specification (must have been previously registered)}

\item{location}{A single character value representing the location for the calibration}

\item{include.first.sim}{Whether to also pull the first simulation from the calibration (will then return a 2-simulation simset)}

\item{chains}{}
}
\description{
Extract the Last Simulation From A Calibration
}
