% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_main.R
\name{get.calibration.progress}
\alias{get.calibration.progress}
\title{Get Calibration Progress}
\usage{
get.calibration.progress(
  version,
  locations,
  calibration.code,
  root.dir = get.jheem.root.directory(),
  round.to.digits = ifelse(as.pct, 0, 2),
  as.pct = T
)
}
\arguments{
\item{version, calibration.code, locations}{The version, locations, and calibration code for which to gauge progress}

\item{root.dir}{The root directory from which to search for calibrations}

\item{round.to.digits}{How many digits to round progress to}

\item{as.pct}{Whether the progress should be shown as a percentage instead of a fraction}
}
\description{
Summarize the progress of a calibration task
}
