% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPERS_age_year_helpers.R
\name{get.cumulative.age.counts.smoother}
\alias{get.cumulative.age.counts.smoother}
\title{Get a function that projects a smoothed cumulative count up to a given age}
\usage{
get.cumulative.age.counts.smoother(
  counts,
  given.age.brackets,
  smooth.infinite.age.to = Inf,
  allow.extrapolation = F,
  method = c("monoH.FC", "hyman")[1],
  error.prefix = ""
)
}
\arguments{
\item{allow.extrapolation}{}
}
\description{
Get a function that produces a cubic spline over the *cumulative* counts by age up to a given age
}
