% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{get.data.outcome.pretty.names}
\alias{get.data.outcome.pretty.names}
\alias{get.data.outcome.labels}
\alias{get.data.outcome.descriptions}
\title{Get pretty names for outcomes}
\usage{
get.data.outcome.pretty.names(
  data.manager = get.default.data.manager(),
  outcomes
)

get.data.outcome.labels(data.manager = get.default.data.manager(), outcomes)

get.data.outcome.descriptions(
  data.manager = get.default.data.manager(),
  outcomes
)
}
\arguments{
\item{data.manager}{A jheem.data.manager object}

\item{outcomes}{A character vector with the names of outcomes previously registered to this data manager}
}
\value{
A character vector of pretty names, labels, or descriptions
}
\description{
Get pretty names for outcomes
}
\details{
Gets the pretty.names, labels, or descriptions registered for the outcomes with the data manager
}
\section{Functions}{
\itemize{
\item \code{get.data.outcome.labels()}: Get the labels for outcomes

\item \code{get.data.outcome.descriptions()}: Get descriptions for outcomes

}}
