% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{get.data.source.full.names}
\alias{get.data.source.full.names}
\alias{get.data.source.short.names}
\title{Get full names for data sources}
\usage{
get.data.source.full.names(data.manager = get.default.data.manager(), sources)

get.data.source.short.names(data.manager = get.default.data.manager(), sources)
}
\arguments{
\item{data.manager}{A jheem.data.manager object}

\item{sources}{A character vector with the names of sources previously registered to this data manager}
}
\value{
A character vector of full names or short names
}
\description{
Get full names for data sources
}
\details{
Gets the full.names or short.names for the data sources in the data manager
}
\section{Functions}{
\itemize{
\item \code{get.data.source.short.names()}: Get short names for data sources

}}
