% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS_style_manager.R
\name{get.default.style.manager}
\alias{get.default.style.manager}
\title{Get the Default Style Manager}
\usage{
get.default.style.manager(type = c("ggplot", "plotly")[1])
}
\description{
Get the Default Style Manager
}
