% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPERS_array_helpers.R
\name{get.expand.array.indices}
\alias{get.expand.array.indices}
\title{Get the Array Indices Under Expanded Dim-names}
\usage{
get.expand.array.indices(to.expand.dim.names, target.dim.names, index.from = 1)
}
\description{
Get the Array Indices Under Expanded Dim-names
}
