% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPERS_array_helpers.R
\name{get.expand.then.collapse.array.indices}
\alias{get.expand.then.collapse.array.indices}
\title{Get a Set of Indices to Collapse One Array into Another, expanding the first array if needed to add dimensions}
\usage{
get.expand.then.collapse.array.indices(
  large.arr.dim.names,
  small.arr.dim.names
)
}
\arguments{
\item{large.arr.dim.names}{The dimnames of the source array}

\item{small.arr.dim.names}{the dimnames of the target array}
}
\value{
A list with four elements: $large.indices, $small.indices, $large.n, $small.n
 large.indices and small.indices are integer vectors of equal length, such that the value at large.indices[i] goes into the value at small.indices[i] if collapsing
}
\description{
Get a Set of Indices to Collapse One Array into Another, expanding the first array if needed to add dimensions
}
