% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{get.identity.ontology.mapping}
\alias{get.identity.ontology.mapping}
\title{Get an "Identity" Ontology Mapping (that does not change data structure when applied)}
\usage{
get.identity.ontology.mapping()
}
\value{
An object of class "ontology.mapping"
}
\description{
Get an "Identity" Ontology Mapping (that does not change data structure when applied)
}
\seealso{
Other Functions to identify why ontologies fail to map: 
\code{\link{apply.ontology.mapping}()},
\code{\link{map.value.ontology}()}
}
\concept{Functions to identify why ontologies fail to map}
