% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILE_MANAGER_file_manager.R
\name{get.jheem.root.directory}
\alias{get.jheem.root.directory}
\title{Get JHEEM Root Directory
#' @description
Gets the directory that calibration directories and files will be written into,
if it has been set with "set.jheem.root.directory"}
\usage{
get.jheem.root.directory(error.prefix = "")
}
\description{
Get JHEEM Root Directory
#' @description
Gets the directory that calibration directories and files will be written into,
if it has been set with "set.jheem.root.directory"
}
