% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_main.R
\name{get.likelihood.instructions}
\alias{get.likelihood.instructions}
\title{Get Likelihood Instructions}
\usage{
get.likelihood.instructions(
  code,
  error.prefix = "Error getting likelihood instructions"
)
}
\arguments{
\item{code}{A valid code containing only letters, numbers, dashes, and periods.}

\item{error.prefix}{(Optional) A message to prepend to error messages.}
}
\description{
Get Likelihood Instructions
}
