% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{get.locations.with.data}
\alias{get.locations.with.data}
\title{Get locations with data for an outcome}
\usage{
get.locations.with.data(
  data.manager = get.default.data.manager(),
  outcome,
  metric = "estimate",
  years,
  exclude.ontology.names = NULL
)
}
\arguments{
\item{outcome}{The outcome for which to find earliest and latest years of data (must be registered in the data manager)}

\item{metric}{The type of measurement to find earliest and latest years of data within the outcome. The default value is "estimate", but other options include "cv", "variance", and "sd".}

\item{years}{The years to include. May be NULL to search all available years, or a numeric vector with no NA's or duplicates.}

\item{exclude.ontology.names}{The names of ontologies to exclude from this search. Must be NULL or a character vector with no NA's or empty values.}
}
\description{
Get locations with data for an outcome
}
\details{
Get locations that have data for a given outcome, metric and set of years. Searches all sources and all ontologies that haven't been specifcally excluded
}
