% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{get.mappings.to.align.ontologies}
\alias{get.mappings.to.align.ontologies}
\title{Get a Pair of Ontology Mappings that Aligns two Data Elements}
\usage{
get.mappings.to.align.ontologies(
  ontology.1,
  ontology.2,
  align.on.dimensions = intersect(names(ontology.1), names(ontology.2)),
  allow.non.overlapping.incomplete.dimensions = F,
  include.dim.names = NULL
)
}
\arguments{
\item{ontology.1, ontology.2}{The ontologies of two data elements to be transformed}

\item{align.on.dimensions}{The dimensions which should match in the transformed data}

\item{allow.non.overlapping.incomplete.dimensions}{Logical indicating whether, when we map to an incomplete dimension, there can be no overlap in the from and to values for that dimension}

\item{include.dim.names}{Optional argument, specifying specific dimension values that must be present in the transformed data}
}
\value{
Either a (1) NULL - if no mappings could align the ontologies or (2) a list with two elements, each an ontology.mapping object, such that applying the first to data conforming to ontology.1 and applying the second to data conforming to ontology.2 yields two data objects with the same dim.names
}
\description{
Get a Pair of Ontology Mappings that Aligns two Data Elements
}
