% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_simulation.R
\name{get.mcmc.mixing.statistic}
\alias{get.mcmc.mixing.statistic}
\title{Get MCMC Mixing Statistic}
\usage{
get.mcmc.mixing.statistic(
  simset,
  match.names = NULL,
  chains = simset$unique.chains,
  sort = T
)
}
\arguments{
\item{match.names}{(Optional) A regex on parameter names for which to compute the mixing statistic. If NULL, all parameters are used.}

\item{chains}{(Optional) Which chains to use, defaulting to all chains.}

\item{sort}{(Optional) Whether to sort the output in decreasing order by mixing statistic value, defaulting to TRUE.}
}
\description{
For one chain, the mixing statistic is the variance in the first
quarter of simulations divided by the variance in the last quarter of simulations.
}
