% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{get.ontology.error.debug.info}
\alias{get.ontology.error.debug.info}
\title{Get Ontology Error Debug Info}
\usage{
get.ontology.error.debug.info()
}
\description{
Returns information useful for debugging ontology mapping errors
}
