% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_nested_proportion_likelihood.R
\name{get.p.bias.estimates}
\alias{get.p.bias.estimates}
\title{Get P Bias Estimates}
\usage{
get.p.bias.estimates(
  data.manager = get.default.data.manager(),
  dimensions,
  levels.of.stratification,
  outcome.for.p,
  outcome.for.n,
  sub.location.type,
  super.location.type,
  main.location.type = "CBSA",
  minimum.sample.size = 12,
  from.year = NULL,
  to.year = NULL,
  main.location.type.p.source = NULL,
  sub.location.type.p.source = NULL,
  super.location.type.p.source = NULL,
  main.location.type.n.source = NULL,
  sub.location.type.n.source = NULL,
  super.location.type.n.source = NULL,
  debug = F
)
}
\arguments{
\item{data.manager}{The name of the data manager}

\item{dimensions}{A character vector of dimensions, excluding year, from which stratifications will be generated.}

\item{levels.of.stratification}{An integer vector specifying how the dimensions should be combined to form strata. '0' indicates totals (not stratified) while '1' indicates strata that are each stratified by one dimension at a time, '2' indicates strat that are each stratified by a combination of two dimensions at a time, etc. May not exceed the number of dimensions. Defaults to NULL, which is equivalent to '0'.}

\item{outcome.for.p, outcome.for.n}{Outcome names for proportion and denominator data, respectively, as found in the data manager}

\item{sub.location.type}{Used to find p-bias "inside" estimates by comparing data for this type against that of the main location type. If NULL, no p-bias "inside" estimates are found.}

\item{super.location.type}{Used to find p-bias "outside" estimates by comparing data for this type against that of the main location type. If NULL, no p-bias "outside" estimates are found.}

\item{main.location.type}{The location type that serves as the basis of comparison.}

\item{minimum.sample.size}{Numeric value specifying the minimum number of sub-main location pairs (or super-main location pairs) that must be found in order to calculate p-bias estimates}

\item{from.year, to.year}{Both NULL or both single numeric values}
}
\value{
A list containing mean and sd for p-bias "inside" and p-bias "outside" estimates. Some or all of these may be NA if 'sub.location.type' and/or 'super.location.type' were left NULL.
}
\description{
Get P Bias Estimates
}
\seealso{
Other Nested Proportion Likelihoods: 
\code{\link{create.nested.proportion.likelihood.instructions}()},
\code{\link{create.nested.proportion.likelihood.instructions.with.included.multiplier}()},
\code{\link{create.time.lagged.comparison.nested.proportion.likelihood.instructions}()}
}
\concept{Nested Proportion Likelihoods}
