% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERSIONS_version_manager.R
\name{get.prior.versions}
\alias{get.prior.versions}
\title{Get Prior JHEEM Model Versions}
\usage{
get.prior.versions(version, recursive = T)
}
\description{
Get Prior JHEEM Model Versions
}
\seealso{
Other JHEEM Version Management Functions: 
\code{\link{get.specification.for.version}()},
\code{\link{register.model.specification}()},
\code{\link{version.has.been.registered}()}
}
\concept{JHEEM Version Management Functions}
