% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{get.registered.ontology}
\alias{get.registered.ontology}
\title{param ontology.name The name of an ontology (must be registered in the data manager)}
\usage{
get.registered.ontology(
  data.manager = get.default.data.manager(),
  ontology.name
)
}
\description{
param ontology.name The name of an ontology (must be registered in the data manager)
}
