% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_simulation.R
\name{get.simset.data}
\alias{get.simset.data}
\title{Get Simulation Data}
\usage{
get.simset.data(
  simset,
  outcomes,
  output = c("value", "numerator", "denominator")[[1]],
  keep.dimensions = NULL,
  dimension.values = list(),
  ...,
  check.consistency = T,
  drop.single.outcome.dimension = T,
  error.prefix = "Error getting simulation results: "
)
}
\arguments{
\item{outcomes}{A character vector with one or more outcomes for which to pull data. Must be a subset of sim$outcomes}

\item{keep.dimensions}{Which dimensions should be retained in the resulting array}

\item{dimension.values}{A set of dimension values according to which to subset the data (will be combined with ...)}

\item{...}{The parameters in ... are interpreted as additional dimension values according to which to subset the data. Each must be named, and either a character, integer, or logical vector}

\item{check.consistency}{Whether to check the consistency of arguments. Setting to FALSE will yield faster performance but unintelligible error messages}

\item{error.prefix}{A character value to prepend to any errors generating in getting these data}

\item{drop.single.outcome.dimensions}{When set to FALSE, the resulting array will have 'outcome' as its last dimension regardless of how many outcomes we are getting data for. If TRUE, the returned array will omit that dimension when only one outcome is requested}
}
\description{
Get Simulation Data
}
\details{
Returns an array whose dimensions are keep.dimensions - plus (if there is more than one outcome or drop.single.outcome.dimensions==FALSE) an 'outcome' dimension at the end
}
