% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILE_MANAGER_file_manager.R
\name{get.simset.filename.and.dir}
\alias{get.simset.filename.and.dir}
\title{Get Simset Filename and Directory}
\usage{
get.simset.filename.and.dir(
  version,
  location,
  calibration.code,
  n.sim,
  intervention.code = NULL,
  sub.version = NULL,
  root.dir = get.jheem.root.directory("Cannot get.simset.filename(): "),
  error.prefix = ""
)
}
\description{
Get Simset Filename and Directory
}
