% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILE_MANAGER_file_manager.R
\name{get.simset.filename}
\alias{get.simset.filename}
\title{Get Simset Filename}
\usage{
get.simset.filename(
  version,
  location,
  calibration.code,
  n.sim,
  intervention.code = NULL,
  sub.version = NULL,
  include.path = T,
  root.dir = get.jheem.root.directory("Cannot get.simset.filename(): "),
  error.prefix = ""
)
}
\description{
Get Simset Filename
}
