% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_simulation.R
\name{get.simset.params}
\alias{get.simset.params}
\title{Get Simulation Set Paramater Values}
\usage{
get.simset.params(
  simset,
  match.names = NULL,
  simulation.indices = simset$n.sim,
  drop = T
)
}
\arguments{
\item{simulation.indices}{An integer or logical vector specifying which simulations to show in output. Defaults to the last simulation.}

\item{drop}{If the default TRUE, then an output with only one simulation will be coerced to a vector.}
}
\description{
Get Simulation Set Paramater Values
}
