% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_simulation.R
\name{get.simulation.metadata}
\alias{get.simulation.metadata}
\title{Get a Simulation Metadata Object}
\usage{
get.simulation.metadata(
  version,
  location,
  from.year = NULL,
  to.year = NULL,
  n.sim = 1,
  sub.version = NULL,
  error.prefix = paste0("Error deriving the simulation-metadata for '", version,
    "' and location '", location, "': ")
)
}
\arguments{
\item{version}{The version of the model specification (must have been previously registered) for which to get metadata}

\item{location}{A single character value representing the location for the metadata}

\item{from.year, to.year}{The years which a corresponding simulation will have data for}

\item{error.prefix}{A string to prepend to any errors generated in getting the metadata object}
}
\description{
Get a Simulation Metadata Object
}
\details{
A simulation.metadata object contains metadata, such as the dim.names to which its contents will conform
}
