% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_entity.R
\name{get.specification.info}
\alias{get.specification.info}
\title{Get the Specification-Info Object for an Instance of a JHEEM Entity}
\usage{
get.specification.info(jheem.entity)
}
\arguments{
\item{jheem.entity}{An object of class jheem.entity (such as a jheem.engine, jheem.simset, or jheem.sim)}
}
\value{
An object of class 'jheem.specification.info'
}
\description{
Get the Specification-Info Object for an Instance of a JHEEM Entity
}
