% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_metadata.R
\name{get.specification.metadata}
\alias{get.specification.metadata}
\title{Get a Specification Metadata Object}
\usage{
get.specification.metadata(version, location, error.prefix = "")
}
\arguments{
\item{version}{The version of the model specification (must have been previously registered) for which to get metadata}

\item{location}{A single character value representing the location for the metadata}

\item{error.prefix}{A string to prepend to any errors generated in getting the metadata object}
}
\description{
Get a Specification Metadata Object
}
\details{
A specification.metadata object contains metadata, such as dim.names of quantities, for a specification
}
