% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{get.year.bounds.for.outcome}
\alias{get.year.bounds.for.outcome}
\title{Get year bounds for outcome}
\usage{
get.year.bounds.for.outcome(
  data.manager = get.default.data.manager(),
  outcome,
  metric = "estimate",
  exclude.ontology.names = NULL
)
}
\arguments{
\item{outcome}{The outcome for which to find earliest and latest years of data (must be registered in the data manager)}

\item{metric}{The type of measurement to find earliest and latest years of data within the outcome. The default value is "estimate", but other options include "cv", "variance", and "sd".}

\item{exclude.ontology.names}{The names of ontologies to exclude from this search. Must be NULL or a character vector with no NA's or empty values.}
}
\description{
Get year bounds for outcome
}
\details{
Gets the earliest and latest years for which a data manager has data for a given outcome and metric. Searches all sources and all ontologies that haven't been specifically excluded
}
