% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{import.data}
\alias{import.data}
\title{Import data from another JHEEM Data Manager}
\usage{
import.data(to.data.manager = get.default.data.manager(), from.data.manager)
}
\arguments{
\item{to.data.manager}{A jheem.data.manager object}

\item{from.data.manager}{A jheem.data.manager object}
}
\description{
Import data from another JHEEM Data Manager
}
