% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology.R
\name{incomplete.dimensions.ontology}
\alias{incomplete.dimensions.ontology}
\title{Get an ontology's incomplete dimensions}
\usage{
\method{incomplete.dimensions}{ontology}(x)
}
\arguments{
\item{ontology}{An ontology, as created by \code{\link{ontology}}}
}
\value{
A vector containing the names of the ontology's incomplete dimensions
}
\description{
Get an ontology's incomplete dimensions
}
