% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology.R
\name{incomplete.dimensions}
\alias{incomplete.dimensions}
\title{Get an ontology's incomplete dimensions}
\usage{
incomplete.dimensions(x)
}
\value{
A vector containing the names of the ontology's incomplete dimensions
}
\description{
Get an ontology's incomplete dimensions
}
