% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{inspect.marginals}
\alias{inspect.marginals}
\title{Inspect marginal sums between stratifications}
\usage{
inspect.marginals(
  data.manager = get.default.data.manager,
  outcome = NULL,
  source = NULL,
  ontology = NULL,
  sub.stratification = NULL,
  super.stratification = NULL,
  years = NULL,
  locations = NULL,
  threshold.percent = 10,
  threshold.magnitude = 50,
  check.up.to.n.way.stratified = 1
)
}
\arguments{
\item{data.manager}{A jheem.data.manager object}

\item{outcome}{The outcome type for the data. Must be an outcome previously registered with \code{\link{register.data.outcome}}}

\item{source}{The (single) character name of the source from which these data derive. Must be a source previously registered with \code{\link{register.data.source}} Note: a source is conceived such that one source cannot contain two sets of data for the same set of dimension values}

\item{ontology}{The name of a registered ontology}

\item{sub.stratification}{The name of a less-aggregated stratification}

\item{super.stratification}{The name of a more-aggregated stratification}

\item{years, locations}{Show results for certain years (numeric, non-NA vector) or locations (character, non-NA vector)}

\item{threshold.percent}{A percentage below which discrepancies will not be reported (default 10).}

\item{threshold.magnitude}{A magnitude of difference between marginal sums and more-aggregated data below which discrepancies will not be reported (default 50).}

\item{check.up.to.n.way.stratified}{An integer greater than 0 indicating the highest level of stratification to report marginal sum discrepancies for (default 1).}
}
\description{
Compares the percentage difference between the marginal sums of 
more stratified data and the data it corresponds to in more aggregated 
stratifications of the same source and ontology.
}
