% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_main.R
\name{instantiate.likelihood}
\alias{instantiate.likelihood}
\title{Instantiate Likelihood}
\usage{
instantiate.likelihood(
  instructions,
  version,
  location,
  sub.version = NULL,
  data.manager = get.default.data.manager(),
  verbose = F,
  throw.error.if.no.data = F,
  error.prefix = "Error instantiating likelihood from instructions: "
)
}
\arguments{
\item{instructions}{A 'jheem.likelihood.instructions' object}

\item{data.manager}{A jheem.data.manager object. May be NULL for Bernoulli likelihood instructions.}

\item{verbose}{Should detailed messages be printed out as we go}

\item{throw.error.if.no.data}{Should an error be thrown if no relevant data is found?}
}
\description{
Instantiate Likelihood
}
