% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERVENTIONS_target_populations.R
\name{intersect.target.populations}
\alias{intersect.target.populations}
\title{Create a Target Population that is the INTERSECTION of Other Target Populations}
\usage{
intersect.target.populations(..., name = NULL)
}
\arguments{
\item{...}{One or more target.population objects or lists of target populations}

\item{name}{A short (<= 20 characters) descriptive name for the target population. Should be nicely formatted for public display}
}
\description{
Create a Target Population that is the INTERSECTION of Other Target Populations
}
\seealso{
Other Functions to create Target Populations: 
\code{\link{create.target.population}()},
\code{\link{diff.target.populations}()},
\code{\link{union.target.populations}()}
}
\concept{Functions to create Target Populations}
