% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology.R
\name{is.ontology}
\alias{is.ontology}
\title{Test if an object is of type 'ontology'}
\usage{
is.ontology(x)
}
\arguments{
\item{x}{Object to be tested}
}
\value{
A single logical T/F value
}
\description{
Test if an object is of type 'ontology'
}
