% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology.R
\name{is_complete.ontology}
\alias{is_complete.ontology}
\title{Get indicators of which dimensions in an ontology are complete}
\usage{
\method{is_complete}{ontology}(x)
}
\value{
A named logical vector - the corresponding to names(ontology) - indicating whether each dimension is complete
}
\description{
Get indicators of which dimensions in an ontology are complete
}
