% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology.R
\name{is_complete}
\alias{is_complete}
\alias{is.complete}
\title{Get indicators of which dimensions in an ontology are complete}
\usage{
is_complete(x)

is.complete(x)
}
\arguments{
\item{ontology}{An ontology, as created by \code{\link{ontology}}}
}
\value{
A named logical vector - the corresponding to names(ontology) - indicating whether each dimension is complete
}
\description{
Get indicators of which dimensions in an ontology are complete
}
