% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{JHEEM.ENGINE}
\alias{JHEEM.ENGINE}
\title{Still need to flesh out this interface}
\description{
Still need to flesh out this interface

Still need to flesh out this interface
}
\section{Super class}{
\code{jheem2::jheem.entity} -> \code{jheem.engine}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-jheem.engine-check}{\code{JHEEM.ENGINE$check()}}
\item \href{#method-jheem.engine-new}{\code{JHEEM.ENGINE$new()}}
\item \href{#method-jheem.engine-crunch}{\code{JHEEM.ENGINE$crunch()}}
\item \href{#method-jheem.engine-run}{\code{JHEEM.ENGINE$run()}}
\item \href{#method-jheem.engine-set.element.value}{\code{JHEEM.ENGINE$set.element.value()}}
\item \href{#method-jheem.engine-set.element.functional.form.main.effect.alphas}{\code{JHEEM.ENGINE$set.element.functional.form.main.effect.alphas()}}
\item \href{#method-jheem.engine-set.element.functional.form.interaction.alphas}{\code{JHEEM.ENGINE$set.element.functional.form.interaction.alphas()}}
\item \href{#method-jheem.engine-set.element.foreground}{\code{JHEEM.ENGINE$set.element.foreground()}}
\item \href{#method-jheem.engine-set.element.functional.form.from.time}{\code{JHEEM.ENGINE$set.element.functional.form.from.time()}}
\item \href{#method-jheem.engine-set.element.functional.form.to.time}{\code{JHEEM.ENGINE$set.element.functional.form.to.time()}}
\item \href{#method-jheem.engine-set.element.ramp.times}{\code{JHEEM.ENGINE$set.element.ramp.times()}}
\item \href{#method-jheem.engine-set.element.ramp.values}{\code{JHEEM.ENGINE$set.element.ramp.values()}}
\item \href{#method-jheem.engine-set.element.taper.times}{\code{JHEEM.ENGINE$set.element.taper.times()}}
\item \href{#method-jheem.engine-set.element.taper.values}{\code{JHEEM.ENGINE$set.element.taper.values()}}
\item \href{#method-jheem.engine-set.element.functional.form.future.slope}{\code{JHEEM.ENGINE$set.element.functional.form.future.slope()}}
\item \href{#method-jheem.engine-set.element.functional.form.future.slope.after.time}{\code{JHEEM.ENGINE$set.element.functional.form.future.slope.after.time()}}
\item \href{#method-jheem.engine-has.been.crunched}{\code{JHEEM.ENGINE$has.been.crunched()}}
\item \href{#method-jheem.engine-test.crunch}{\code{JHEEM.ENGINE$test.crunch()}}
\item \href{#method-jheem.engine-clone}{\code{JHEEM.ENGINE$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jheem2" data-topic="jheem.entity" data-id="get.filename"><a href='../../jheem2/html/jheem.entity.html#method-jheem.entity-get.filename'><code>jheem2::jheem.entity$get.filename()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jheem2" data-topic="jheem.entity" data-id="get.specification.info"><a href='../../jheem2/html/jheem.entity.html#method-jheem.entity-get.specification.info'><code>jheem2::jheem.entity$get.specification.info()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-check"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-check}{}}}
\subsection{Method \code{check()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$check()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-new"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$new(version, location)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-crunch"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-crunch}{}}}
\subsection{Method \code{crunch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$crunch(
  start.year,
  end.year,
  check.consistency = !self$has.been.crunched()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-run"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$run(
  start.year,
  end.year,
  check.consistency = !self$has.been.crunched(),
  max.run.time.seconds = Inf,
  prior.sim = NULL,
  keep.years = start.year:end.year,
  atol = 1e-04,
  rtol = 1e-04
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-set.element.value"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-set.element.value}{}}}
\subsection{Method \code{set.element.value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$set.element.value(
  element.name,
  value,
  check.consistency = !self$has.been.crunched(),
  error.prefix = ""
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-set.element.functional.form.main.effect.alphas"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-set.element.functional.form.main.effect.alphas}{}}}
\subsection{Method \code{set.element.functional.form.main.effect.alphas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$set.element.functional.form.main.effect.alphas(
  element.name,
  alpha.name,
  values,
  applies.to.dimension.values = names(values),
  dimensions = names(applies.to.dimension.values),
  check.consistency = !self$has.been.crunched()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-set.element.functional.form.interaction.alphas"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-set.element.functional.form.interaction.alphas}{}}}
\subsection{Method \code{set.element.functional.form.interaction.alphas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$set.element.functional.form.interaction.alphas(
  element.name,
  alpha.name,
  value,
  applies.to.dimension.values = names(values),
  dimensions = names(applies.to.dimension.values),
  check.consistency = !self$has.been.crunched()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-set.element.foreground"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-set.element.foreground}{}}}
\subsection{Method \code{set.element.foreground()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$set.element.foreground(
  element.name,
  foreground,
  check.consistency = !self$has.been.crunched()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-set.element.functional.form.from.time"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-set.element.functional.form.from.time}{}}}
\subsection{Method \code{set.element.functional.form.from.time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$set.element.functional.form.from.time(
  element.name,
  from.time,
  check.consistency = !self$has.been.crunched()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-set.element.functional.form.to.time"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-set.element.functional.form.to.time}{}}}
\subsection{Method \code{set.element.functional.form.to.time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$set.element.functional.form.to.time(
  element.name,
  to.time,
  check.consistency = !self$has.been.crunched()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-set.element.ramp.times"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-set.element.ramp.times}{}}}
\subsection{Method \code{set.element.ramp.times()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$set.element.ramp.times(
  element.name,
  times,
  indices = 1:length(times),
  check.consistency = !self$has.been.crunched()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-set.element.ramp.values"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-set.element.ramp.values}{}}}
\subsection{Method \code{set.element.ramp.values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$set.element.ramp.values(
  element.name,
  values,
  indices = 1:length(values),
  check.consistency = !self$has.been.crunched()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-set.element.taper.times"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-set.element.taper.times}{}}}
\subsection{Method \code{set.element.taper.times()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$set.element.taper.times(
  element.name,
  times,
  indices = 1:length(times),
  check.consistency = !self$has.been.crunched()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-set.element.taper.values"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-set.element.taper.values}{}}}
\subsection{Method \code{set.element.taper.values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$set.element.taper.values(
  element.name,
  values,
  indices = 1:length(values),
  check.consistency = !self$has.been.crunched()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-set.element.functional.form.future.slope"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-set.element.functional.form.future.slope}{}}}
\subsection{Method \code{set.element.functional.form.future.slope()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$set.element.functional.form.future.slope(
  element.names,
  slope,
  check.consistency = !self$has.been.crunched()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-set.element.functional.form.future.slope.after.time"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-set.element.functional.form.future.slope.after.time}{}}}
\subsection{Method \code{set.element.functional.form.future.slope.after.time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$set.element.functional.form.future.slope.after.time(
  element.names,
  after.time,
  check.consistency = !self$has.been.crunched()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-has.been.crunched"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-has.been.crunched}{}}}
\subsection{Method \code{has.been.crunched()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$has.been.crunched()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-test.crunch"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-test.crunch}{}}}
\subsection{Method \code{test.crunch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$test.crunch()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem.engine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-jheem.engine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM.ENGINE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
