% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{JHEEM}
\alias{JHEEM}
\title{Still need to flesh out this interface}
\description{
Still need to flesh out this interface

Still need to flesh out this interface
}
\section{Super class}{
\code{jheem2::jheem.entity} -> \code{jheem}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-jheem-check}{\code{JHEEM$check()}}
\item \href{#method-jheem-new}{\code{JHEEM$new()}}
\item \href{#method-jheem-crunch}{\code{JHEEM$crunch()}}
\item \href{#method-jheem-run}{\code{JHEEM$run()}}
\item \href{#method-jheem-crunch.for.transmutation}{\code{JHEEM$crunch.for.transmutation()}}
\item \href{#method-jheem-transmute}{\code{JHEEM$transmute()}}
\item \href{#method-jheem-test}{\code{JHEEM$test()}}
\item \href{#method-jheem-extract.quantity.values}{\code{JHEEM$extract.quantity.values()}}
\item \href{#method-jheem-extract.diffeq.settings}{\code{JHEEM$extract.diffeq.settings()}}
\item \href{#method-jheem-set.element.value}{\code{JHEEM$set.element.value()}}
\item \href{#method-jheem-set.element.functional.form.main.effect.alphas}{\code{JHEEM$set.element.functional.form.main.effect.alphas()}}
\item \href{#method-jheem-set.element.functional.form.interaction.alphas}{\code{JHEEM$set.element.functional.form.interaction.alphas()}}
\item \href{#method-jheem-set.quantity.foreground}{\code{JHEEM$set.quantity.foreground()}}
\item \href{#method-jheem-set.parameters}{\code{JHEEM$set.parameters()}}
\item \href{#method-jheem-set.element.functional.form.from.time}{\code{JHEEM$set.element.functional.form.from.time()}}
\item \href{#method-jheem-set.element.functional.form.to.time}{\code{JHEEM$set.element.functional.form.to.time()}}
\item \href{#method-jheem-set.element.ramp.times}{\code{JHEEM$set.element.ramp.times()}}
\item \href{#method-jheem-set.element.ramp.values}{\code{JHEEM$set.element.ramp.values()}}
\item \href{#method-jheem-set.element.taper.times}{\code{JHEEM$set.element.taper.times()}}
\item \href{#method-jheem-set.element.taper.values}{\code{JHEEM$set.element.taper.values()}}
\item \href{#method-jheem-set.element.functional.form.future.slope}{\code{JHEEM$set.element.functional.form.future.slope()}}
\item \href{#method-jheem-set.element.functional.form.future.slope.after.time}{\code{JHEEM$set.element.functional.form.future.slope.after.time()}}
\item \href{#method-jheem-set.intervention}{\code{JHEEM$set.intervention()}}
\item \href{#method-jheem-set.calibration.code}{\code{JHEEM$set.calibration.code()}}
\item \href{#method-jheem-has.been.crunched}{\code{JHEEM$has.been.crunched()}}
\item \href{#method-jheem-clone}{\code{JHEEM$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jheem2" data-topic="jheem.entity" data-id="is.out.of.date"><a href='../../jheem2/html/jheem.entity.html#method-jheem.entity-is.out.of.date'><code>jheem2::jheem.entity$is.out.of.date()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-check"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-check}{}}}
\subsection{Method \code{check()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$check()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-new"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$new(
  jheem.kernel,
  sub.version,
  transmute.from.kernel,
  transmute.from.sub.version,
  error.prefix = "Cannot create JHEEM Instance: "
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-crunch"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-crunch}{}}}
\subsection{Method \code{crunch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$crunch(
  start.year,
  end.year,
  keep.from.year,
  keep.to.year,
  prior.simulation.set,
  prior.sim.index,
  solver.metadata,
  check.consistency = !self$has.been.crunched()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-run"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$run(
  start.year,
  end.year,
  check.consistency,
  max.run.time.seconds,
  prior.simulation.set,
  prior.sim.index,
  keep.from.year,
  keep.to.year,
  solver.metadata,
  finalize
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-crunch.for.transmutation"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-crunch.for.transmutation}{}}}
\subsection{Method \code{crunch.for.transmutation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$crunch.for.transmutation(
  keep.from.year,
  keep.to.year,
  prior.simulation.set,
  prior.sim.index,
  check.consistency = !self$has.been.crunched()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-transmute"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-transmute}{}}}
\subsection{Method \code{transmute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$transmute(
  prior.simulation.set,
  prior.sim.index,
  keep.from.year,
  keep.to.year,
  finalize,
  check.consistency
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-test"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-test}{}}}
\subsection{Method \code{test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$test()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-extract.quantity.values"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-extract.quantity.values}{}}}
\subsection{Method \code{extract.quantity.values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$extract.quantity.values()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-extract.diffeq.settings"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-extract.diffeq.settings}{}}}
\subsection{Method \code{extract.diffeq.settings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$extract.diffeq.settings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-set.element.value"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-set.element.value}{}}}
\subsection{Method \code{set.element.value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$set.element.value(
  element.name,
  value,
  check.consistency = !self$has.been.crunched(),
  wrt.version = self$version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-set.element.functional.form.main.effect.alphas"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-set.element.functional.form.main.effect.alphas}{}}}
\subsection{Method \code{set.element.functional.form.main.effect.alphas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$set.element.functional.form.main.effect.alphas(
  element.name,
  alpha.name,
  values,
  dimension,
  applies.to.dimension.values = names(values),
  check.consistency = !self$has.been.crunched(),
  wrt.version = self$version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-set.element.functional.form.interaction.alphas"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-set.element.functional.form.interaction.alphas}{}}}
\subsection{Method \code{set.element.functional.form.interaction.alphas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$set.element.functional.form.interaction.alphas(
  element.name,
  alpha.name,
  value,
  applies.to.dimension.values,
  check.consistency = !self$has.been.crunched(),
  wrt.version = self$version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-set.quantity.foreground"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-set.quantity.foreground}{}}}
\subsection{Method \code{set.quantity.foreground()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$set.quantity.foreground(
  foreground,
  check.consistency = !self$has.been.crunched()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-set.parameters"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-set.parameters}{}}}
\subsection{Method \code{set.parameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$set.parameters(parameters, check.consistency, wrt.version = self$version)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-set.element.functional.form.from.time"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-set.element.functional.form.from.time}{}}}
\subsection{Method \code{set.element.functional.form.from.time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$set.element.functional.form.from.time(
  element.name,
  from.time,
  check.consistency = !self$has.been.crunched(),
  wrt.version = self$version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-set.element.functional.form.to.time"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-set.element.functional.form.to.time}{}}}
\subsection{Method \code{set.element.functional.form.to.time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$set.element.functional.form.to.time(
  element.name,
  to.time,
  check.consistency = !self$has.been.crunched(),
  wrt.version = self$version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-set.element.ramp.times"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-set.element.ramp.times}{}}}
\subsection{Method \code{set.element.ramp.times()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$set.element.ramp.times(
  element.name,
  times,
  indices = 1:length(times),
  check.consistency = !self$has.been.crunched(),
  wrt.version = self$version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-set.element.ramp.values"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-set.element.ramp.values}{}}}
\subsection{Method \code{set.element.ramp.values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$set.element.ramp.values(
  element.name,
  values,
  indices = 1:length(values),
  check.consistency = !self$has.been.crunched(),
  wrt.version = self$version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-set.element.taper.times"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-set.element.taper.times}{}}}
\subsection{Method \code{set.element.taper.times()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$set.element.taper.times(
  element.name,
  times,
  indices = 1:length(times),
  check.consistency = !self$has.been.crunched(),
  wrt.version = self$version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-set.element.taper.values"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-set.element.taper.values}{}}}
\subsection{Method \code{set.element.taper.values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$set.element.taper.values(
  element.name,
  values,
  indices = 1:length(values),
  check.consistency = !self$has.been.crunched(),
  wrt.version = self$version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-set.element.functional.form.future.slope"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-set.element.functional.form.future.slope}{}}}
\subsection{Method \code{set.element.functional.form.future.slope()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$set.element.functional.form.future.slope(
  element.names,
  slope,
  check.consistency = !self$has.been.crunched(),
  wrt.version = self$version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-set.element.functional.form.future.slope.after.time"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-set.element.functional.form.future.slope.after.time}{}}}
\subsection{Method \code{set.element.functional.form.future.slope.after.time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$set.element.functional.form.future.slope.after.time(
  element.names,
  after.time,
  check.consistency = !self$has.been.crunched(),
  wrt.version = self$version
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-set.intervention"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-set.intervention}{}}}
\subsection{Method \code{set.intervention()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$set.intervention(intervention.code)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-set.calibration.code"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-set.calibration.code}{}}}
\subsection{Method \code{set.calibration.code()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$set.calibration.code(calibration.code, outcome.location.mapping)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-has.been.crunched"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-has.been.crunched}{}}}
\subsection{Method \code{has.been.crunched()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$has.been.crunched()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jheem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-jheem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JHEEM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
