% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jheem2-parameter-documentation.R
\name{jheem2-data-manager-params}
\alias{jheem2-data-manager-params}
\title{JHEEM Data Manager Parameters}
\arguments{
\item{metric}{The type of measurement. The default value is "estimate", but other options include "cv", "variance", and "sd".}

\item{ont}{An ontology object as created by \code{\link{ontology()}}}

\item{outcomes}{A character vector with the names of outcomes previously registered to this data manager}

\item{sources}{A character vector with the names of sources previously registered to this data manager}
}
\description{
JHEEM Data Manager Parameters
}
