% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jheem2-parameter-documentation.R
\name{jheem2-likelihood-params}
\alias{jheem2-likelihood-params}
\title{JHEEM Likelihood Parameters}
\arguments{
\item{log}{Whether to use log likelihood}

\item{instructions}{A 'jheem.likelihood.instructions' object}
}
\description{
JHEEM Likelihood Parameters
}
