% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jheem2-parameter-documentation.R
\name{jheem2-params}
\alias{jheem2-params}
\title{JHEEM Parameters}
\arguments{
\item{version}{The version of the model specification (must have been previously registered)}

\item{data.manager}{A jheem.data.manager object}

\item{error.prefix}{(Optional) A message to prepend to error messages.}

\item{sim}{A 'jheem.simulation.set' object}

\item{simset}{A 'jheem.simulation.set' object}

\item{likelihood}{A 'jheem.likelihood' object}

\item{likelihood.instructions}{A 'jheem.likelihood.instructions' object}

\item{debug}{(dev argument)}
}
\description{
JHEEM Parameters
}
