% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jheem2-parameter-documentation.R
\name{jheem2-simset-params}
\alias{jheem2-simset-params}
\title{JHEEM Simulation Set Parameters}
\arguments{
\item{match.names}{A regex to select parameters by name. If NULL, all parameters are used.}

\item{chains}{Which chains to use, defaulting to all.}
}
\description{
JHEEM Simulation Set Parameters
}
