% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jheem2-package-documentation.R
\docType{package}
\name{jheem2}
\alias{jheem2-package}
\alias{jheem2}
\title{jheem2: A package for setting up and running simulations using the JHEEM Engine
#'}
\description{
The engine for the JHEEM (Johns Hopkins Epidemiological and Economic Model) family of compartmental models of HIV and STI transmission. Provides tools to specify versioned model structures, compile and run them as differential-equation simulations, manage surveillance data, map between data ontologies, define interventions, and calibrate models to data via likelihoods.
}
\section{jheem2 functions}{

The jheem2 functions ...
}

