% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERVENTIONS_main.R
\name{join.interventions}
\alias{join.interventions}
\title{Join Multiple Interventions into a Single Intervention}
\usage{
join.interventions(
  ...,
  code = NULL,
  name = NULL,
  parameter.distribution = NULL,
  overwrite.existing.intervention = F,
  sequential = F
)
}
\arguments{
\item{...}{One or more interventions to join. These may be either objects of class jheem.intervention or lists which contain only objects of class jheem.intervention}

\item{code}{A brief (2-30 character) string that will uniquely identify this intervention}

\item{parameter.distribution}{}

\item{sequential}{}
}
\description{
Join Multiple Interventions into a Single Intervention
}
