% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_joint_likelihood.R
\name{join.likelihood.instructions}
\alias{join.likelihood.instructions}
\title{Make a single (joint) set of likelihood instructions out of two or more likelihood instructions objects}
\usage{
join.likelihood.instructions(...)
}
\arguments{
\item{...}{One or more jheem.likelihood.instructions objects or lists which contain only jheem.likelihood.instructions objects #OR it could just be codes}
}
\description{
Make a single (joint) set of likelihood instructions out of two or more likelihood instructions objects
}
