% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_outcome_location_mappings.R
\name{join.outcome.location.mappings}
\alias{join.outcome.location.mappings}
\title{Join Outcome Location Mappings}
\usage{
join.outcome.location.mappings(mappings.to.join)
}
\arguments{
\item{mappings.to.join}{A list of outcome.location.mappings objects}
}
\description{
Join Outcome Location Mappings
}
