% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_simulation.R
\name{join.simulation.sets}
\alias{join.simulation.sets}
\title{Join multiple simulation sets into a single set}
\usage{
join.simulation.sets(...)
}
\arguments{
\item{...}{One or more jheem.simulation.set objects or lists that contain only jheem.simulation.set objects}
}
\value{
A jheem.simulation.set object
}
\description{
Join multiple simulation sets into a single set
}
