% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{load.data.manager}
\alias{load.data.manager}
\title{Load a cached data manager}
\usage{
load.data.manager(file, name = NULL, description = NULL, set.as.default = F)
}
\arguments{
\item{file}{A file storing a cached data manager object (only)}

\item{name}{A name for the local data manager. If NULL, the name of the cached data manager.}

\item{description}{A description for the local data manager. If NULL, the description of the cached data manager.}
}
\description{
Load a cached data manager
}
