% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_simulation.R
\name{load.simulation.set}
\alias{load.simulation.set}
\title{Load a Simulation Set}
\usage{
load.simulation.set(file)
}
\arguments{
\item{file}{The file containing the saved simulation set}
}
\description{
Load a Simulation Set
}
