% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPERS_age_race_helpers.R,
%   R/HELPERS_age_year_helpers.R
\name{make.age.strata.names}
\alias{make.age.strata.names}
\title{Make Names for a Set of Age Strata}
\usage{
make.age.strata.names(endpoints = NULL, lowers = NULL, uppers = NULL)

make.age.strata.names(endpoints = NULL, lowers = NULL, uppers = NULL)

make.age.strata.names(endpoints = NULL, lowers = NULL, uppers = NULL)
}
\arguments{
\item{endpoints}{A numeric vector of at least two points. endpoints[1] is the lower bound (inclusive) of the first stratum, endpoints[2] is the upper bound (exclusive) for the first stratum and the lower bound for the second stratum, etc.}
}
\value{
A character vector with length(endpoints)-1 values

A character vector with length(endpoints)-1 values

A character vector with length(endpoints)-1 values
}
\description{
Make Names for a Set of Age Strata

Make Names for a Set of Age Strata

Make Names for a Set of Age Strata
}
