% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{map.value.ontology}
\alias{map.value.ontology}
\title{Map Value's Ontology}
\usage{
map.value.ontology(
  value,
  target.dim.names,
  fun = "sum",
  allow.expand.values = T,
  allow.map.to.subset.of.target = T,
  allow.restratify.ages = T,
  allow.age.extrapolation = F,
  na.rm = T,
  smooth.infinite.age.to = 101,
  error.prefix = ""
)
}
\arguments{
\item{allow.map.to.subset.of.target}{}

\item{error.prefix}{}
}
\description{
Map Value's Ontology
}
\seealso{
Other Functions to identify why ontologies fail to map: 
\code{\link{apply.ontology.mapping}()},
\code{\link{get.identity.ontology.mapping}()}
}
\concept{Functions to identify why ontologies fail to map}
