% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_simulation.R
\name{mcmc.traceplot}
\alias{mcmc.traceplot}
\title{Create Traceplot of MCMC}
\usage{
mcmc.traceplot(simset, match.names, chains = simset$unique.chains, burn = 0)
}
\arguments{
\item{burn}{How many iterations to burn and thus ignore on the plot, defaulting to 0.}
}
\description{
Create Traceplot of MCMC
}
