% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{ONTOLOGY.MAPPING}
\alias{ONTOLOGY.MAPPING}
\title{The Constructor}
\description{
The Constructor

The Constructor

The Print Function

A test for equality - to be overwritten in subclass --#

Test if this ontology.mapping can apply to the given dim.names

Test if this ontology.mapping can apply to the given ontology

Get the dimnames that would be generated after applying this mapping to some data with the given dim.names

Get the ontology that would be generated after applying this mapping to some data
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ontology.mapping-new}{\code{ONTOLOGY.MAPPING$new()}}
\item \href{#method-ontology.mapping-print}{\code{ONTOLOGY.MAPPING$print()}}
\item \href{#method-ontology.mapping-equals}{\code{ONTOLOGY.MAPPING$equals()}}
\item \href{#method-ontology.mapping-can.apply.to.dim.names}{\code{ONTOLOGY.MAPPING$can.apply.to.dim.names()}}
\item \href{#method-ontology.mapping-can.apply.to.ontology}{\code{ONTOLOGY.MAPPING$can.apply.to.ontology()}}
\item \href{#method-ontology.mapping-apply.to.dim.names}{\code{ONTOLOGY.MAPPING$apply.to.dim.names()}}
\item \href{#method-ontology.mapping-apply.to.ontology}{\code{ONTOLOGY.MAPPING$apply.to.ontology()}}
\item \href{#method-ontology.mapping-apply}{\code{ONTOLOGY.MAPPING$apply()}}
\item \href{#method-ontology.mapping-get.matrix}{\code{ONTOLOGY.MAPPING$get.matrix()}}
\item \href{#method-ontology.mapping-get.mapping.indices}{\code{ONTOLOGY.MAPPING$get.mapping.indices()}}
\item \href{#method-ontology.mapping-get.reverse.mapping.indices}{\code{ONTOLOGY.MAPPING$get.reverse.mapping.indices()}}
\item \href{#method-ontology.mapping-get.required.from.dim.names}{\code{ONTOLOGY.MAPPING$get.required.from.dim.names()}}
\item \href{#method-ontology.mapping-get.required.from.dimensions}{\code{ONTOLOGY.MAPPING$get.required.from.dimensions()}}
\item \href{#method-ontology.mapping-clone}{\code{ONTOLOGY.MAPPING$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology.mapping-new"></a>}}
\if{latex}{\out{\hypertarget{method-ontology.mapping-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ONTOLOGY.MAPPING$new(name, component.names = name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology.mapping-print"></a>}}
\if{latex}{\out{\hypertarget{method-ontology.mapping-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ONTOLOGY.MAPPING$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology.mapping-equals"></a>}}
\if{latex}{\out{\hypertarget{method-ontology.mapping-equals}{}}}
\subsection{Method \code{equals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ONTOLOGY.MAPPING$equals(other)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology.mapping-can.apply.to.dim.names"></a>}}
\if{latex}{\out{\hypertarget{method-ontology.mapping-can.apply.to.dim.names}{}}}
\subsection{Method \code{can.apply.to.dim.names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ONTOLOGY.MAPPING$can.apply.to.dim.names(
  from.dim.names,
  to.dim.names = NULL,
  throw.errors = F,
  error.prefix = ""
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology.mapping-can.apply.to.ontology"></a>}}
\if{latex}{\out{\hypertarget{method-ontology.mapping-can.apply.to.ontology}{}}}
\subsection{Method \code{can.apply.to.ontology()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ONTOLOGY.MAPPING$can.apply.to.ontology(
  from.ontology,
  to.ontology = NULL,
  throw.errors = F,
  error.prefix = ""
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology.mapping-apply.to.dim.names"></a>}}
\if{latex}{\out{\hypertarget{method-ontology.mapping-apply.to.dim.names}{}}}
\subsection{Method \code{apply.to.dim.names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ONTOLOGY.MAPPING$apply.to.dim.names(
  from.dim.names,
  error.prefix = "Cannot apply ontology.mapping to dim names: "
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology.mapping-apply.to.ontology"></a>}}
\if{latex}{\out{\hypertarget{method-ontology.mapping-apply.to.ontology}{}}}
\subsection{Method \code{apply.to.ontology()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ONTOLOGY.MAPPING$apply.to.ontology(
  ontology,
  error.prefix = "Cannot apply ontology.mapping to ontology: "
)}\if{html}{\out{</div>}}
}

\subsection{Details}{
An alias of the apply.to.dim.names function (which appropriately handles if dim.names is additionally an ontology)
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology.mapping-apply"></a>}}
\if{latex}{\out{\hypertarget{method-ontology.mapping-apply}{}}}
\subsection{Method \code{apply()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ONTOLOGY.MAPPING$apply(
  from.arr,
  to.dim.names = NULL,
  fun = "sum",
  na.rm = F,
  error.prefix = "Cannot apply ontology.mapping: "
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology.mapping-get.matrix"></a>}}
\if{latex}{\out{\hypertarget{method-ontology.mapping-get.matrix}{}}}
\subsection{Method \code{get.matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ONTOLOGY.MAPPING$get.matrix(
  from.dim.names,
  to.dim.names,
  error.prefix = "Cannot get matrix from ontology.mapping: "
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology.mapping-get.mapping.indices"></a>}}
\if{latex}{\out{\hypertarget{method-ontology.mapping-get.mapping.indices}{}}}
\subsection{Method \code{get.mapping.indices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ONTOLOGY.MAPPING$get.mapping.indices(
  from.dim.names,
  to.dim.names,
  error.prefix = "Cannot get mapping indices from ontology.mapping: "
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list, with length the same as the length that would be generated by
applying the ontology mapping to data with the given from.dim.names
The ith element of the return value is an integer vector containing
the indices into the from data which are combined
ie - to[i] would be equal to fun(from[ get.mapping.indices()[[i]] ])
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology.mapping-get.reverse.mapping.indices"></a>}}
\if{latex}{\out{\hypertarget{method-ontology.mapping-get.reverse.mapping.indices}{}}}
\subsection{Method \code{get.reverse.mapping.indices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ONTOLOGY.MAPPING$get.reverse.mapping.indices(
  from.dim.names,
  to.dim.names,
  error.prefix = "Cannot get mapping indices from ontology.mapping: "
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology.mapping-get.required.from.dim.names"></a>}}
\if{latex}{\out{\hypertarget{method-ontology.mapping-get.required.from.dim.names}{}}}
\subsection{Method \code{get.required.from.dim.names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ONTOLOGY.MAPPING$get.required.from.dim.names(to.dim.names)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology.mapping-get.required.from.dimensions"></a>}}
\if{latex}{\out{\hypertarget{method-ontology.mapping-get.required.from.dimensions}{}}}
\subsection{Method \code{get.required.from.dimensions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ONTOLOGY.MAPPING$get.required.from.dimensions(to.dimensions)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology.mapping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ontology.mapping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ONTOLOGY.MAPPING$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
