% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology.R
\name{ontology}
\alias{ontology}
\title{Create an ontology object}
\usage{
ontology(..., incomplete.dimensions = NULL)
}
\arguments{
\item{...}{One or more character vectors which must be NAMED}

\item{incomplete.dimensions}{A character vector indicating which dimensions are incomplete. Must be a subset of the names of ...}
}
\description{
Create an ontology object
}
\details{
An ontology is a set of dimnames with specific characteristics: (1) the dimnames must be named, (2) some of the dimensions may be noted as "incomplete". An incomplete dimension is one for which the given values for a dimension may not represent all possible values for that dimension
}
