% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPERS_dim_names_helpers.R
\name{outer.join.dim.names}
\alias{outer.join.dim.names}
\title{Join multiple dim.names lists}
\usage{
outer.join.dim.names(..., error.prefix = "")
}
\arguments{
\item{...}{can be either (1) dim.names or (2) lists of dim.names}
}
\value{
A named list, with an element for each name that appears in any of the dim.names in ... - rv[[d]] is the union of all values in the d dimension of any dim.names in ...
}
\description{
Join multiple dim.names lists
}
