% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPERS_age_race_helpers.R
\name{parse.age.strata.names}
\alias{parse.age.strata.names}
\title{Convert Age Strata Names into Lower and Upper Bounds for Each Stratum}
\usage{
parse.age.strata.names(strata.names)

parse.age.strata.names(strata.names)
}
\arguments{
\item{strata.names}{Names of age brackets. The function knows how to parse names generated in the format given by \link{make.age.strata.names}, as well as some other common formats}
}
\value{
A list with two elements, $lowers and $uppers, representing the lower (inclusive) and upper (exclusive) bounds of each age stratum

A list with two elements, $lowers and $uppers, representing the lower (inclusive) and upper (exclusive) bounds of each age stratum
}
\description{
Convert Age Strata Names into Lower and Upper Bounds for Each Stratum

Convert Age Strata Names into Lower and Upper Bounds for Each Stratum
}
