% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILE_MANAGER_file_manager.R
\name{parse.simset.filename}
\alias{parse.simset.filename}
\title{Parse Simset Filename}
\usage{
parse.simset.filename(filename, throw.error.if.malformed = T)
}
\description{
Parses the calibration information that was compressed into the filename of a
properly saved calibration simset.
}
