% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_main.R
\name{percentage.cache.complete}
\alias{percentage.cache.complete}
\title{Get Percentage of Run Cache Completed}
\usage{
percentage.cache.complete(version, calibration.code, locations, chains)
}
\value{
Matrix of percentages shape [chain x location]
}
\description{
Get Percentage of Run Cache Completed
}
