% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS_simplot.R
\name{plot.data.validation}
\alias{plot.data.validation}
\title{Cleaning and verifying the plotting data so it can be properly plotted}
\usage{
\method{plot}{data.validation}(
  simset.args,
  deparsed.substituted.args.simset.args,
  outcomes,
  corresponding.data.outcomes,
  plot.which,
  summary.type,
  simset.names = NULL
)
}
\arguments{
\item{simset.args}{A list of the ... parameter in the previous function, containing simset data.  Will be processed and returned as $simset.list}

\item{deparsed.substituted.args.simset.args}{The ... parameter with each argument converted to a character exactly as it is, without evaluating any expressions. Used to give un-named simsets names later.}

\item{outcomes}{The outcomes asked for by the plotting function.  Will be processed and returned as $outcomes}

\item{corresponding.data.outcomes}{Checked for proper value, not passed out of the function}

\item{plot.which}{Checked for proper value, not passed out of the function}

\item{summary.type}{Checked for proper value, not passed out of the function}
}
\value{
A list containing processed $simset.list and $outcomes (a vector)
}
\description{
Cleaning and verifying the plotting data so it can be properly plotted
}
