% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS_plot_simulations.R
\name{plot.simulations}
\alias{plot.simulations}
\title{Plot Simulations With Plotly}
\usage{
\method{plot}{simulations}(
  ...,
  outcomes,
  corresponding.data.outcomes = NULL,
  split.by = NULL,
  facet.by = NULL,
  dimension.values = list(),
  target.ontology = NULL,
  summary.type = c("individual.simulation", "mean.and.interval",
    "median.and.interval")[1],
  plot.year.lag.ratio = F,
  title = "location",
  n.facet.rows = NULL,
  interval.coverate = 0.95,
  data.manager = get.default.data.manager(),
  style.manager = get.default.style.manager(),
  hide.legend = FALSE
)
}
\arguments{
\item{...}{One or more of either (1) jheem.simulation.set objects or (2) lists containing only jheem.simulation or jheem.simset objects}

\item{outcomes}{A character vector of which simulation outcomes to plot}

\item{split.by}{A character vector of dimensions for which to make different lines}

\item{facet.by}{A character vector of dimensions for which to make different panels}

\item{data.manager}{The data.manager from which to draw real-world data for the plots}

\item{style.manager}{An object of class 'jheem.style.manager' specifying style attributes}
}
\description{
Plot Simulations With Plotly
}
\details{
See "Simplot" for ggplot2-based plotting.
}
