% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS_simplot.R
\name{prepare.plot}
\alias{prepare.plot}
\title{Prepare Plot Data
A utility that performs the data-pulling half of the simplot operation
for applications like webtools that need the step to be separate.}
\usage{
prepare.plot(
  simset.list = NULL,
  outcomes = NULL,
  locations = NULL,
  corresponding.data.outcomes = NULL,
  split.by = NULL,
  facet.by = NULL,
  dimension.values = list(),
  dimension.values.post.mapping = list(),
  target.ontology = NULL,
  plot.which = c("sim.and.data", "sim.only", "data.only")[1],
  summary.type = c("individual.simulation", "mean.and.interval",
    "median.and.interval")[1],
  data.locations = NULL,
  plot.year.lag.ratio = F,
  label.function = NULL,
  title = "location",
  title.suffix = NULL,
  append.url = F,
  data.manager = get.default.data.manager(),
  style.manager = get.default.style.manager(),
  show.data.pull.error = F,
  debug = F
)
}
\arguments{
\item{simset.list}{A list of jheem.simulation.set objects}

\item{outcomes}{A character vector of which simulation outcomes to plot}

\item{corresponding.data.outcomes}{Specify directly which data outcomes should be plotted against simulation outcomes. Must be NULL or a character vector with outcomes as names; all of those outcomes must be present in either the 'outcomes' argument or in '...'"}

\item{split.by}{At most one dimension}

\item{facet.by}{Any number of dimensions but cannot include the split.by dimension}

\item{dimension.values.post.mapping}{How to subset data in the form it will take after mappings have been applied to align sim and data ontologies. A list.}

\item{plot.which}{Should simulation data and calibration data be plotted ('sim.and.data'), or only simulation data ('sim.only'), or only calibration data ('data.only')}

\item{data.locations}{Specifies locations for which to pull data. May be 1. NULL, in which case data locations will be inferred from the simsets' likelihoods, 2. a character vector of locations, or 3. a list of character vectors with outcomes as list names. For 3., outcomes not included will have their data locations inferred as in 1.}

\item{label.function}{A function to reformat labels. If NULL, will use the first provided simset's "get.labels" property, or will do no transformation if plotting data only.}

\item{title}{NULL or a single, non-NA character value. If "location", the location of the first provided simset (if any) will be used for the title.}

\item{title.suffix}{NULL or a single, non-NA character value. Something to append to the end of the title that would be generated automatically based on the 'title' argument. Note that this pastes with no space, so add the space in if you want it.}

\item{data.manager}{The data.manager from which to draw real-world data for the plots}

\item{style.manager}{We are going to have to define this down the road. It's going to govern how we do lines and sizes and colors. For now, just hard code those in, and we'll circle back to it}

\item{show.data.pull.error}{Not finding data does not block the plot from showing simulation projections, but if you'd like to see in error when data doesn't appear, set this to TRUE.}
}
\description{
Prepare Plot Data
A utility that performs the data-pulling half of the simplot operation
for applications like webtools that need the step to be separate.
}
