% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{put.data.long.form}
\alias{put.data.long.form}
\title{Put long-form data into a data manager

#'@inheritParams distribute.dimension.values}
\usage{
put.data.long.form(
  data.manager = get.default.data.manager(),
  data,
  outcome,
  metric,
  source,
  ontology.name,
  dimension.values = NULL,
  url,
  details,
  allow.na.to.overwrite = F,
  dimension.values.to.distribute = list(),
  round.distributed.dimension.values = T,
  debug = F,
  printouts = F
)
}
\arguments{
\item{data.manager}{A jheem.data.manager object}

\item{data}{A data frame or other 2-dimensional data structure with named columns. Must contain a column named 'value' of numeric values. If the 'source' argument is NULL, then data must also have a column named 'source' that gives the source for each row. May contain additional columns with names matching the specified ontology, which have the dimension values for each row. Any other columns are ignored}

\item{outcome}{The outcome type for the data. Must be an outcome previously registered with \code{\link{register.data.outcome}}}

\item{metric}{The type of measurement. The default value is "estimate", but other options include "cv", "variance", and "sd".}

\item{source}{The (single) character name of the source from which these data derive. Must be a source previously registered with \code{\link{register.data.source}} Note: a source is conceived such that one source cannot contain two sets of data for the same set of dimension values}

\item{ontology.name}{The name of the ontology which the data follow. Must be an ontology previously registered with \code{\link{register.data.ontology}}}

\item{dimension.values}{A named list that indicates what subset of a bigger data element these particular data should be stored into. Each element must be a named character, numeric, or logical vector.}

\item{url}{A character vector with at least one element giving one or more URLs indicating where the data derive from}

\item{details}{A character vector with at least one element giving one or more URLs giving data collection details. In general, data collected and tabulated using the same approach should have the same details, whereas data with different methods/tabulation should have different details}

\item{allow.na.to.overwrite}{A logical indicator for whether NA values in data should be allowed to overwrite previous values put to the data manager (if data have been put previously)}

\item{dimension.values.to.distribute}{A named list of the dimensions and their values which should be redistributed}

\item{round.distributed.dimension.values}{Should distributed quantities be rounded?}
}
\description{
Put long-form data into a data manager

#'@inheritParams distribute.dimension.values
}
