% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{register.aging}
\alias{register.aging}
\title{Register Information about Aging for a Model Specification}
\usage{
register.aging(
  specification,
  aging.rate.value,
  tag = "aging",
  groups = c("uninfected", "infected"),
  applies.to = list()
)
}
\arguments{
\item{specification}{The jheem.specification object to modify}

\item{aging.rate.value}{The values for the aging rate. Either: (1) a numeric object, (2) a character string giving the name of a model quantity or a model element, (3) an 'call' or 'expression' object (generated by expression() or expr()) containing an expression comprising other model quantities or model elements, and the operators +, -, *, /, ^, sqrt, log, and exp, or (4) a function whose arguments are some subset of 'specification.metadata', 'location', other model quantities and model elements, and ...}

\item{tag}{A tag used to follow track this component}

\item{groups}{The group(s) to which the aging apply. Either 'infected' or 'uninfected' or both}

\item{applies.to}{A named list of character or integer vectors, denoting what subset of the group aging applies to}
}
\description{
Register Information about Aging for a Model Specification
}
\seealso{
Other Defining Core Components for a Model Specification: 
\code{\link{register.model.mechanism}()},
\code{\link{register.mortality}()},
\code{\link{register.natality}()},
\code{\link{register.remission}()},
\code{\link{register.transition}()},
\code{\link{register.transmission}()}
}
\concept{Defining Core Components for a Model Specification}
