% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERSIONS_version_manager.R
\name{register.apply.parameters.to.engine.function}
\alias{register.apply.parameters.to.engine.function}
\title{Register a Function to Apply Parameters to a JHEEM Engine}
\usage{
register.apply.parameters.to.engine.function(
  version,
  fn,
  join.with.previous.version.function = T
)
}
\arguments{
\item{version}{The name of a JHEEM version for which a specification has already been registered (using \code{\link{}})}

\item{fn}{A function which takes three arguments: (1) 'jheem.engine', an object of class jheem.engine, (2) 'parameters', a named numeric vector, and (3) 'check.consistency', a logical indicating whether consistency checks should be performed. If check.consistency==T, the function should return a character vector of parameter names that were used. If F, any return value from the function will be ignored}

\item{join.with.previous.version.function}{Whether the apply.parameters.to.engine.function should be called first when this function is invoked (in other words, whether this function should 'inherit' the behavior of the previous function)}
}
\description{
Register a Function to Apply Parameters to a JHEEM Engine
}
\seealso{
Other JHEEM Version Management Functions: 
\code{\link{get.specification.for.version}()},
\code{\link{register.model.specification}()}
}
\concept{JHEEM Version Management Functions}
