% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_main.R
\name{register.calibration.info}
\alias{register.calibration.info}
\title{Register Calibration Info}
\usage{
register.calibration.info(
  code,
  likelihood.instructions,
  is.preliminary,
  end.year,
  parameter.names,
  parameter.aliases = NULL,
  n.iter,
  thin,
  description,
  special.case.likelihood.instructions = list(),
  fixed.initial.parameter.values = numeric(),
  max.run.time.seconds = Inf,
  solver.metadata = NULL,
  n.chains = if (is.preliminary) 1 else 4,
  n.burn = if (is.preliminary) 0 else floor(n.iter/2),
  data.manager = get.default.data.manager(),
  preceding.calibration.codes = character(),
  draw.from.parent.version = F,
  weight.to.preceding.variance.estimates = 0.5,
  error.prefix = "Error registering calibration info: "
)
}
\arguments{
\item{parameter.aliases}{A named list of character vectors. Each character vector must be one or more parameter names from the parameter distribution, and the names represent the aliases to be used in sampling}

\item{error.prefix}{}
}
\description{
Register Calibration Info
}
