% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{register.data.ontology}
\alias{register.data.ontology}
\title{Register a new data ontology to a data manager before putting data to it}
\usage{
register.data.ontology(data.manager = get.default.data.manager(), name, ont)
}
\arguments{
\item{data.manager}{A jheem.data.manager object}

\item{name}{The name of the data group}

\item{ont}{An ontology object as created by \code{\link{ontology()}}}
}
\description{
Register a new data ontology to a data manager before putting data to it
}
