% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{register.default.parameter.values}
\alias{register.default.parameter.values}
\title{Register Default Parameter Values}
\usage{
register.default.parameter.values(specification, parameter.values)
}
\description{
Register Default Parameter Values
}
