% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{register.fixed.model.strata}
\alias{register.fixed.model.strata}
\title{Set Whether to Fix Strata Sizes During a Time Period}
\usage{
register.fixed.model.strata(
  specification,
  applies.after.time,
  applies.before.time,
  fix.strata = T,
  dimensions.to.fix = NULL
)
}
\arguments{
\item{specification}{The jheem.specification object}

\item{applies.after.time, applies.before.time}{Single numeric values giving the time frame over whether this setting applies}

\item{fix.strata}{A single logical value indicating whether strata should be fixed during this time frame}

\item{dimensions.to.fix}{A character vector denoting which dimensions should be fixed. These should be dimensions common to both infected and uninfected groups. These are only used if fix.strata==TRUE}
}
\description{
Set Whether to Fix Strata Sizes During a Time Period
}
