% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{register.initial.population}
\alias{register.initial.population}
\title{Set The Initial Model Population}
\usage{
register.initial.population(specification, value, group)
}
\arguments{
\item{specification}{The jheem.specification object}

\item{value}{Either: (1) a numeric object, (2) a character string giving the name of a model quantity or a model element, (3) an 'call' or 'expression' object (generated by expression() or expr()) containing an expression comprising other model quantities or model elements, and the operators +, -, *, /, ^, sqrt, log, and exp, or (4) a function whose arguments are some subset of 'specification.metadata', 'location', other model quantities and model elements, and ...}

\item{group}{Which group ("infected" or "uninfected") this initial population is for}
}
\description{
Set The Initial Model Population
}
