% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_main.R
\name{register.likelihood.instructions}
\alias{register.likelihood.instructions}
\title{Register Likelihood Instructions}
\usage{
register.likelihood.instructions(
  instructions,
  code,
  description,
  error.prefix = "Error registering likelihood instructions: "
)
}
\arguments{
\item{instructions}{A previously unregistered jheem.likelihood.instructions.object.}

\item{code}{A valid code containing only letters, numbers, dashes, and periods.}

\item{description}{A description of the likelihood for user reference.}

\item{error.prefix}{(Optional) A message to prepend to error messages.}
}
\description{
Register Likelihood Instructions
}
