% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{register.model.foreground}
\alias{register.model.foreground}
\title{Register a Foreground for a Model Specification}
\usage{
register.model.foreground(specification, foreground, name)
}
\arguments{
\item{specification}{The jheem.specification object to modify}

\item{foreground}{An object of class 'jheem.model.foreground', as created by \code{\link{create.model.foreground}}}

\item{name}{A single character value giving the name by which the foreground can be referred to}
}
\description{
Register a Foreground for a Model Specification
}
