% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{register.model.quantity.subset}
\alias{register.model.quantity.subset}
\title{Register a Value to a Subset of a Model Quantity or Transition}
\usage{
register.model.quantity.subset(
  specification,
  name,
  value,
  applies.to,
  na.replacement = as.numeric(NA),
  apply.function = c("overwrite", "add", "subtract", "divide", "multiply")[1],
  ...
)
}
\arguments{
\item{specification}{The jheem.specification object to modify}

\item{name}{The name of the model quantity or transition to which to register a subset}

\item{value}{Either: (1) a numeric object, (2) a character string giving the name of a model quantity or a model element, (3) an 'call' or 'expression' object (generated by expression() or expr()) containing an expression comprising other model quantities or model elements, and the operators +, -, *, /, ^, sqrt, log, and exp, or (4) a function whose arguments are some subset of 'specification.metadata', 'location', other model quantities and model elements, and ...}

\item{applies.to}{A named list representing the subset of value in question. The names of applies.to should correspond to dimensions of value, and the values must be either character, integer, or logical vectors accessing values in each dimension}

\item{na.replacement}{A single numeric value that should be used to replace NA values generated in calculating the quantity's value (note, if na.replacement==NA, no handling of NAs is performed and NAs will cause an error in model execution)}

\item{apply.function}{A character value indicating how the value for this subset should be incorporated into the main quantity. Options are 'overwrite' (the value for the subset overwrites the value in the quantity), 'add' (subset value is added to the quantity), 'subtract' (subset value subtracted from the quantity), 'multiply' (subset value multiplied by the quantity), or 'divide' (quantity is divided by the subset value)}

\item{...}{Additional parameters to be passed to value if a function is passed as the value}
}
\description{
Register a Value to a Subset of a Model Quantity or Transition
}
