% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERSIONS_version_manager.R
\name{register.model.specification}
\alias{register.model.specification}
\title{Register a JHEEM Model Specification}
\usage{
register.model.specification(specification)
}
\arguments{
\item{specification}{A jheem.specification object, as created by \code{\link{create.jheem.specification}}}
}
\description{
Register a JHEEM Model Specification
}
\details{
Once a specification object is registered, it can no longer be modified
}
\seealso{
Other JHEEM Version Management Functions: 
\code{\link{get.specification.for.version}()}
}
\concept{JHEEM Version Management Functions}
