% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{register.natality}
\alias{register.natality}
\title{Register Information about Natality (Births) for a Model Specification}
\usage{
register.natality(
  specification,
  parent.groups,
  child.groups,
  fertility.rate.value,
  birth.proportions.value,
  parent.child.concordant.dimensions = character(),
  all.births.from.compartments = list(),
  all.births.into.compartments = character(),
  tag = "natality",
  applies.to = list()
)
}
\arguments{
\item{specification}{The jheem.specification object to modify}

\item{parent.groups}{The group representing parents (infected or uninfected)}

\item{child.groups}{The group into which births are distributed (infected or uninfected)}

\item{fertility.rate.value, birth.proportions.value}{The values for the fertility rate and the proportions by which births are distributed into the to.group. Either: (1) a numeric object, (2) a character string giving the name of a model quantity or a model element, (3) an 'call' or 'expression' object (generated by expression() or expr()) containing an expression comprising other model quantities or model elements, and the operators +, -, *, /, ^, sqrt, log, and exp, or (4) a function whose arguments are some subset of 'specification.metadata', 'location', other model quantities and model elements, and ...}

\item{parent.child.concordant.dimensions}{A character vector of the dimensions for which all new births share the same value as their parents}

\item{all.births.from.compartments}{A named list of character or integer values. The names represent dimensions, and the elements represent the compartments from those dimensions that can give birth}

\item{all.births.into.compartments}{A named list of single character or integer values. The names represent dimensions and the elements represent the single compartment in that dimension into which all births go}

\item{tag}{A tag used to follow track this component}

\item{applies.to}{A named list of character or integer vectors, denoting what subset of the from.group fertility applies to}
}
\description{
Register Information about Natality (Births) for a Model Specification
}
\seealso{
Other Defining Core Components for a Model Specification: 
\code{\link{register.aging}()},
\code{\link{register.model.mechanism}()},
\code{\link{register.mortality}()},
\code{\link{register.remission}()},
\code{\link{register.transition}()},
\code{\link{register.transmission}()}
}
\concept{Defining Core Components for a Model Specification}
