% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology_mappings.R
\name{register.ontology.mapping}
\alias{register.ontology.mapping}
\title{Create and register a mapping that allows transformations of arrays}
\usage{
register.ontology.mapping(
  name,
  mappings,
  from.dimensions,
  to.dimensions = from.dimensions
)
}
\arguments{
\item{name}{A descriptive single character value}

\item{mappings}{A character matrix, with a number of columns == length(from.dimensions) + length(to.dimensions). The first length(from.dimensions) columns represent values for the from.dimensions (in the same order as from.dimensions), and the last length(to.dimensions) colums represent values for to.dimensions (in the order of to.dimensions). Each row represents a mapping from a set of values for from.dimension to a set of values for to.dimensions. NB: any dimnames on mappings will be ignored}

\item{from.dimensions}{A character vector of names of the dimensions we are mapping from}

\item{to.dimensions}{A character vector of names of the dimensions we are mapping to}
}
\description{
Create and register a mapping that allows transformations of arrays
}
\details{
The mappings must contain every possible combination of 'from' values (ie, every combo of from values must map to something). A row of 'mappings' may have NA values in ALL the 'from' columns (indicating there is no set of 'from' values that maps to the corresponding 'to' values), or may have NA values in ALL the 'to' columns (indicating the that the given set of 'from' values will be discarded in the transformation)
}
