% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{register.parent.data.source}
\alias{register.parent.data.source}
\title{Register a parent data source to a data manager}
\usage{
register.parent.data.source(
  data.manager = get.default.data.manager(),
  parent.source,
  full.name,
  short.name = full.name
)
}
\arguments{
\item{data.manager}{A jheem.data.manager object}

\item{parent.source}{The name (a single character value) of the parent source. This is the 'internal' name by which the parent source will be referenced in accessing the data manager}

\item{full.name}{A descriptive, fully-formatted and capitalized name to use in generating figures and tables (eg in popovers). Avoid abbreviations. Should be unique to the parent data source (although this is not enforced)}

\item{short.name}{A name for the parent data source to use in setting where brevity is important. Ideal to use abbreviations, but can be the same as full.name}
}
\description{
Register a parent data source to a data manager
}
