% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{register.remission}
\alias{register.remission}
\title{Register Information about Remission from Infection for a Model Specification}
\usage{
register.remission(
  specification,
  remission.rate.value,
  remission.proportions.value,
  tag = "remission",
  applies.to = list(),
  all.remissions.into.compartments = character()
)
}
\arguments{
\item{specification}{The jheem.specification object to modify}

\item{remission.rate.value, remission.proportions.value}{The values for the remission rate and the proportions by which remitted cases are distributed into the uninfected group. Either: (1) a numeric object, (2) a character string giving the name of a model quantity or a model element, (3) an 'call' or 'expression' object (generated by expression() or expr()) containing an expression comprising other model quantities or model elements, and the operators +, -, *, /, ^, sqrt, log, and exp, or (4) a function whose arguments are some subset of 'specification.metadata', 'location', other model quantities and model elements, and ...}

\item{tag}{A tag used to follow track this component}

\item{applies.to}{A named list of character or integer vectors, denoting what subset of the infected group remissions apply to}
}
\description{
Register Information about Remission from Infection for a Model Specification
}
\seealso{
Other Defining Core Components for a Model Specification: 
\code{\link{register.aging}()},
\code{\link{register.model.mechanism}()},
\code{\link{register.mortality}()},
\code{\link{register.natality}()},
\code{\link{register.transition}()},
\code{\link{register.transmission}()}
}
\concept{Defining Core Components for a Model Specification}
