% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERSIONS_version_manager.R
\name{register.sampled.parameters.for.version}
\alias{register.sampled.parameters.for.version}
\title{Register Parameters to Be Sampled Prior to Making Projections from a JHEEM Version}
\usage{
register.sampled.parameters.for.version(
  version,
  distribution,
  apply.function,
  join.with.previous.version
)
}
\arguments{
\item{version}{The name of an EHE version for which a specification has been registered with \code{\link{register.model.specification}}}

\item{distribution}{An object of class 'distribution' that specifies the prior distribution for the parameters}

\item{apply.function}{A function that knows how to apply parameter values to a jheem.model.settings object. The function should take two arguments: (1) 'model.settings', an object of class jheem.model.settings, and (2) 'parameters', a named numeric vector}

\item{join.with.previous.version}{Whether the distribution and apply.function should be merged with those of the previous version}
}
\description{
Register Parameters to Be Sampled Prior to Making Projections from a JHEEM Version
}
\details{
*Sampled* parameters have their values randomly chosen from a distribution prior to running projections PAST the calibration period
}
