% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERSIONS_version_manager.R
\name{register.set.parameters.for.version}
\alias{register.set.parameters.for.version}
\title{Register Parameters to Be Set for Running a Simulation}
\usage{
register.set.parameters.for.version(
  version,
  parameter.names,
  apply.function,
  join.with.previous.version
)
}
\arguments{
\item{version}{The name of an EHE version for which a specification has been registered with \code{\link{register.model.specification}}}

\item{parameter.names}{A character vector giving the names of parameters which will be set}

\item{apply.function}{A function that knows how to apply parameter values to a jheem.model.settings object. The function should take two arguments: (1) 'model.settings', an object of class jheem.model.settings, and (2) 'parameters', a named numeric vector}

\item{join.with.previous.version}{Whether the parameter.names and apply.function should be merged with those of the previous version}
}
\description{
Register Parameters to Be Set for Running a Simulation
}
\details{
*Set* parameters have their values set explicitly (neither sampled nor calibrated)
}
