% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{register.top.level.quantity}
\alias{register.top.level.quantity}
\title{Register a Top-Level Quantity (ie a quantity required or pre-specified as optional for running a JHEEM simulation)}
\usage{
register.top.level.quantity(
  specification,
  name,
  value,
  groups = "all",
  na.replacement = as.numeric(NA),
  ...
)
}
\arguments{
\item{specification}{The jheem.specification object to modify}

\item{name}{The name of the quantity. Must be one of specification$top.level.quantity.names}

\item{value}{Either: (1) a numeric object, (2) a character string giving the name of another model quantity or a model element, (3) an 'call' or 'expression' object (generated by expression() or expr()) containing an expression comprising other model quantities or model elements, and the operators +, -, *, /, ^, sqrt, log, and exp, or (4) a function whose arguments are some subset of 'specification.info', 'location', other model quantities and model elements, and ...}

\item{groups}{The groups to which this quantity applies. Either 'uninfected', 'infected', or 'all' ('all' applies the quantity to all groups which it could apply to)}

\item{na.replacement}{A single numeric value that should be used to replace NA values generated in calculating the quantity's value (note, if na.replacement==NA, no handling of NAs is performed and NAs will cause an error in model execution)}

\item{...}{Additional parameters to be passed to value if a function is passed as the value}
}
\description{
Register a Top-Level Quantity (ie a quantity required or pre-specified as optional for running a JHEEM simulation)
}
