% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{register.transition}
\alias{register.transition}
\title{Register Information about Disease Transmission for a Model Specification}
\usage{
register.transition(
  specification,
  dimension,
  from.compartments,
  to.compartments,
  value,
  groups,
  applies.to = list(),
  tag = "transition",
  na.replacement = as.numeric(NA),
  ...
)
}
\arguments{
\item{dimension}{The name of the dimension in which this transition operates}

\item{from.compartments, to.compartments}{The compartments from and to which this transition applies. May be (1) one or more names of compartments, (2) one or more character compartment.value.aliases, (3) one or more integers denoting compartments, (4) a logical vector, which, when applied to the dimension names, denotes the compartments}

\item{groups}{The group(s) to which this transition applies. Either 'uninfected', 'infected', or both}
}
\description{
Register Information about Disease Transmission for a Model Specification
}
\seealso{
Other Defining Core Components for a Model Specification: 
\code{\link{register.aging}()},
\code{\link{register.model.mechanism}()},
\code{\link{register.mortality}()},
\code{\link{register.natality}()},
\code{\link{register.remission}()},
\code{\link{register.transmission}()}
}
\concept{Defining Core Components for a Model Specification}
