% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{register.transmission}
\alias{register.transmission}
\title{Register Information about Disease Transmission for a Model Specification}
\usage{
register.transmission(
  specification,
  contact.value,
  susceptibility.value,
  transmissibility.value,
  new.infection.proportions.value,
  tag = "transmission",
  all.new.infections.into.compartments = list(),
  from.applies.to = list(),
  to.applies.to = list(),
  transmission.applies.to = list(),
  new.infections.applies.to = list()
)
}
\arguments{
\item{specification}{The jheem.specification object to modify}

\item{contact.value, susceptibility.value, transmissibility.value, new.infection.proportions.value}{The values for the contact matrix, susceptibility, transmissibility, and proportions by which new infections are distributed into the infected group. Either: (1) a numeric object, (2) a character string giving the name of a model quantity or a model element, (3) an 'call' or 'expression' object (generated by expression() or expr()) containing an expression comprising other model quantities or model elements, and the operators +, -, *, /, ^, sqrt, log, and exp, or (4) a function whose arguments are some subset of 'specification.metadata', 'location', other model quantities and model elements, and ...}

\item{tag}{A tag used to follow track this component}

\item{from.applies.to}{A named list of character or integer vectors, denoting what subset of the population PARTNERS may come from (regardless of whether or not they actually transmit). Must be a subset of compartments.for.infected.and.uninfected}

\item{to.applies.to}{A named list of character or integer vectors, denoting what subset of the uninfected group are susceptible to infection. Must be a subset of compartments.for.infected.and.uninfected + compartments.for.uninfected.only}

\item{transmission.applies.to}{A named list of character or integer vectors, denoting what subset of the infected group may transmit infection. Must be a subset of compartments.for.infected.and.uninfected + compartments.for.infected.only}

\item{new.infections.applies.to}{A named list of character or integer vectors, denoting what subset of the infected group new infections may enter into. Must be a subset of compartments.for.infected.only}
}
\description{
Register Information about Disease Transmission for a Model Specification
}
\seealso{
Other Defining Core Components for a Model Specification: 
\code{\link{register.aging}()},
\code{\link{register.model.mechanism}()},
\code{\link{register.mortality}()},
\code{\link{register.natality}()},
\code{\link{register.remission}()},
\code{\link{register.transition}()}
}
\concept{Defining Core Components for a Model Specification}
