% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_transmute_calibration.R
\name{register.transmute.calibration.info}
\alias{register.transmute.calibration.info}
\title{Register Transmute Calibration Info}
\usage{
register.transmute.calibration.info(
  transmute.code,
  from.version,
  to.version,
  n.iter.first.sim,
  n.iter.subsequent.sims,
  likelihood.instructions,
  prior.distribution,
  sampling.blocks,
  fixed.initial.parameter.values = numeric(),
  from.sub.version = NULL,
  to.sub.version = NULL,
  max.lookback.attempts = 20
)
}
\description{
Register Transmute Calibration Info
}
