% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{remove.data}
\alias{remove.data}
\title{Remove Data From Data Manager}
\usage{
remove.data(
  data.manager,
  outcome,
  source,
  ontology.name,
  dimension.values,
  metric = "estimate",
  details.for.removal = "removed",
  url.for.removal = "url"
)
}
\arguments{
\item{data.manager}{A jheem.data.manager object}

\item{outcome}{The outcome type for the data. Must be an outcome previously registered with \code{\link{register.data.outcome}}}

\item{source}{The (single) character name of the source from which these data derive. Must be a source previously registered with \code{\link{register.data.source}} Note: a source is conceived such that one source cannot contain two sets of data for the same set of dimension values}

\item{ontology.name}{The name of the ontology which the data follow. Must be an ontology previously registered with \code{\link{register.data.ontology}}}

\item{dimension.values}{A named list that indicates what subset of a bigger data element these particular data should be stored into. Each element must be a named character, numeric, or logical vector.}

\item{metric}{The type of measurement. The default value is "estimate", but other options include "cv", "variance", and "sd".}

\item{details.for.removal, url.for.removal}{Single character values to use for these points' 'details' and 'url'}
}
\description{
Replaces data with NA if data exists
}
