% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERVENTIONS_intervention_effects.R
\name{resolve.intervention.effect}
\alias{resolve.intervention.effect}
\title{#@export}
\usage{
resolve.intervention.effect(intervention.effect, bindings, error.prefix = "")
}
\description{
#@export
}
