% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology.R
\name{resolve.ontology.dimension.values}
\alias{resolve.ontology.dimension.values}
\title{Resolve dimension.values into a set of dimnames for an ontology}
\usage{
resolve.ontology.dimension.values(
  ont,
  dimension.values,
  error.prefix,
  throw.error.if.unresolvable = T
)
}
\arguments{
\item{dimension.values}{A list of dimension.values (ie, a named list of either character, integer, or logical vectors that indexes into the dimnames given in ont)}

\item{error.prefix}{Text to prepend to any error messages that are thrown}

\item{throw.error.if.unresolvable}{A logical indicating what to do if unable to resolve against the ontology. When FALSE, will return NULL if unresolvable}
}
\value{
If the dimension.values can be resolved against ontology ont, returns a list of dimnames (ie, a named list of character vectors). If cannot resolve will either throw an error (if throw.error.if.unresolvable) or return NULL (NB, if passed dimension.values==NULL, will also return NULL)
}
\description{
Resolve dimension.values into a set of dimnames for an ontology
}
