% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILE_MANAGER_file_manager.R
\name{retrieve.simulation.set}
\alias{retrieve.simulation.set}
\title{Retrieve a Saved Simulation Set}
\usage{
retrieve.simulation.set(
  version,
  location,
  calibration.code,
  n.sim,
  intervention.code = NULL,
  sub.version = NULL,
  root.dir = get.jheem.root.directory("Cannot get.simset.filename(): "),
  error.prefix = "",
  throw.error.if.missing = T
)
}
\description{
Loads a simulation set that has been properly saved in JHEEM's calibration
file structure.
}
