% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_main.R
\name{run.calibration}
\alias{run.calibration}
\title{Run A Calibration}
\usage{
run.calibration(
  version,
  location,
  calibration.code,
  root.dir = get.jheem.root.directory("Cannot set up calibration: "),
  chains,
  update.frequency = 500,
  update.detail = "low"
)
}
\arguments{
\item{version}{The version of the model specification (must have been previously registered)}

\item{location}{A single character value representing the location for the calibration}

\item{update.detail}{}
}
\value{
An MCMC object
}
\description{
Run A Calibration
}
