% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{run.jheem.engine}
\alias{run.jheem.engine}
\title{Produce a Simulation from a JHEEM Engine}
\usage{
run.jheem.engine(
  jheem.engine,
  start.year,
  end.year,
  check.consistency = T,
  max.run.time.seconds = Inf,
  prior.sim = NULL,
  keep.years = start.year:end.year,
  atol = 1e-04,
  rtol = 1e-04
)
}
\arguments{
\item{jheem.engine}{A jheem.engine object (created by \code{\link{create.jheem.engine}})}

\item{start.year, end.year}{The years across which the simulation is to run}

\item{check.consistency}{Whether to perform consistency checks during calculations, which can catch errors, but imposes a computational cost. In general, this is likely only necessary on the first time an engine is crunched or run}

\item{max.run.time.seconds}{The maximum number of seconds a simulation is allowed to run before being terminated}

\item{prior.sim}{A jheem.simulation object from which to start this simulation. NULL if not running from a prior simulation}

\item{keep.years}{The years for which to keep simulation data}

\item{atol, rtol}{The absolute and relative tolerance parameters to be passed along to the differential equation solver}
}
\value{
A jheem.simulation object
}
\description{
Produce a Simulation from a JHEEM Engine
}
\seealso{
Other Functions to create and modify a jheem.engine: 
\code{\link{create.jheem.engine}()},
\code{\link{crunch.jheem.engine}()},
\code{\link{set.element.functional.form.from.time}()},
\code{\link{set.element.functional.form.future.slope}()},
\code{\link{set.element.functional.form.interaction.alphas}()},
\code{\link{set.element.functional.form.main.effect.alphas}()},
\code{\link{set.element.ramp.times}()},
\code{\link{set.element.values.from.parameters}()},
\code{\link{set.element.value}()}
}
\concept{Functions to create and modify a jheem.engine}
