% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_simulation.R
\name{run.simulations.from.parameters}
\alias{run.simulations.from.parameters}
\title{Run a Simulations from a Set of Parameters}
\usage{
run.simulations.from.parameters(
  version,
  location,
  parameters,
  end.year,
  sub.version = NULL,
  max.run.time.seconds = Inf,
  calibration.code = NULL,
  solver.metadata = create.solver.metadata(),
  keep.from.year = start.year,
  keep.to.year = end.year,
  verbose = F
)
}
\arguments{
\item{parameters}{The parameters to run simulations for. Can be either (a) a matrix with parameters on the rows, and one column for each set of parameters/simulation to run, with named rows or (b) a named numeric vector of parameters for a single simulation}

\item{calibration.code}{If parameters were derived from a previous calibration, the calibration code that should be attached to the resulting simulation set}
}
\description{
Run a Simulations from a Set of Parameters
}
