% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_transmute_calibration.R
\name{run.transmute.calibration}
\alias{run.transmute.calibration}
\title{Run Transmute Calibration}
\usage{
run.transmute.calibration(
  transmute.code,
  location,
  n.sim,
  chunks = NULL,
  verbose = T,
  update.frequency = 500,
  update.detail = "low",
  ignore.errors = T,
  rerun = T,
  root.dir = get.jheem.root.directory()
)
}
\description{
Run Transmute Calibration
}
