% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILE_MANAGER_file_manager.R
\name{save.simulation.set}
\alias{save.simulation.set}
\title{Save a JHEEM Simulation Set}
\usage{
save.simulation.set(
  simset,
  root.dir = get.jheem.root.directory("Cannot save simulation set: ")
)
}
\arguments{
\item{simset}{A jheem.simulation.set object}

\item{root.dir}{The directory in which the folder tree for saving the simset will be rooted}
}
\description{
Save a JHEEM Simulation Set
}
