% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{set.default.data.manager}
\alias{set.default.data.manager}
\title{Set the default data manager}
\usage{
set.default.data.manager(data.manager)
}
\description{
Set the default data manager
}
