% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STYLE_MANAGER_style_manager.R
\name{set.default.style.manager}
\alias{set.default.style.manager}
\title{Set the default style manager}
\usage{
set.default.style.manager(style.manager)
}
\description{
Set the default style manager
}
