% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{set.element.functional.form.alphas.from.parameters}
\alias{set.element.functional.form.alphas.from.parameters}
\title{Set multiple functional form main effect alphas from parameters}
\usage{
set.element.functional.form.alphas.from.parameters(
  jheem.engine,
  element.name,
  alpha.name,
  parameters,
  parameter.name.prefix,
  parameter.name.suffix,
  dimensions.with.values.referred.to.by.name = character(),
  dimensions.with.values.referred.to.by.index = character(),
  throw.error.if.no.parameters = T,
  check.consistency = !jheem.engine$has.been.crunched()
)
}
\arguments{
\item{jheem.engine}{A jheem.engine object (created by \code{\link{create.jheem.engine}})}

\item{element.name}{The name of the model element whose functional form we wish to modify}

\item{alpha.name}{The name of the functional form parameter that we want to set alphas for}

\item{parameters}{A named numeric vector with values for interaction terms}

\item{parameter.name.prefix, parameter.name.suffix}{Character values to be prepended and addended to names of specific dimension values to generate specific parameter names}

\item{dimensions.with.values.referred.to.by.name}{Dimensions for which we should look for main effects in parameters, and make parameter names by combining parameter.name.prefix and parameter.name.suffix with the actual dimension values for the dimension}

\item{dimensions.with.values.referred.to.by.index}{Dimensions for which we should look for main effects in parameters, and make parameter names by combining parameter.name.prefix and parameter.name.suffix with indices of the values in each dimension}

\item{throw.error.if.no.parameters}{A logical indicator - if no parameter names match the pattern parameter.name.prefix<x>parameter.name.suffix, whether an error should be thrown}

\item{check.consistency}{A logical indicator whether to sanitize arguments and check for consistency during calculations. Setting to F increases performance}
}
\value{
The names of parameters that were used
}
\description{
Set multiple functional form main effect alphas from parameters
}
\details{
This function will search for parameters with names in the format of either 
         parameter.name.prefix<dimension values for dimensions.with.values.referred.to.by.name>parameter.name.suffix 
         OR
         parameter.name.prefix<dimension in dimension.values.referred.to.by.name><1:length of dimension>parameter.name.suffix
}
