% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{set.element.functional.form.future.slope}
\alias{set.element.functional.form.future.slope}
\alias{set.element.functional.form.future.slope.after.time}
\title{Set a "Future Slope" and the time after which it applies, for a model element}
\usage{
set.element.functional.form.future.slope(
  model.settings,
  element.names,
  slope,
  wrt.version = model.settings$version
)

set.element.functional.form.future.slope.after.time(
  model.settings,
  element.names,
  after.time,
  wrt.version = model.settings$version
)
}
\arguments{
\item{model.settings}{An object of class 'jheem.model.settings'}

\item{element.names}{One or more names of the model element to set the future.slope for}

\item{slope}{A single numeric value representing the slope}

\item{wrt.version}{The version relative to which the element name is resolved (if an element.name is overwritten in a descendant version, using wrt.version can be used to select which specification's model element is to be modified)}

\item{after.time}{The year after which the additional slope takes effect}
}
\description{
Set a "Future Slope" and the time after which it applies, for a model element
}
\details{
A future slope is applied to model elements with a functional.form set AFTER a given point in time
}
\section{Functions}{
\itemize{
\item \code{set.element.functional.form.future.slope.after.time()}: 

}}
\seealso{
Other Functions to create and modify a model settings: 
\code{\link{set.element.functional.form.from.time}()},
\code{\link{set.element.functional.form.interaction.alphas}()},
\code{\link{set.element.ramp.times}()}
}
\concept{Functions to create and modify a model settings}
