% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{set.element.ramp.times}
\alias{set.element.ramp.times}
\alias{set.element.ramp.values}
\alias{set.element.taper.times}
\alias{set.element.taper.values}
\title{Set a ramp or taper (times and values) for a model element}
\usage{
set.element.ramp.times(
  jheem.engine,
  element.name,
  times,
  indices = 1:length(times),
  check.consistency = !jheem.engine$has.been.crunched()
)

set.element.ramp.values(
  jheem.engine,
  element.name,
  values,
  indices = 1:length(values),
  check.consistency = !jheem.engine$has.been.crunched()
)

set.element.taper.times(
  jheem.engine,
  element.name,
  times,
  indices = 1:length(times),
  check.consistency = !jheem.engine$has.been.crunched()
)

set.element.taper.values(
  jheem.engine,
  element.name,
  values,
  indices = 1:length(values),
  check.consistency = !jheem.engine$has.been.crunched()
)
}
\arguments{
\item{jheem.engine}{A jheem.engine object (created by \code{\link{create.jheem.engine}})}

\item{element.name}{The name of the model element to set the ramp/taper for}

\item{times, values}{The times or values to set for the ramp/taper}

\item{indices}{The indices of the ramp/taper to set times or values for. Can be either numeric indices, character indices (if the ramp/taper was named when setting up the specification), or logical indices}

\item{check.consistency}{A logical indicator whether to sanitize arguments and check for consistency during calculations. Setting to F increases performance}
}
\description{
Set a ramp or taper (times and values) for a model element
}
\details{
A ramp is a set of multipliers which are multiplied by the first functional.form value (ie, the value produced at functional.form.from.time) for an element. A taper is a set of multipliers which are multiplied by the last functional.form value (ie, the value produced at functional.form.to.time) for an element
}
\section{Functions}{
\itemize{
\item \code{set.element.ramp.values()}: 

\item \code{set.element.taper.times()}: 

\item \code{set.element.taper.values()}: 

}}
\seealso{
Other Functions to create and modify a jheem.engine: 
\code{\link{create.jheem.engine}()},
\code{\link{crunch.jheem.engine}()},
\code{\link{run.jheem.engine}()},
\code{\link{set.element.functional.form.from.time}()},
\code{\link{set.element.functional.form.future.slope}()},
\code{\link{set.element.functional.form.interaction.alphas}()},
\code{\link{set.element.functional.form.main.effect.alphas}()},
\code{\link{set.element.values.from.parameters}()},
\code{\link{set.element.value}()}
}
\concept{Functions to create and modify a jheem.engine}
