% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{set.element.value}
\alias{set.element.value}
\title{Set the Value of a Model Element}
\usage{
set.element.value(
  jheem.engine,
  element.name,
  value,
  check.consistency = !jheem.engine$has.been.crunched()
)
}
\arguments{
\item{jheem.engine}{A jheem.engine object (created by \code{\link{create.jheem.engine}})}

\item{element.name}{The name of the model element to update a value for}

\item{value}{The new value for the model element}

\item{check.consistency}{A logical indicator whether to sanitize arguments and check for consistency during calculations. Setting to F increases performance}
}
\description{
Set the Value of a Model Element
}
\seealso{
Other Functions to create and modify a jheem.engine: 
\code{\link{create.jheem.engine}()},
\code{\link{crunch.jheem.engine}()},
\code{\link{run.jheem.engine}()},
\code{\link{set.element.functional.form.from.time}()},
\code{\link{set.element.functional.form.future.slope}()},
\code{\link{set.element.functional.form.interaction.alphas}()},
\code{\link{set.element.functional.form.main.effect.alphas}()},
\code{\link{set.element.ramp.times}()},
\code{\link{set.element.values.from.parameters}()}
}
\concept{Functions to create and modify a jheem.engine}
