% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{set.element.values.from.parameters}
\alias{set.element.values.from.parameters}
\title{Set Values for Multiple Model Elements}
\usage{
set.element.values.from.parameters(
  jheem.engine,
  element.names,
  parameters,
  check.consistency = !jheem.engine$has.been.crunched()
)
}
\arguments{
\item{jheem.engine}{A jheem.engine object (created by \code{\link{create.jheem.engine}})}

\item{element.names}{A character vector containing the names of the elements to set values for}

\item{parameters}{A named numeric vector with values for each of the elements of 'element.names'}

\item{check.consistency}{A logical indicator whether to sanitize arguments and check for consistency during calculations. Setting to F increases performance}
}
\value{
The names of the elements that had values set
}
\description{
Set Values for Multiple Model Elements
}
\seealso{
Other Functions to create and modify a jheem.engine: 
\code{\link{create.jheem.engine}()},
\code{\link{crunch.jheem.engine}()},
\code{\link{run.jheem.engine}()},
\code{\link{set.element.functional.form.from.time}()},
\code{\link{set.element.functional.form.future.slope}()},
\code{\link{set.element.functional.form.interaction.alphas}()},
\code{\link{set.element.functional.form.main.effect.alphas}()},
\code{\link{set.element.ramp.times}()},
\code{\link{set.element.value}()}
}
\concept{Functions to create and modify a jheem.engine}
