% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILE_MANAGER_file_manager.R
\name{set.jheem.root.directory}
\alias{set.jheem.root.directory}
\title{Set JHEEM Root Directory}
\usage{
set.jheem.root.directory(dir)
}
\description{
Set the directory that calibration directories and files will be written into.
}
