% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_main.R
\name{set.up.calibration}
\alias{set.up.calibration}
\title{Set Up Calibration}
\usage{
set.up.calibration(
  version,
  location,
  calibration.code,
  root.dir = get.jheem.root.directory("Cannot set up calibration: "),
  sub.version = NULL,
  cache.frequency = 500,
  allow.overwrite.cache = F,
  verbose = T
)
}
\arguments{
\item{version}{The version of the model specification (must have been previously registered)}

\item{location}{A single character value representing the location for the calibration}

\item{verbose}{}
}
\description{
Set Up Calibration
}
