% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_transmute_calibration.R
\name{set.up.transmute.calibration}
\alias{set.up.transmute.calibration}
\title{Set Up Transmute Calibration}
\usage{
set.up.transmute.calibration(
  transmute.code,
  from.calibration.code,
  location,
  n.sim,
  n.chunks,
  allow.overwrite.cache = F,
  root.dir = get.jheem.root.directory(),
  return.simulations = F,
  verbose = T
)
}
\description{
Set Up Transmute Calibration
}
