% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS_simplot.R
\name{simplot.data.only}
\alias{simplot.data.only}
\title{Simplot Data Only}
\usage{
simplot.data.only(
  outcomes,
  locations,
  split.by = NULL,
  facet.by = NULL,
  dimension.values = list(),
  dimension.values.post.mapping = list(),
  target.ontology = NULL,
  label.function = NULL,
  plot.year.lag.ratio = F,
  title = "location",
  title.suffix = NULL,
  n.facet.rows = NULL,
  append.url = F,
  data.manager = get.default.data.manager(),
  style.manager = get.default.style.manager(),
  show.data.pull.error = F,
  debug = F
)
}
\arguments{
\item{outcomes}{A character vector of which simulation outcomes to plot}

\item{locations}{A character vector of locations for which to pull data.}

\item{split.by}{At most one dimension}

\item{facet.by}{Any number of dimensions but cannot include the split.by dimension}

\item{dimension.values.post.mapping}{How to subset data in the form it will take after mappings have been applied to align sim and data ontologies. A list.}

\item{label.function}{A function to reformat labels. If NULL, will use the first provided simset's "get.labels" property, or will do no transformation if plotting data only.}

\item{title}{NULL or a single, non-NA character value. If "location", the location of the first provided simset (if any) will be used for the title.}

\item{title.suffix}{NULL or a single, non-NA character value. Something to append to the end of the title that would be generated automatically based on the 'title' argument. Note that this pastes with no space, so add the space in if you want it.}

\item{data.manager}{The data.manager from which to draw real-world data for the plots}

\item{style.manager}{We are going to have to define this down the road. It's going to govern how we do lines and sizes and colors. For now, just hard code those in, and we'll circle back to it}

\item{show.data.pull.error}{Not finding data does not block the plot from showing simulation projections, but if you'd like to see in error when data doesn't appear, set this to TRUE.}
}
\description{
Simplot Data Only
}
