% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPERS_age_year_helpers.R
\name{standardize.age.strata.names}
\alias{standardize.age.strata.names}
\title{Convert Age Strata Names to a Standard Form}
\usage{
standardize.age.strata.names(strata.names)
}
\arguments{
\item{strata.names}{A character vector of names of age strata}
}
\description{
Convert Age Strata Names to a Standard Form
}
\details{
Calls \code{\link{parse.age.strata.names}} and feeds the output into \code{\link{make.age.strata.names}}
}
