% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPERS_age_race_helpers.R
\name{standardize.race.names}
\alias{standardize.race.names}
\title{Convert Race Names to a Standard Form}
\usage{
standardize.race.names(race.names)

standardize.race.names(race.names)
}
\arguments{
\item{race.names}{A character vector containing race names}
}
\description{
Convert Race Names to a Standard Form

Convert Race Names to a Standard Form
}
