% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_functional_forms.R
\name{STATIC.FUNCTIONAL.FORM}
\alias{STATIC.FUNCTIONAL.FORM}
\title{Create an Static (Time-Invariant) Functional Form}
\description{
Create an Static (Time-Invariant) Functional Form

Create an Static (Time-Invariant) Functional Form
}
\section{Super class}{
\code{jheem2::functional.form} -> \code{static.functional.form}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-static.functional.form-new}{\code{STATIC.FUNCTIONAL.FORM$new()}}
\item \href{#method-static.functional.form-clone}{\code{STATIC.FUNCTIONAL.FORM$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jheem2" data-topic="functional.form" data-id="apply.ontology.mapping"><a href='../../jheem2/html/functional.form.html#method-functional.form-apply.ontology.mapping'><code>jheem2::functional.form$apply.ontology.mapping()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jheem2" data-topic="functional.form" data-id="check"><a href='../../jheem2/html/functional.form.html#method-functional.form-check'><code>jheem2::functional.form$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jheem2" data-topic="functional.form" data-id="project"><a href='../../jheem2/html/functional.form.html#method-functional.form-project'><code>jheem2::functional.form$project()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jheem2" data-topic="functional.form" data-id="project.static"><a href='../../jheem2/html/functional.form.html#method-functional.form-project.static'><code>jheem2::functional.form$project.static()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-static.functional.form-new"></a>}}
\if{latex}{\out{\hypertarget{method-static.functional.form-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STATIC.FUNCTIONAL.FORM$new(
  value,
  link = c("identity", "log", "logistic")[1],
  min = NA,
  max = NA,
  value.name = "value",
  value.is.on.transformed.scale = F,
  overwrite.parameters.with.alphas = F,
  error.prefix = "Cannot create static functional.form: "
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{array representing the value}

\item{\code{link}}{The name of the transformation to the scale on which alphas are added to the value. One of "identity", "log", "logistic"}

\item{\code{min, max}}{The minimum and maximum values this functional form can create. The default (NA) sets to the min/max for the link specified (eg, 0-Inf for a log link, 0-1 for a logistic link)}

\item{\code{value.is.on.transformed.scale}}{Logical indicating whether the the value is already on the transformed scale at which alphas apply}

\item{\code{overwrite.parameters.with.alphas}}{Logical indicating whether, when alphas are applied, they OVERWRITE values. If F, alphas are added to the transformed value (on the transformed scale)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-static.functional.form-clone"></a>}}
\if{latex}{\out{\hypertarget{method-static.functional.form-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{STATIC.FUNCTIONAL.FORM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
