% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology.R
\name{[[<-.ontology}
\alias{[[<-.ontology}
\title{Modify an ontology}
\usage{
\method{[[}{ontology}(ont, i) <- value
}
\arguments{
\item{ont}{An ontology, as created by \code{\link{ontology}}}

\item{i}{What to subset}

\item{value}{New value to insert}
}
\value{
An ontology
}
\description{
Modify an ontology
}
