% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTOLOGY_ontology.R
\name{[<-.ontology}
\alias{[<-.ontology}
\title{Modify an ontology}
\usage{
\method{[}{ontology}(ont, i) <- value
}
\arguments{
\item{i}{What to subset}

\item{value}{New values to insert}
}
\value{
An ontology
}
\description{
Modify an ontology
}
