% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_simulation.R
\name{subset.simset}
\alias{subset.simset}
\title{Subset a Simulation Set}
\usage{
\method{subset}{simset}(simset, simulation.indices)
}
\arguments{
\item{simulation.indices}{An integer or logical vector specifying which simulations to keep.}
}
\description{
Subset a Simulation Set
}
