% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{subset_data}
\alias{subset_data}
\title{Subset a JHEEM Data Manager}
\usage{
subset_data(
  data.manager = get.default.data.manager(),
  dimension.values,
  ontology.name
)
}
\arguments{
\item{data.manager}{A jheem.data.manager object}

\item{dimension.values}{A named list that indicates what subset of the data should be kept. The values of dimension.values must be character vectors unless an 'ontology.name' is specified, in which case logical and numeric vectors may also be used.}

\item{ontology.name}{The name of a registered ontology for which data should be subset. If NULL, data from all ontologies will be subset.}
}
\description{
Subset a JHEEM Data Manager
}
