% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_simulation.R
\name{thin.simset}
\alias{thin.simset}
\title{Thin A Simulation Set}
\usage{
thin.simset(simset, n = NULL, keep = NULL, allow.expand = F)
}
\arguments{
\item{n}{Every Nth simulation will be kept, counting backwards from the last sim (ie., if n=3, the last simulation and the fourth-to-last simulation will be kept, etc.).}

\item{allow.expand}{Whether to duplicate simulations if keep > simset$n.sim}

\item{k}{Either a number of simulations to keep or a fraction to keep (between 0 and 1). Simulations will be drawn uniformly from the set.}
}
\description{
Subset a simulation set either by keeping every Nth simulation
or keeping a fraction of all simulations. One of 'n' or 'keep' can be specified,
with the other left as NULL.
}
\details{
Decimal values above 1 are rounded down to the nearest whole number.
}
