% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{track.cumulative.outcome}
\alias{track.cumulative.outcome}
\title{Track a Combination of Cumulative Outcomes}
\usage{
track.cumulative.outcome(
  specification,
  name,
  outcome.metadata,
  value,
  value.is.numerator = F,
  denominator.outcome = NULL,
  corresponding.data.outcome = NULL,
  keep.dimensions = NULL,
  exclude.dimensions = NULL,
  subset.dimension.values = NULL,
  rename.dimension.values = NULL,
  force.dim.names.to.keep.dimensions = F,
  allow.expand.dimensions = NULL,
  allow.expand.denominator.dimensions = NULL,
  dimension.aliases = NULL,
  dimension.alias.suffix = NULL,
  scale = NULL,
  from.year = -Inf,
  to.year = Inf,
  save = T
)
}
\arguments{
\item{specification}{The jheem.specification object to modify}

\item{name}{The name by which to access the outcome value after the simulation runs. Cannot be 'infected', 'uninfected' or the name of another outcome}

\item{outcome.metadata}{An object created by \code{\link{create.outcome.metadata}}, giving information on the outcome}

\item{value}{Either (1) a single character value referencing a separately-registered cumulative outcome or a static (value does not change over time) quantity, or (2) an expression that includes only the names of cumulative outcomes and static model quantities and operators +, -, *, /, ^, sqrt, log, and exp}

\item{value.is.numerator}{In the case where the outcome's scale is proportion, rate, or ratio (any scale that results from a numerator or a denominator), an indicator of whether the value represents just the numerator or the actual proportion/rate/ratio}

\item{denominator.outcome}{A denominator for aggregating the outcome or calculating proportions. Required if the scale of this outcome is not 'number' or 'non.negative.number', otherwise optional. Must be the name of a separately registered cumulative outcome that has type 'non.negative.number'}

\item{corresponding.data.outcome}{The (optional) name of a real-world outcome (with observations registered to a data.manager) to which this model outcome corresponds}

\item{keep.dimensions}{The dimensions which to keep in storing outcome values. All other dimensions are marginalized. Either keep.dimensions OR exclude.dimensions but not both can be set; if both are NULL, will keep all possible dimensions for the outcome.}

\item{exclude.dimensions}{Dimensions to be marginalized out when storing outcome values. Either keep.dimensions OR exclude.dimensions but not both can be set; if both are NULL, will keep all possible dimensions for the outcome.}

\item{subset.dimension.values}{A list of dimension values indicating for which values of which dimensions the outcome should be tracked}

\item{rename.dimension.values}{A list of names character vectors indicating which dimension values should be renamed in the outcome ontology}

\item{force.dim.names.to.keep.dimensions}{A logical indicating whether the dim.names of the outcome should be forced to match keep.dimensions (based on dimension values in compartments.for.infected.only, compartments.for.uninfected.only, and compartments.for.infected.and.uninfected) regardless of the dimensions of the underlying model quantities that it comprises.}

\item{allow.expand.dimensions, allow.expand.denominator.dimensions}{A character vector of dimensions which can be added to the outcome (or outcome's denominator) even if not present in its component outcomes or quantities. The outcome will have the same value for all values in that dimension}

\item{dimension.aliases}{A named character vector indicating what dimensions in the inputs to the outcome (the names of the vector) should be converted to (the values of the vector)}

\item{dimension.alias.suffix}{A single character value, indicating a suffix that should be removed from the dimensions of the inputs to the outcome. Eg, if ".to" is specified, then dimension "race.to" in outcome inputs becomes "race" in the outcome}

\item{scale}{The scale of this outcome. Can be NULL when outcome.metadata is not NULL (only required when not saving and not using outcome.metadata)}

\item{from.year, to.year}{The time span during which the outcome should be recorded}

\item{save}{A logical indicator of whether this outcome should be stored in simulations. If FALSE, the outcome will be available for calculating other outcomes (with \code{\link{track.cumulative.outcome}} or \code{\link{track.point.outcome}}), but will not be retrievable afterwards}
}
\description{
Track a Combination of Cumulative Outcomes
}
\details{
Creates a cumulative outcome to track by combining other cumulative outcomes and static quantities
}
