% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{track.quantity}
\alias{track.quantity}
\title{Indicate a Quantity to Track for Simulations}
\usage{
track.quantity(specification, name, value, cumulative = F)
}
\arguments{
\item{name}{The name by which to access the quantity after the simulation runs. Cannot be 'infected', 'uninfected' or the name of a tracked transition}

\item{value}{Either a character string or an expression. If dynamic==T, value must be one of 'infected', 'uninfected', 'incidence', 'mortality.infected', 'mortality.uninfected', or the name of a tracked transition (set by \code{\link{track.transition}}), or an expression in which one of those values is multiplied or divided by a combination of registered model.quantities/model.elements. If dynamic==F, value must contain only 'infected', 'uninfected', registered model.quantities/model.elements, other tracked quantities, or tracked transitions. If cumulative==T and dynamic==F, then value must furthermore be either a 'infected', 'uninfected', a tracked transition, or cumulative tracked quantity multiplied by some combination of registered model.quantities/model.elements (or a sum of such terms)}

\item{cumulative}{Whether this quantity should evaluate to the integral of value over time, or value itself}

\item{scale}{The scale to which getting this quantity from a simulation should evaluate.}

\item{dynamic}{If cumulative==T, whether value should be integrated WITHIN the diffeq solver (if dynamic==T) or AFTER it. Setting dynamic=T yields a (pretty) exact solution to the integral, whereas setting dynamic=F, yields an approximate (but computationally faster at run-time) solution}

\item{value.scale}{The scale to which the value denoted in value should evaluate. NB: if dynamic==T or cumulative==T, then value.scale, by definition, must be 'non.negative.number'}

\item{denominator}{If scale or value.scale requires a denominator to aggregate (ie 'rate', 'time', 'proportion'), then the character name of the denominator used in aggregating. Must be either the name of atracked transition or a tracked quantity with scale=='non.negative.number'}
}
\description{
Indicate a Quantity to Track for Simulations
}
