% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{track.sub.version.outcomes}
\alias{track.sub.version.outcomes}
\title{Track Sub Version Outcomes for a Specification}
\usage{
track.sub.version.outcomes(
  specification,
  sub.versions,
  outcome.names,
  keep.dimensions = NULL,
  subset.dimension.values = NULL,
  ontology.mapping = NULL,
  from.year = NULL,
  to.year = NULL
)
}
\description{
Track Sub Version Outcomes for a Specification
}
