% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPERS_dim_names_helpers.R
\name{union.dimension.values}
\alias{union.dimension.values}
\title{Get the union of dimension values sets}
\usage{
union.dimension.values(..., error.prefix = "")
}
\arguments{
\item{...}{can be either (1) dimension values or (2) lists of dimension values}
}
\description{
Get the union of dimension values sets
}
