% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{update.outcome.keep.dimensions}
\alias{update.outcome.keep.dimensions}
\title{Change the keep.dimensions of an outcome registered to an ancestor specification}
\usage{
\method{update}{outcome.keep.dimensions}(specification, outcome.name, keep.dimensions, exclude.dimensions = NULL)
}
\arguments{
\item{specification}{The jheem.specification object to modify}

\item{outcome.name}{The name of the outcome to update keep.dimensions for}

\item{keep.dimensions}{The dimensions which to keep in storing outcome values. All other dimensions are marginalized. Either keep.dimensions OR exclude.dimensions but not both can be set; if both are NULL, will keep all possible dimensions for the outcome.}

\item{exclude.dimensions}{Dimensions to be marginalized out when storing outcome values. Either keep.dimensions OR exclude.dimensions but not both can be set; if both are NULL, will keep all possible dimensions for the outcome.}
}
\description{
Change the keep.dimensions of an outcome registered to an ancestor specification
}
