% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_simulation.R
\name{copy.simulation.set}
\alias{copy.simulation.set}
\title{Make a copy of a Simulation Set}
\usage{
copy.simulation.set(simset)
}
\arguments{
\item{simset}{The simulation.set object to copy}
}
\description{
Make a copy of a Simulation Set
}
