% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_basic_likelihood.R
\name{create.basic.likelihood.instructions.with.specified.outcome}
\alias{create.basic.likelihood.instructions.with.specified.outcome}
\title{Create JHEEM Basic Likelihood Instructions With Specified Outcome}
\usage{
create.basic.likelihood.instructions.with.specified.outcome(
  outcome.for.sim,
  denominator.outcome.for.sim = NULL,
  outcome.value,
  from.year,
  to.year,
  omit.years = NULL,
  correlation.different.years = 0.5,
  observation.correlation.form = c("compound.symmetry", "autoregressive.1")[1],
  error.variance.term = NULL,
  error.variance.type = NULL,
  weights = list(),
  equalize.weight.by.year = T
)
}
\arguments{
\item{outcome.for.sim}{A single character vector specifying the simulation outcome to use.}

\item{denominator.outcome.for.sim}{(Optional) A single character vector specifying the simulation outcome to use a denominator if the outcome.for.sim is a ratio, proportion, or rate.}

\item{outcome.value}{A single, non-NA numeric value that is the value of the outcome at the totals level (i.e., not stratified).}

\item{from.year}{-Inf or a single integer value specifying the earliest year for which data should be pulled.}

\item{to.year}{Inf or a single integer value specifying the latest year for which data should be pulled.}

\item{omit.years}{(Optional) An integer vector of years to ignore.}

\item{correlation.different.years}{A single numeric value specifying the correlation between observations of different years.}

\item{observation.correlation.form}{Which correlation form to use in building the measurement error covariance matrix. May be 'compound.symmetry' or 'autoregressive.1'.}

\item{error.variance.term}{If applicable for the 'error.variance.type' used, a single numeric value.}

\item{error.variance.type}{A single character value specifying how the error variance is described. May refer to a single quantity with "sd" (standard deviation), "variance", "cv" (covariance); or may request the error variance be sourced from data with "data.sd", "data.variance", and "data.ci" (confidence interval).}

\item{weights}{A list containing only numeric vectors and objects of class 'jheem.likelihood.weights', which are generated by \code{\link{create.likelihood.weights}}}

\item{equalize.weight.by.year}{A Boolean indicating whether every year should be weighted equally before the weights from 'weights' are applied.}
}
\description{
Create JHEEM Basic Likelihood Instructions With Specified Outcome
}
