% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_basic_likelihood.R
\name{create.basic.likelihood.instructions.with.specified.outcome}
\alias{create.basic.likelihood.instructions.with.specified.outcome}
\title{Create JHEEM Basic Likelihood Instructions With Specified Outcome}
\usage{
create.basic.likelihood.instructions.with.specified.outcome(
  outcome.for.sim,
  outcome.value,
  from.year,
  to.year,
  omit.years = NULL,
  minimum.error.sd = 0,
  correlation.different.years = 0.5,
  observation.correlation.form = c("compound.symmetry", "autoregressive.1")[1],
  error.variance.term = NULL,
  error.variance.type = NULL,
  weights = list(),
  equalize.weight.by.year = T,
  name = outcome.for.sim
)
}
\arguments{
\item{outcome.for.sim}{A single character vector specifying the simulation outcome to use.}

\item{outcome.value}{A single, non-NA numeric value that is the value of the outcome at the totals level (i.e., not stratified).}

\item{from.year}{-Inf or a single integer value specifying the earliest year for which data should be pulled.}

\item{to.year}{Inf or a single integer value specifying the latest year for which data should be pulled.}

\item{omit.years}{(Optional) An integer vector of years to ignore.}

\item{minimum.error.sd}{A single, positive numeric value. If any standard deviations are calculated to a value below this, they will be replaced with this value. This can help keep the likelihood from computing to negative infinity.}

\item{correlation.different.years}{A single numeric value specifying the correlation between observations of different years.}

\item{observation.correlation.form}{Which correlation form to use in building the measurement error covariance matrix. May be 'compound.symmetry' or 'autoregressive.1'.}

\item{error.variance.type, error.variance.term}{The type of error variance(s) to be used and the corresponding value or function required for each. See details for the allowable values. For more than one type, supply them in a vector. If you therefore need more than one term, supply these in a list.}

\item{weights}{A list containing only numeric vectors and objects of class 'jheem.likelihood.weights', which are generated by \code{\link{create.likelihood.weights}}}

\item{equalize.weight.by.year}{A Boolean indicating whether every year should be weighted equally before the weights from 'weights' are applied.}
}
\description{
Create JHEEM Basic Likelihood Instructions With Specified Outcome
}
\details{
Error variance may be specified in a variety of ways, and more than
one can be used in the same likelihood. 'error.variance.type' is a vector
containing single character values describing the type(s) of variance to use,
and there is one 'error.variance.term' corresponding to each. The types "sd"
(standard deviation), "variance", "cv" (covariance), and "exp.of.variance"
(exponent of variance) require a single numeric value as their term. The
types "data.sd", "data.variance", and "data.cv" indicate that error data will
be pulled from the data manager and matched directly to data points, and they
require NULL as their term. Finally, the type "function.sd" requires as its
term a function that takes only arguments "data", "details", "version", and "location"
and returns a
numeric array of the same dimensions as "data". Supplying multiple types (and
therefore terms) results in the variances created by each being summed prior
to use in the measurement error covariance matrix.

As an example, one type could be specified with an 'error.variance.type' of
"cv" and an 'error.variance.term' of 0.13 (or list(0.13)). Two or more
types might look like an 'error.variance.type' of `c("data.sd",
"function.sd") and an 'error.variance.term' of list(NULL, my_sd_function). It
is possible to use the same type multiple times, although this is not
recommended.
}
\seealso{
Other Basic Likelihoods: 
\code{\link{create.basic.likelihood.instructions.with.included.multiplier}()},
\code{\link{create.basic.likelihood.instructions}()},
\code{\link{create.time.lagged.comparison.likelihood.instructions.with.included.multiplier}()},
\code{\link{create.time.lagged.comparison.likelihood.instructions}()}
}
\concept{Basic Likelihoods}
