% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_ifelse_likelihood_instructions.R
\name{create.ifelse.likelihood.instructions}
\alias{create.ifelse.likelihood.instructions}
\title{Create If-Else Likelihood Instructions}
\usage{
create.ifelse.likelihood.instructions(..., name = NULL)
}
\arguments{
\item{...}{'jheem.likelihood.instructions' objects that all have the same 'outcome.for.sim'.}

\item{name}{Name for the ifelse; will default to}
}
\description{
Create instructions that can try a sequence of likelihood instructions until the first one that works without error.
}
\details{
The instructions are attempted in the order that they are supplied.
}
