% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{create.jheem.engine}
\alias{create.jheem.engine}
\title{Create an Engine to Run JHEEM Simulations}
\usage{
create.jheem.engine(
  version,
  location,
  end.year,
  sub.version = NULL,
  max.run.time.seconds = Inf,
  solver.metadata = NULL
)
}
\arguments{
\item{version, location, sub.version}{The JHEEM version, sub.version, and location for which to run simulations}

\item{end.year}{The year to run simulations to}

\item{max.run.time.seconds}{The maximum amount of time to run simulations before terminating}

\item{solver.metadata}{A solver.metadata object, as created by \code{\link{create.solver.metadata}}, specifying what ODE solver to use in running simulations}
}
\description{
Create an Engine to Run JHEEM Simulations
}
