% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_main.R
\name{create.likelihood.weights}
\alias{create.likelihood.weights}
\title{Create Weights to Be Applied in Calculating a JHEEM Likelihood}
\usage{
create.likelihood.weights(
  total.weight = 1,
  dimension.values = list(),
  is.recursive = T
)
}
\arguments{
\item{total.weight}{The weight that is applied to all observations in the}

\item{is.recursive}{Should this weight apply to deeper levels of stratification? For exampel, if F and dimension.values is the default empty list, this weight would only apply to data points that are totals.}
}
\description{
Create Weights to Be Applied in Calculating a JHEEM Likelihood
}
