% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKELIHOODS_nested_proportion_likelihood.R
\name{create.nested.proportion.likelihood.instructions.with.included.multiplier}
\alias{create.nested.proportion.likelihood.instructions.with.included.multiplier}
\title{Created Nested Proportion Likelihood Instructions With Included Multiplier}
\usage{
create.nested.proportion.likelihood.instructions.with.included.multiplier(
  outcome.for.data,
  denominator.outcome.for.data,
  outcome.for.sim,
  denominator.outcome.for.sim = NULL,
  outcome.for.n.multipliers = denominator.outcome.for.data,
  location.types,
  minimum.geographic.resolution.type,
  maximum.locations.per.type = 3,
  dimensions = NULL,
  levels.of.stratification = 0:length(dimensions),
  from.year = -Inf,
  to.year = Inf,
  omit.years = NULL,
  sources.to.use = NULL,
  exclude.denominator.ontology.names = NULL,
  redundant.location.threshold = 5,
  included.multiplier,
  included.multiplier.sd,
  included.multiplier.correlation = NULL,
  included.multiplier.correlation.structure = c("compound.symmetry",
    "autoregressive.1")[1],
  p.bias.inside.location,
  p.bias.outside.location,
  p.bias.sd.inside.location,
  p.bias.sd.outside.location,
  within.location.p.error.correlation = 0.5,
  within.location.n.error.correlation = 0.5,
  correlation.different.locations = 0,
  correlation.different.years = 0.5,
  correlation.different.strata = 0.1,
  correlation.different.sources = 0.3,
  correlation.same.source.different.details = 0.3,
  observation.correlation.form = c("compound.symmetry", "autoregressive.1")[1],
  n.multiplier.cv = 0.1,
  p.error.variance.term = NULL,
  p.error.variance.type = NULL,
  n.error.variance.term = 0.05,
  n.error.variance.type = "cv",
  weights,
  equalize.weight.by.year = T,
  partitioning.function
)
}
\arguments{
\item{outcome.for.data}{A single character vector specifying the outcome that data managers will use to find data.}

\item{outcome.for.sim}{A single character vector specifying the simulation outcome to use.}

\item{denominator.outcome.for.sim}{(Optional) A single character vector specifying the simulation outcome to use a denominator if the outcome.for.sim is a ratio, proportion, or rate.}

\item{location.types}{The types of the locations that contain or are contained by the model location.}

\item{minimum.geographic.resolution.type}{The type of location used to partition locations. The type of the model location AND 'location.types' types must all completely enclose regions of this type}

\item{dimensions}{A character vector of dimensions, excluding year, from which stratifications will be generated.}

\item{levels.of.stratification}{An integer vector specifying how the dimensions should be combined to form strata. '0' indicates totals (not stratified) while '1' indicates strata that are each stratified by one dimension at a time, '2' indicates strat that are each stratified by a combination of two dimensions at a time, etc. May not exceed the number of dimensions. Defaults to NULL, which is equivalent to '0'.}

\item{from.year}{-Inf or a single integer value specifying the earliest year for which data should be pulled.}

\item{to.year}{Inf or a single integer value specifying the latest year for which data should be pulled.}

\item{omit.years}{(Optional) An integer vector of years to ignore.}

\item{sources.to.use}{A character vector of sources from which to pull data.}

\item{included.multiplier.correlation.structure}{}

\item{p.bias.inside.location}{A single numeric value specifying the bias in the outcome proportion between locations inside the model location and the model location itself}

\item{p.bias.outside.location}{A single numeric value specifying the bias in the outcome proportion between locations outside the model location and the model location itself}

\item{p.bias.sd.inside.location}{The standard deviation associated with 'p.bias.inside.location'}

\item{within.location.p.error.correlation}{}

\item{correlation.different.years}{A single numeric value specifying the correlation between observations of different years.}

\item{correlation.different.strata}{A single numeric value specifying the correlation between observations from different strata.}

\item{correlation.different.sources}{A single numeric value specifying the correlation between observations from different sources.}

\item{correlation.same.source.different.details}{A single numeric value specifying the correlation between observations of the same source but which have different 'details'.}

\item{observation.correlation.form}{Which correlation form to use in building the measurement error covariance matrix. May be 'compound.symmetry' or 'autoregressive.1'.}

\item{weights}{A list containing only numeric vectors and objects of class 'jheem.likelihood.weights', which are generated by \code{\link{create.likelihood.weights}}}

\item{equalize.weight.by.year}{A Boolean indicating whether every year should be weighted equally before the weights from 'weights' are applied.}

\item{partitioning.function}{A function that partitions values in the data ontology into values in the model ontology. Must have two arguments, "arr" and "version" and return an array with the same dimnames as the input array.}
}
\description{
Created Nested Proportion Likelihood Instructions With Included Multiplier
}
