% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STYLE_MANAGER_style_manager.R
\name{create.style.manager}
\alias{create.style.manager}
\title{Create a JHEEM Style Manager}
\usage{
create.style.manager(
  name,
  description,
  aes.for.simset = "linetype",
  aes.for.split.by = "color",
  aes.for.simset.single.point = "shape",
  aes.for.location = "shape",
  name.for.sim.color.legend = "sim color",
  name.for.data.color.legend = "data color",
  name.for.data.shape.legend = "shape",
  alpha.ribbon = 0.2,
  set.as.default = F
)
}
\arguments{
\item{name}{The name of the style manager}

\item{description}{A short description}
}
\description{
Create a JHEEM Style Manager
}
