% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS_style_manager.R
\name{create.style.manager}
\alias{create.style.manager}
\title{Create A Style Manager}
\usage{
create.style.manager(
  color.sim.by = "stratum",
  color.data.by = "stratum",
  shade.data.by = "location",
  linetype.sim.by = "simset",
  shape.data.by = "source",
  shape.sim.by = "simset",
  sim.palette = ggsci::pal_jama(),
  data.palette = ggsci::pal_jama(),
  linetypes = c("solid", "dashed", "dotted", "dotdash"),
  shapes = 21:25,
  shade.increment = -16,
  alpha.ribbon = 0.2
)
}
\description{
Create A Style Manager
}
