% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATION_main.R
\name{get.calibration.progress}
\alias{get.calibration.progress}
\title{Get Calibration Progress}
\usage{
get.calibration.progress(
  version,
  locations,
  calibration.code,
  root.dir = get.jheem.root.directory(),
  round.to.digits = 0,
  as.pct = T
)
}
\description{
Get Calibration Progress
}
