% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERSIONS_version_manager.R
\name{get.specification.for.version}
\alias{get.specification.for.version}
\title{Get a JHEEM Specification Object}
\usage{
get.specification.for.version(version)
}
\arguments{
\item{version}{The name of the version under which the specification was registered}
}
\value{
An object of class 'jheem.specification'
}
\description{
Get a JHEEM Specification Object
}
\seealso{
Other JHEEM Version Management Functions: 
\code{\link{get.prior.versions}()},
\code{\link{register.model.specification}()},
\code{\link{version.has.been.registered}()}
}
\concept{JHEEM Version Management Functions}
