% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS_simplot.R
\name{prepare.plot}
\alias{prepare.plot}
\title{Prepare Plot Data
A utility that performs the data-pulling half of the simplot operation
for applications like webtools that need the step to be separate.}
\usage{
prepare.plot(
  simset.list = NULL,
  outcomes = NULL,
  locations = NULL,
  corresponding.data.outcomes = NULL,
  split.by = NULL,
  facet.by = NULL,
  dimension.values = list(),
  target.ontology = NULL,
  plot.which = c("sim.and.data", "sim.only", "data.only")[1],
  summary.type = c("individual.simulation", "mean.and.interval",
    "median.and.interval")[1],
  plot.year.lag.ratio = F,
  title = "location",
  data.manager = get.default.data.manager(),
  debug = F
)
}
\arguments{
\item{simset.list}{A list of jheem.simulation.set objects}

\item{outcomes}{A character vector of which simulation outcomes to plot}

\item{split.by}{At most one dimension}

\item{facet.by}{Any number of dimensions but cannot include the split.by dimension}

\item{plot.which}{Should simulation data and calibration data be plotted ('sim.and.data'), or only simulation data ('sim.only'), or only calibration data ('data.only')}

\item{title}{NULL or a single, non-NA character value. If "location", the location of the first provided simset (if any) will be used for the title.}

\item{data.manager}{The data.manager from which to draw real-world data for the plots}
}
\description{
Prepare Plot Data
A utility that performs the data-pulling half of the simplot operation
for applications like webtools that need the step to be separate.
}
