% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{put.data}
\alias{put.data}
\title{Put data into a data manager}
\usage{
put.data(
  data.manager = get.default.data.manager(),
  data,
  outcome,
  metric,
  source,
  ontology.name,
  dimension.values,
  url,
  details,
  allow.na.to.overwrite = F
)
}
\arguments{
\item{data.manager}{A jheem.data.manager object}

\item{data}{A numeric array or scalar value containing the data to store. If it is an array, it must have named dimnames set}

\item{outcome}{The outcome type for the data. Must be an outcome previously registered with \code{\link{register.data.outcome}}}

\item{source}{The (single) character name of the source from which these data derive. Must be a source previously registered with \code{\link{register.data.source}} Note: a source is conceived such that one source cannot contain two sets of data for the same set of dimension values}

\item{ontology.name}{The name of the ontology which the data follow. Must be an ontology previously registered with \code{\link{register.data.ontology}}}

\item{dimension.values}{A named list that indicates what subset of a bigger data element these particular data should be stored into. The names of dimension values. The values of dimension.values can be either (1) character vectors}

\item{url}{A character vector with at least one element giving one or more URLs indicating where the data derive from}

\item{details}{A character vector with at least one element giving one or more URLs giving data collection details. In general, data collected and tabulated using the same approach should have the same details, whereas data with different methods/tabulation should have different details}

\item{allow.na.to.overwrite}{A logical indicator for whether NA values in data should be allowed to overwrite previous values put to the data manager (if data have been put previously)}
}
\description{
Put data into a data manager
}
