% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{register.data.outcome}
\alias{register.data.outcome}
\title{Register an outcome to a data manager before putting data for that outcome}
\usage{
register.data.outcome(
  data.manager = get.default.data.manager(),
  outcome,
  metadata,
  denominator.outcome = NULL,
  denominator.lags.by.one.year = F,
  overwrite = F,
  allow.missing.denominator.outcome = F
)
}
\arguments{
\item{data.manager}{A jheem.data.manager object}

\item{outcome}{The name (a single character value) of the outcome. This is the 'internal' name by which the outcome will be referenced in accessing the data manager}

\item{metadata}{An object of class 'outcome.metadata', as created by \code{\link{create.outcome.metadata}} that contains information about how to display the outcome}

\item{denominator.outcome}{The denominator outcome type that should be used when aggregating data (taking a weighted average) for this outcome type. Must be a previously registered outcome with scale='non.negative.number'. Only applies if scale is 'rate', 'proportion', or 'time'}

\item{denominator.lags.by.one.year}{Flag indicating whether the denominator data is shifting one year earlier than this outcome's data. For example, data for this outcome in year 2020 will use the 2019 denominator data if this flag is set to TRUE.}

\item{overwrite}{A logical indicating whether the information on this outcome should overwrite previously-registered information about this outcome. However, this registration must include the same metadata$scale and denominator.outcome (ie, can't change the structure of the outcome, only the display 'trappings')}

\item{allow.missing.denominator.outcome}{A logical allowing the user to register an outcome of scale 'rate', 'proportion' or 'time' without supplying a denominator outcome}
}
\description{
Register an outcome to a data manager before putting data for that outcome
}
