% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{register.model.quantity}
\alias{register.model.quantity}
\title{Register a Model Quantity}
\usage{
register.model.quantity(
  specification,
  name,
  value,
  scale = NULL,
  dimensions = names(dimension.values),
  dimension.values = NULL,
  apply.aliases.to.dimension.values = F,
  ...,
  na.replacement = as.numeric(NA)
)
}
\arguments{
\item{specification}{The jheem.specification object to modify}

\item{name}{The name of the model quantity. Cannot overlap with names of model elements}

\item{value}{Either: (1) a numeric object, (2) a character string giving the name of another model quantity or a model element, (3) an 'call' or 'expression' object (generated by expression() or expr()) containing an expression comprising other model quantities or model elements, and the operators +, -, *, /, ^, sqrt, log, and exp, or (4) a function whose arguments are some subset of 'specification.info', 'location', other model quantities and model elements, and ...}

\item{scale}{An optional parameter, indicating the scale for this model.quantity. If specified, running the model will throw an error if evaluating the quantity results in an invalid value for the scale (eg a negative rate or a proportion > 1). Can be 'rate', 'ratio', 'proportion', 'time', 'number', 'non.negative.number'}

\item{dimensions}{An optional parameter, indicating the dimensions the quantity is expected to have (if dim.names is not specified, the specific values of the dimensions will be inferred automatically)}

\item{dimension.values}{An optional parameter, indicating (some of) the elements that the dimnames of the quantity are supposed to contain. This must be a named list of character vectors, whose names are a subset of dimensions.}

\item{apply.aliases.to.dimension.values}{A logical indicating whether, if dimension.values are supplied, they should have the compartment.value.aliases for the model specification applied to them}

\item{...}{Additional parameters to be passed to value if a function is passed as the value}

\item{na.replacement}{A single numeric value that should be used to replace NA values generated in calculating the quantity's value (note, if na.replacement==NA, no handling of NAs is performed and NAs will cause an error in model execution)}
}
\description{
Register a Model Quantity
}
