% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERSIONS_version_manager.R
\name{register.sampled.parameters.for.version}
\alias{register.sampled.parameters.for.version}
\title{Register Parameters to Be Sampled Prior to Making Projections from a JHEEM Version}
\usage{
register.sampled.parameters.for.version(version, distribution, apply.function)
}
\arguments{
\item{version}{The name of an EHE version for which a specification has been registered with \code{\link{register.model.specification}}}

\item{distribution}{An object of class 'distribution' that specifies the prior distribution for the parameters}

\item{apply.function}{}
}
\description{
Register Parameters to Be Sampled Prior to Making Projections from a JHEEM Version
}
\details{
*Sampled* parameters have their values randomly chosen from a distribution prior to running projections PAST the calibration period
}
