% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{register.transition}
\alias{register.transition}
\title{Register a Transition between Compartments in a Dimension}
\usage{
register.transition(
  specification,
  name = NULL,
  dimension,
  from.compartments,
  to.compartments,
  value,
  groups,
  na.replacement = as.numeric(NA),
  ...
)
}
\arguments{
\item{specification}{The jheem.specification object to modify}

\item{name}{The name by which to refer to this transition. Optional, but if not specified (NULL), you will not be able to refer to this transition in other quantities or override it in descendant specifications}

\item{dimension}{The name of the dimension in which this transition operates}

\item{from.compartments, to.compartments}{The compartments from and to which this transition applies. May be (1) one or more names of compartments, (2) one or more character compartment.value.aliases, (3) one or more integers denoting compartments, (4) a logical vector, which, when applied to the dimension names, denotes the compartments}

\item{value}{Either: (1) a numeric object, (2) a character string giving the name of another model quantity or a model element, (3) an 'call' or 'expression' object (generated by expression() or expr()) containing an expression comprising other model quantities or model elements, and the operators +, -, *, /, ^, sqrt, log, and exp, or (4) a function whose arguments are some subset of 'specification.info', 'location', other model quantities and model elements, and ...}

\item{groups}{The groups to which this transition applies. Either 'uninfected', 'infected', or 'all' ('all' applies the transition to all groups which it could apply to)}

\item{na.replacement}{A single numeric value that should be used to replace NA values generated in calculating the quantity's value (note, if na.replacement==NA, no handling of NAs is performed and NAs will cause an error in model execution)}

\item{...}{Additional parameters to be passed to value if a function is passed as the value}
}
\description{
Register a Transition between Compartments in a Dimension
}
