% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_simulation.R
\name{rerun.simulations}
\alias{rerun.simulations}
\title{Rerun a Simulation Set}
\usage{
rerun.simulations(
  simset,
  max.run.time.seconds = Inf,
  sub.version = simset$sub.version,
  from.year = simset$from.year,
  to.year = simset$to.year,
  verbose = T
)
}
\arguments{
\item{simset}{The JHEEM simulation set to re-run}

\item{max.run.time.seconds}{The maximum amount of time to run simulations before terminating}

\item{sub.version}{The sub-version to run}

\item{from.year, to.year}{The years to keep in the simulation object}

\item{verbose}{Whether to print updates as simulations are run}
}
\description{
Rerun a Simulation Set
}
\details{
Helpful if the specification has changed
}
