% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS_style_manager.R
\name{set.default.style.manager}
\alias{set.default.style.manager}
\title{Set the Default Style Manager}
\usage{
set.default.style.manager(style.manager, type = c("ggplot", "plotly")[1])
}
\description{
Set the Default Style Manager
}
