% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{set.element.functional.form.from.time}
\alias{set.element.functional.form.from.time}
\alias{set.element.functional.form.to.time}
\title{Set the Times From and To Which the Functional Form Determines a Model Element's Value}
\usage{
set.element.functional.form.from.time(
  model.settings,
  element.name,
  from.time,
  wrt.version = model.settings$version
)

set.element.functional.form.to.time(
  model.settings,
  element.name,
  to.time,
  wrt.version = model.settings$version
)
}
\arguments{
\item{model.settings}{An object of class 'jheem.model.settings'}

\item{element.name}{The name of the model element to set functional.form times for}

\item{from.time, to.time}{The time from or to which the functional form is active}

\item{wrt.version}{The version relative to which the element name is resolved (if an element.name is overwritten in a descendant version, using wrt.version can be used to select which specification's model element is to be modified)}
}
\description{
Set the Times From and To Which the Functional Form Determines a Model Element's Value
}
\section{Functions}{
\itemize{
\item \code{set.element.functional.form.to.time()}: 

}}
\seealso{
Other Functions to create and modify a model settings: 
\code{\link{set.element.functional.form.future.slope}()},
\code{\link{set.element.functional.form.interaction.alphas}()},
\code{\link{set.element.ramp.times}()}
}
\concept{Functions to create and modify a model settings}
