% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{set.element.functional.form.interaction.alphas}
\alias{set.element.functional.form.interaction.alphas}
\title{Set Interaction Alpha Values for a Functional Form for a Model Element}
\usage{
set.element.functional.form.interaction.alphas(
  model.settings,
  element.name,
  alpha.name,
  value,
  applies.to.dimension.values,
  wrt.version = model.settings$version
)
}
\arguments{
\item{model.settings}{An object of class 'jheem.model.settings'}

\item{element.name}{The name of the model element whose functional form we wish to modify}

\item{alpha.name}{The name of the functional form parameter that we want to set alphas for}

\item{value}{A single numeric value, which will apply to every combination of applies.to.dimension.values which are from different dimensions}

\item{applies.to.dimension.values}{Either a character vector where each element is either a character vector. Must be named with the dimension to which each dimension value applies}

\item{wrt.version}{The version relative to which the element name is resolved (if an element.name is overwritten in a descendant version, using wrt.version can be used to select which specification's model element is to be modified)}
}
\description{
Set Interaction Alpha Values for a Functional Form for a Model Element
}
\details{
Alphas for an 'interaction' will apply to every combination of applies.to.dimension.values from different dimensions (as opposed to 'main effect' apply to all values in a dimension
}
\seealso{
Other Functions to create and modify a model settings: 
\code{\link{set.element.functional.form.from.time}()},
\code{\link{set.element.functional.form.future.slope}()},
\code{\link{set.element.ramp.times}()}
}
\concept{Functions to create and modify a model settings}
