% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{set.element.functional.form.main.effect.alphas}
\alias{set.element.functional.form.main.effect.alphas}
\title{Set Main Effect Alpha Values for a Functional Form for a Model Element}
\usage{
set.element.functional.form.main.effect.alphas(
  model.settings,
  element.name,
  alpha.name,
  values,
  dimension,
  applies.to.dimension.values = names(values),
  wrt.version = model.settings$version
)
}
\arguments{
\item{model.settings}{An object of class 'jheem.model.settings'}

\item{element.name}{The name of the model element whose functional form we wish to modify}

\item{alpha.name}{The name of the functional form parameter that we want to set alphas for}

\item{values}{A numeric vector of values (either a single value applied to all dimension values OR a vector of the same length as applies.to.dimension.values)}

\item{dimension}{The dimension to which values in applies.to.dimension.values apply}

\item{applies.to.dimension.values}{A character vector of values within the given dimension to which the alphas should apply}

\item{wrt.version}{The version relative to which the element name is resolved (if an element.name is overwritten in a descendant version, using wrt.version can be used to select which specification's model element is to be modified)}
}
\description{
Set Main Effect Alpha Values for a Functional Form for a Model Element
}
\details{
Alphas for a 'main effect' apply to all values in a dimension (as opposed to 'interaction effects', which apply to combinations of values)
}
\concept{Functions to create and modify model settings}
