% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JHEEM_engine.R
\name{set.element.value}
\alias{set.element.value}
\title{Set the Value of a Model Element}
\usage{
set.element.value(
  model.settings,
  element.name,
  value,
  wrt.version = model.settings$version
)
}
\arguments{
\item{model.settings}{An object of class 'jheem.model.settings'}

\item{element.name}{The name of the model element to update a value for}

\item{value}{The new value for the model element}

\item{wrt.version}{The version relative to which the element name is resolved (if an element.name is overwritten in a descendant version, using wrt.version can be used to select which specification's model element is to be modified)}
}
\description{
Set the Value of a Model Element
}
\concept{Functions to modify model settings}
