% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS_simplot.R
\name{simplot.data.only}
\alias{simplot.data.only}
\title{Simplot Data Only}
\usage{
simplot.data.only(
  outcomes,
  locations,
  split.by = NULL,
  facet.by = NULL,
  dimension.values = list(),
  target.ontology = NULL,
  plot.year.lag.ratio = F,
  title = "location",
  n.facet.rows = NULL,
  append.url = F,
  data.manager = get.default.data.manager(),
  style.manager = get.default.style.manager(),
  show.data.pull.error = F,
  debug = F
)
}
\arguments{
\item{outcomes}{A character vector of which simulation outcomes to plot}

\item{split.by}{At most one dimension}

\item{facet.by}{Any number of dimensions but cannot include the split.by dimension}

\item{title}{NULL or a single, non-NA character value. If "location", the first location provided in "locations" will be used for the title.}

\item{data.manager}{The data.manager from which to draw real-world data for the plots}

\item{style.manager}{We are going to have to define this down the road. It's going to govern how we do lines and sizes and colors. For now, just hard code those in, and we'll circle back to it}

\item{show.data.pull.error}{Not finding data does not block the plot from showing simulation projections, but if you'd like to see in error when data doesn't appear, set this to TRUE.}

\item{A}{character vector of which simulation outcomes to plot. Note: This should be the names of the outcomes in the provided data manager, NOT in the simulations.}
}
\description{
Simplot Data Only
}
