% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOTS_simplot.R
\name{simplot}
\alias{simplot}
\title{Plot Simulations And Data}
\usage{
simplot(
  ...,
  outcomes = NULL,
  corresponding.data.outcomes = NULL,
  split.by = NULL,
  facet.by = NULL,
  dimension.values = list(),
  dimension.values.post.mapping = list(),
  target.ontology = NULL,
  plot.which = c("sim.and.data", "sim.only")[1],
  summary.type = c("individual.simulation", "mean.and.interval",
    "median.and.interval")[1],
  label.function = NULL,
  plot.year.lag.ratio = F,
  title = "location",
  n.facet.rows = NULL,
  append.url = F,
  data.manager = get.default.data.manager(),
  style.manager = get.default.style.manager(),
  show.data.pull.error = F,
  debug = F
)
}
\arguments{
\item{...}{One or more jheem.simulation.set objects and at most one character vector of outcomes (as an alternative to the 'outcomes' argument)}

\item{outcomes}{A character vector of which simulation outcomes to plot}

\item{corresponding.data.outcomes}{Specify directly which data outcomes should be plotted against simulation outcomes. Must be NULL or a character vector with outcomes as names; all of those outcomes must be present in either the 'outcomes' argument or in '...'"}

\item{split.by}{At most one dimension}

\item{facet.by}{Any number of dimensions but cannot include the split.by dimension}

\item{dimension.values.post.mapping}{How to subset data in the form it will take after mappings have been applied to align sim and data ontologies. A list.}

\item{plot.which}{Should simulation data and calibration data be plotted ('sim.and.data'), or only simulation data ('sim.only')}

\item{label.function}{A function to reformat labels. If NULL, will use the first provided simset's "get.labels" property, or will do no transformation if plotting data only.}

\item{title}{NULL or a single, non-NA character value. If "location", the location of the first provided simset (if any) will be used for the title.}

\item{data.manager}{The data.manager from which to draw real-world data for the plots}

\item{style.manager}{We are going to have to define this down the road. It's going to govern how we do lines and sizes and colors. For now, just hard code those in, and we'll circle back to it}

\item{show.data.pull.error}{Not finding data does not block the plot from showing simulation projections, but if you'd like to see in error when data doesn't appear, set this to TRUE.}
}
\description{
Plot Simulations And Data
}
\details{
Returns a ggplot object:
 - With one panel for each combination of outcome x facet.by
 - x-axis is year
 - y-axis is outcome
}
