% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_MANAGER_data_manager.R
\name{subset_data}
\alias{subset_data}
\title{Subset a JHEEM Data Manager}
\usage{
subset_data(
  data.manager = get.default.data.manager(),
  outcomes.to.keep = NULL,
  sources.to.keep = NULL,
  ontologies.to.keep = NULL,
  incomplete.dimension.values.to.keep = NULL,
  retain.registrations = T
)
}
\arguments{
\item{outcomes.to.keep, sources.to.keep, ontologies.to.keep}{Character vectors of which outcomes, sources, and/or ontologies to retain data for.}

\item{incomplete.dimension.values.to.keep}{A list of which years, locations, etc. to retain data for. Must be character vectors with the dimensions as names.}

\item{retain.registrations}{Should registered outcomes (or sources, ontologies, etc.) be un-registered if they aren't being kept?}

\item{A}{jheem.data.manager object}
}
\description{
Subset a JHEEM Data Manager
}
