% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{track.cumulative.proportion.from.rate}
\alias{track.cumulative.proportion.from.rate}
\title{Track an Outcome Formed by Integrating a Time-Varying Rate into a Proportion}
\usage{
track.cumulative.proportion.from.rate(
  specification,
  name,
  outcome.metadata,
  rate.value,
  denominator.outcome,
  corresponding.data.outcome = NULL,
  calculate.proportion.leaving = T,
  keep.dimensions = NULL,
  exclude.dimensions = NULL,
  subset.dimension.values = NULL,
  rename.dimension.values = NULL,
  force.dim.names.to.keep.dimensions = F,
  dimension.aliases = NULL,
  dimension.alias.suffix = NULL,
  from.year = -Inf,
  to.year = Inf,
  save = T
)
}
\arguments{
\item{specification}{The jheem.specification object to modify}

\item{name}{The name by which to access the outcome value after the simulation runs. Cannot be 'infected', 'uninfected' or the name of another outcome}

\item{outcome.metadata}{An object created by \code{\link{create.outcome.metadata}}, giving information on the outcome}

\item{rate.value}{The rate to be integrated. Either (1) a single character value referencing a separately-registered "point" (non-cumulative) outcome or a quantity, or (2) an expression that includes only the names of point outcomes and model quantities and operators +, -, *, /, ^, sqrt, log, and exp}

\item{denominator.outcome}{A denominator for aggregating the outcome or calculating proportions. Required if the scale of this outcome is not 'number' or 'non.negative.number', otherwise optional. Must be the name of a separately registered cumulative outcome that has type 'non.negative.number'}

\item{corresponding.data.outcome}{The (optional) name of a real-world outcome (with observations registered to a data.manager) to which this model outcome corresponds}

\item{calculate.proportion.leaving}{If TRUE, integrates to get the proportion of those who "leave" the imaginary compartment at the given rate. If FALSE, calculates the proportion who have not left (ie 1-proportion.leaving)}

\item{keep.dimensions}{The dimensions which to keep in storing outcome values. All other dimensions are marginalized. Either keep.dimensions OR exclude.dimensions but not both can be set; if both are NULL, will keep all possible dimensions for the outcome.}

\item{exclude.dimensions}{Dimensions to be marginalized out when storing outcome values. Either keep.dimensions OR exclude.dimensions but not both can be set; if both are NULL, will keep all possible dimensions for the outcome.}

\item{subset.dimension.values}{A list of dimension values indicating for which values of which dimensions the outcome should be tracked}

\item{rename.dimension.values}{A list of names character vectors indicating which dimension values should be renamed in the}

\item{force.dim.names.to.keep.dimensions}{A logical indicating whether the dim.names of the outcome should be forced to match keep.dimensions (based on dimension values in compartments.for.infected.only, compartments.for.uninfected.only, and compartments.for.infected.and.uninfected) regardless of the dimensions of the underlying model quantities that it comprises.}

\item{dimension.aliases}{A named character vector indicating what dimensions in the inputs to the outcome (the names of the vector) should be converted to (the values of the vector)}

\item{dimension.alias.suffix}{A single character value, indicating a suffix that should be removed from the dimensions of the inputs to the outcome. Eg, if ".to" is specified, then dimension "race.to" in outcome inputs becomes "race" in the outcome}

\item{from.year, to.year}{The time span during which the outcome should be recorded}

\item{save}{A logical indicator of whether this outcome should be stored in simulations. If FALSE, the outcome will be available for calculating other outcomes (with \code{\link{track.cumulative.outcome}} or \code{\link{track.point.outcome}}), but will not be retrievable afterwards}
}
\description{
Track an Outcome Formed by Integrating a Time-Varying Rate into a Proportion
}
\details{
Calculates a proportion derived from having a 'compartment' which people leave at a given rate, and calculates the proportion who either leave or remain in that compartment over time
}
