% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{track.transition}
\alias{track.transition}
\title{Indicate a Transition to Track for Simulations}
\usage{
track.transition(
  specification,
  name,
  dimension,
  from.compartments,
  to.compartments,
  multiply.by = NULL,
  keep.dimensions = NULL
)
}
\arguments{
\item{specification}{The jheem.specification object to modify}

\item{name}{The name by which to access the transition value after the simulation runs. Cannot be 'infected', 'uninfected' or the name of a tracked quantity}

\item{dimension}{The name of the (single) dimension along which we are tracking a transition}

\item{from.compartments, to.compartments}{The names (or aliases) of the from and to compartments of the transition to track that gives information about the resulting quantity}

\item{multiply.by}{An optional quantity to multiply the transition by at each time step before storing. Can be either (1) a numeric value, (2) a single character value referencing a model quantity or element, or (3) an expression that includes only the names of model quantities and elements and operators +, -, *, /, ^, sqrt, log, and exp}

\item{keep.dimensions}{The dimensions over which to keep in storing values. By default (keep.dimensions==NULL), will keep all possible dimensions for the value}

\item{groups}{The groups to include in this transition. Must be a subset of 'infected', 'uninfected', or 'all' ('all' selects all possible groups to which the transition could apply)}

\item{metadata}{An object of class 'outcome.metadata', as created by \code{\link{create.outcome.metadata}}}
}
\description{
Indicate a Transition to Track for Simulations
}
\details{
Integrates the transition (or the product of transition x multiply by), such that the simulation stores, for each year y, the integral from y to y-1 of (transition * multiply.by)
}
