% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIFICATION_model_specification.R
\name{track.transition}
\alias{track.transition}
\title{Mark a Transition for Tracking}
\usage{
track.transition(
  specification,
  name,
  outcome.metadata,
  dimension,
  from.compartments,
  to.compartments,
  denominator.outcome = NULL,
  include.tags = NULL,
  exclude.tags = NULL,
  groups = NULL,
  sub.versions = NULL,
  multiply.by = NULL,
  corresponding.data.outcome = NULL,
  keep.dimensions = NULL,
  exclude.dimensions = NULL,
  subset.dimension.values = NULL,
  rename.dimension.values = NULL,
  dimension.aliases = NULL,
  dimension.alias.suffix = NULL,
  scale = NULL,
  from.year = -Inf,
  to.year = Inf,
  save = T
)
}
\arguments{
\item{specification}{The jheem.specification object to modify}

\item{name}{The name by which to access the outcome value after the simulation runs. Cannot be 'infected', 'uninfected' or the name of another outcome}

\item{outcome.metadata}{An object created by \code{\link{create.outcome.metadata}}, giving information on the outcome}

\item{dimension}{The name of the (single) dimension along which we are tracking a transition}

\item{from.compartments, to.compartments}{The names (or aliases) of the from and to compartments of the transition to track that gives information about the resulting quantity}

\item{denominator.outcome}{Optional: A denominator for calculating proportions. Must be the name of a separately registered cumulative outcome that has type 'non.negative.number'}

\item{include.tags}{The tags to track for (as given in the call to \code{\link{register.transition}}). Choosing NULL (the default value) tracks for all tags for which the outcome could apply minus excluded tags}

\item{exclude.tags}{Tags to exclude from tracking.}

\item{groups}{The groups to track for ('infected' or 'uninfected'). Choosing NULL (the default value) tracks for all groups for which the outcome could apply}

\item{sub.versions}{The sub-versions of the model for which this outcome will be recorded. If NULL, the outcome will be recorded for all sub-versions}

\item{multiply.by}{An optional quantity to multiply the outcome value by at each time step before storing. Can be either (1) a numeric value, (2) a single character value referencing a model quantity or element, or (3) an expression that includes only the names of model quantities and elements and operators +, -, *, /, ^, sqrt, log, and exp}

\item{corresponding.data.outcome}{The (optional) name of a real-world outcome (with observations registered to a data.manager) to which this model outcome corresponds}

\item{keep.dimensions}{The dimensions which to keep in storing outcome values. All other dimensions are marginalized. Either keep.dimensions OR exclude.dimensions but not both can be set; if both are NULL, will keep all possible dimensions for the outcome.}

\item{exclude.dimensions}{Dimensions to be marginalized out when storing outcome values. Either keep.dimensions OR exclude.dimensions but not both can be set; if both are NULL, will keep all possible dimensions for the outcome.}

\item{subset.dimension.values}{A list of dimension values indicating for which values of which dimensions the outcome should be tracked}

\item{rename.dimension.values}{A list of names character vectors indicating which dimension values should be renamed in the}

\item{dimension.aliases}{A named character vector indicating what dimensions in the inputs to the outcome (the names of the vector) should be converted to (the values of the vector)}

\item{dimension.alias.suffix}{A single character value, indicating a suffix that should be removed from the dimensions of the inputs to the outcome. Eg, if ".to" is specified, then dimension "race.to" in outcome inputs becomes "race" in the outcome}

\item{scale}{The scale of this outcome. Can be NULL when outcome.metadata is not NULL (only required when not saving and not using outcome.metadata)}

\item{from.year, to.year}{The time span during which the outcome should be recorded}

\item{save}{A logical indicator of whether this outcome should be stored in simulations. If FALSE, the outcome will be available for calculating other outcomes (with \code{\link{track.cumulative.outcome}} or \code{\link{track.point.outcome}}), but will not be retrievable afterwards}
}
\description{
Mark a Transition for Tracking
}
\details{
Integrates the transition (or the product of transition x multiply by), such that the simulation stores, for each year y, the integral from y to y+1 of (transition * multiply.by)
}
